poise::builtins

Function help

pub async fn help<U, E>(
    ctx: Context<'_, U, E>,
    command: Option<&str>,
    config: HelpConfiguration<'_>,
) -> Result<(), Error>

A help command that outputs text in a code block, groups commands by categories, and annotates commands with a slash if they exist as slash commands.

Example usage from Ferris, the Discord bot running in the Rust community server:

/// Show this menu
#[poise::command(prefix_command, track_edits, slash_command)]
pub async fn help(
    ctx: Context<'_>,
    #[description = "Specific command to show help about"] command: Option<String>,
) -> Result<(), Error> {
    let config = poise::builtins::HelpConfiguration {
        extra_text_at_bottom: "\
Type ?help command for more info on a command.
You can edit your message to the bot and the bot will edit its response.",
        ..Default::default()
    };
    poise::builtins::help(ctx, command.as_deref(), config).await?;
    Ok(())
}

Output:

Playground:
  ?play        Compile and run Rust code in a playground
  ?eval        Evaluate a single Rust expression
  ?miri        Run code and detect undefined behavior using Miri
  ?expand      Expand macros to their raw desugared form
  ?clippy      Catch common mistakes using the Clippy linter
  ?fmt         Format code using rustfmt
  ?microbench  Benchmark small snippets of code
  ?procmacro   Compile and use a procedural macro
  ?godbolt     View assembly using Godbolt
  ?mca         Run performance analysis using llvm-mca
  ?llvmir      View LLVM IR using Godbolt
Crates:
  /crate       Lookup crates on crates.io
  /doc         Lookup documentation
Moderation:
  /cleanup     Deletes the bot's messages for cleanup
  /ban         Bans another person
  ?move        Move a discussion to another channel
  /rustify     Adds the Rustacean role to members
Miscellaneous:
  ?go          Evaluates Go code
  /source      Links to the bot GitHub repo
  /help        Show this menu

Type ?help command for more info on a command.
You can edit your message to the bot and the bot will edit its response.

Examples found in repository?

examples/feature_showcase/builtins.rs (line 16)

11pub async fn help(ctx: Context<'_>, command: Option<String>) -> Result<(), Error> {
12    let configuration = poise::builtins::HelpConfiguration {
13        // [configure aspects about the help message here]
14        ..Default::default()
15    };
16    poise::builtins::help(ctx, command.as_deref(), configuration).await?;
17    Ok(())
18}

Hide additional examples

examples/basic_structure/commands.rs (lines 11-18)

5pub async fn help(
6    ctx: Context<'_>,
7    #[description = "Specific command to show help about"]
8    #[autocomplete = "poise::builtins::autocomplete_command"]
9    command: Option<String>,
10) -> Result<(), Error> {
11    poise::builtins::help(
12        ctx,
13        command.as_deref(),
14        poise::builtins::HelpConfiguration {
15            extra_text_at_bottom: "This is an example bot made to showcase features of my custom Discord bot framework",
16            ..Default::default()
17        },
18    )
19    .await?;
20    Ok(())
21}

examples/help_generation/main.rs (line 314)

287async fn help(
288    ctx: Context<'_>,
289    #[description = "Command to get help for"]
290    #[rest]
291    mut command: Option<String>,
292) -> Result<(), Error> {
293    // This makes it possible to just make `help` a subcommand of any command
294    // `/fruit help` turns into `/help fruit`
295    // `/fruit help apple` turns into `/help fruit apple`
296    if ctx.invoked_command_name() != "help" {
297        command = match command {
298            Some(c) => Some(format!("{} {}", ctx.invoked_command_name(), c)),
299            None => Some(ctx.invoked_command_name().to_string()),
300        };
301    }
302    let extra_text_at_bottom = "\
303Type `?help command` for more info on a command.
304You can edit your `?help` message to the bot and the bot will edit its response.";
305
306    let config = HelpConfiguration {
307        show_subcommands: true,
308        show_context_menu_commands: true,
309        ephemeral: true,
310        extra_text_at_bottom,
311
312        ..Default::default()
313    };
314    poise::builtins::help(ctx, command.as_deref(), config).await?;
315    Ok(())
316}