poise::builtins

Function on_error

pub async fn on_error<U, E: Into<Box<dyn Error + Send + Sync>>>(
    error: FrameworkError<'_, U, E>,
) -> Result<(), Error>

An error handler that logs errors either via the [tracing] crate or via a Discord message. Set up a logger like tracing subscriber (e.g. tracing_subscriber::fmt::init()) to see the logged errors from this method.

If the user invoked the command wrong (crate::FrameworkError::ArgumentParse), the command help is displayed and the user is directed to the help menu.

Can return an error if sending the Discord error message failed. You can decide for yourself how to handle this, for example:

if let Err(e) = poise::builtins::on_error(error).await {
    tracing::error!("Fatal error while sending error message: {}", e);
}

Examples found in repository?

examples/basic_structure/main.rs (line 32)

22async fn on_error(error: poise::FrameworkError<'_, Data, Error>) {
23    // This is our custom error handler
24    // They are many errors that can occur, so we only handle the ones we want to customize
25    // and forward the rest to the default handler
26    match error {
27        poise::FrameworkError::Setup { error, .. } => panic!("Failed to start bot: {:?}", error),
28        poise::FrameworkError::Command { error, ctx, .. } => {
29            println!("Error in command `{}`: {:?}", ctx.command().name, error,);
30        }
31        error => {
32            if let Err(e) = poise::builtins::on_error(error).await {
33                println!("Error while handling error: {}", e)
34            }
35        }
36    }
37}

Hide additional examples

examples/invocation_data/main.rs (line 110)

51async fn main() {
52    let token = std::env::var("TOKEN").unwrap();
53    let intents =
54        serenity::GatewayIntents::non_privileged() | serenity::GatewayIntents::MESSAGE_CONTENT;
55
56    let framework = poise::Framework::builder()
57        .setup(move |ctx, _, framework| {
58            Box::pin(async move {
59                let commands =
60                    poise::builtins::create_application_commands(&framework.options().commands);
61
62                serenity::GuildId::new(703332075914264606)
63                    .set_commands(ctx, commands)
64                    .await
65                    .unwrap();
66                Ok(())
67            })
68        })
69        .options(poise::FrameworkOptions {
70            pre_command: |ctx| {
71                Box::pin(async move {
72                    println!(
73                        "In pre_command: {:?}",
74                        ctx.invocation_data::<&str>().await.as_deref()
75                    );
76                })
77            },
78            command_check: Some(|ctx| {
79                Box::pin(async move {
80                    // Global command check is the first callback that's invoked, so let's set the
81                    // data here
82                    println!("Writing invocation data!");
83                    ctx.set_invocation_data("hello").await;
84
85                    println!(
86                        "In global check: {:?}",
87                        ctx.invocation_data::<&str>().await.as_deref()
88                    );
89
90                    Ok(true)
91                })
92            }),
93            post_command: |ctx| {
94                Box::pin(async move {
95                    println!(
96                        "In post_command: {:?}",
97                        ctx.invocation_data::<&str>().await.as_deref()
98                    );
99                })
100            },
101            on_error: |err| {
102                Box::pin(async move {
103                    match err {
104                        poise::FrameworkError::Command { ctx, .. } => {
105                            println!(
106                                "In on_error: {:?}",
107                                ctx.invocation_data::<&str>().await.as_deref()
108                            );
109                        }
110                        err => poise::builtins::on_error(err).await.unwrap(),
111                    }
112                })
113            },
114
115            commands: vec![invocation_data_test()],
116            prefix_options: poise::PrefixFrameworkOptions {
117                prefix: Some("~".into()),
118                ..Default::default()
119            },
120            ..Default::default()
121        })
122        .build();
123
124    let client = serenity::ClientBuilder::new(token, intents)
125        .framework(framework)
126        .await;
127
128    client.unwrap().start().await.unwrap();
129}

examples/feature_showcase/main.rs (line 104)

31async fn main() {
32    tracing_subscriber::fmt::init();
33
34    let framework = poise::Framework::builder()
35        .options(poise::FrameworkOptions {
36            commands: vec![
37                attachment_parameter::file_details(),
38                attachment_parameter::totalsize(),
39                autocomplete::greet(),
40                bool_parameter::oracle(),
41                #[cfg(feature = "cache")]
42                builtins::servers(),
43                checks::shutdown(),
44                checks::modonly(),
45                checks::delete(),
46                checks::ferrisparty(),
47                checks::cooldowns(),
48                checks::minmax(),
49                checks::get_guild_name(),
50                checks::only_in_dms(),
51                checks::lennyface(),
52                checks::permissions_v2(),
53                choice_parameter::choice(),
54                choice_parameter::inline_choice(),
55                choice_parameter::inline_choice_int(),
56                code_block_parameter::code(),
57                collector::boop(),
58                context_menu::user_info(),
59                context_menu::echo(),
60                inherit_checks::parent_checks(),
61                localization::welcome(),
62                modal::modal(),
63                modal::component_modal(),
64                paginate::paginate(),
65                panic_handler::div(),
66                parameter_attributes::addmultiple(),
67                parameter_attributes::voiceinfo(),
68                parameter_attributes::say(),
69                parameter_attributes::punish(),
70                parameter_attributes::stringlen(),
71                raw_identifiers::r#move(),
72                response_with_reply::reply(),
73                subcommands::parent(),
74                subcommand_required::parent_subcommand_required(),
75                track_edits::test_reuse_response(),
76                track_edits::add(),
77                user_apps::everywhere(),
78                user_apps::everywhere_context(),
79                user_apps::user_install(),
80                user_apps::not_in_guilds(),
81                user_apps::user_install_guild(),
82            ],
83            prefix_options: poise::PrefixFrameworkOptions {
84                prefix: Some("~".into()),
85                non_command_message: Some(|_, msg| {
86                    Box::pin(async move {
87                        println!("non command message!: {}", msg.content);
88                        Ok(())
89                    })
90                }),
91                ..Default::default()
92            },
93            on_error: |error| {
94                Box::pin(async move {
95                    println!("what the hell");
96                    match error {
97                        poise::FrameworkError::ArgumentParse { error, .. } => {
98                            if let Some(error) = error.downcast_ref::<serenity::RoleParseError>() {
99                                println!("Found a RoleParseError: {:?}", error);
100                            } else {
101                                println!("Not a RoleParseError :(");
102                            }
103                        }
104                        other => poise::builtins::on_error(other).await.unwrap(),
105                    }
106                })
107            },
108            ..Default::default()
109        })
110        .setup(move |ctx, _ready, framework| {
111            Box::pin(async move {
112                poise::builtins::register_globally(ctx, &framework.options().commands).await?;
113                Ok(Data {})
114            })
115        })
116        .build();
117
118    let token = std::env::var("DISCORD_TOKEN").unwrap();
119    let intents =
120        serenity::GatewayIntents::non_privileged() | serenity::GatewayIntents::MESSAGE_CONTENT;
121
122    let client = serenity::ClientBuilder::new(token, intents)
123        .framework(framework)
124        .await;
125
126    client.unwrap().start().await.unwrap()
127}