Function poise::builtins::register_application_commands

pub async fn register_application_commands<U, E>(
    ctx: Context<'_, U, E>, 
    global: bool
) -> Result<(), Error>

Generic function to register application commands, either globally or in a guild. Only bot owners can register globally, only guild owners can register in guild.

If you want, you can copy paste this help message:

Register application commands in this guild or globally

Run with no arguments to register in guild, run with argument "global" to register globally.

Examples found in repository

examples/framework_usage/main.rs (line 43)

42
43
44
45
46

async fn register(ctx: Context<'_>, #[flag] global: bool) -> Result<(), Error> {
    poise::builtins::register_application_commands(ctx, global).await?;

    Ok(())
}

examples/serenity_example_port/main.rs (line 128)

127
128
129
130
131

async fn register(ctx: Context<'_>, #[flag] global: bool) -> Result<(), Error> {
    poise::builtins::register_application_commands(ctx, global).await?;

    Ok(())
}