poise::builtins

Function register_globally

pub async fn register_globally<U, E>(
    http: impl AsRef<Http>,
    commands: &[Command<U, E>],
) -> Result<(), Error>

Registers the given list of application commands to Discord as global commands.

Thin wrapper around create_application_commands that funnels the returned builder into serenity::Command::set_global_commands.

Examples found in repository?

examples/quickstart/main.rs (line 31)

20async fn main() {
21    let token = std::env::var("DISCORD_TOKEN").expect("missing DISCORD_TOKEN");
22    let intents = serenity::GatewayIntents::non_privileged();
23
24    let framework = poise::Framework::builder()
25        .options(poise::FrameworkOptions {
26            commands: vec![age()],
27            ..Default::default()
28        })
29        .setup(|ctx, _ready, framework| {
30            Box::pin(async move {
31                poise::builtins::register_globally(ctx, &framework.options().commands).await?;
32                Ok(Data {})
33            })
34        })
35        .build();
36
37    let client = serenity::ClientBuilder::new(token, intents)
38        .framework(framework)
39        .await;
40    client.unwrap().start().await.unwrap();
41}

Hide additional examples

examples/advanced_cooldowns/main.rs (line 43)

31async fn main() {
32    let token = std::env::var("DISCORD_TOKEN").expect("missing DISCORD_TOKEN");
33    let framework = poise::Framework::builder()
34        .options(poise::FrameworkOptions {
35            commands: vec![dynamic_cooldowns()],
36            // This is important! Or else, the command will be marked as invoked before our custom
37            // cooldown code has run - even if the command ends up not running!
38            manual_cooldowns: true,
39            ..Default::default()
40        })
41        .setup(|ctx, _ready, framework| {
42            Box::pin(async move {
43                poise::builtins::register_globally(ctx, &framework.options().commands).await?;
44                Ok(Data {})
45            })
46        })
47        .build();
48
49    let client = serenity::Client::builder(token, serenity::GatewayIntents::non_privileged())
50        .framework(framework)
51        .await;
52
53    client.unwrap().start().await.unwrap();
54}

examples/help_generation/main.rs (line 347)

319async fn main() {
320    let token = std::env::var("DISCORD_TOKEN").expect("missing DISCORD_TOKEN");
321    let intents =
322        serenity::GatewayIntents::non_privileged() | serenity::GatewayIntents::MESSAGE_CONTENT;
323
324    let framework = poise::Framework::builder()
325        .options(poise::FrameworkOptions {
326            commands: vec![
327                fruit(),
328                vegetable(),
329                beer(),
330                meat(),
331                dairy(),
332                help(),
333                context_food(),
334                context_fruit(),
335                context_vegetable(),
336                context_meat(),
337                food_react(),
338            ],
339            prefix_options: poise::PrefixFrameworkOptions {
340                prefix: Some("?".into()),
341                ..Default::default()
342            },
343            ..Default::default()
344        })
345        .setup(|ctx, _ready, framework| {
346            Box::pin(async move {
347                poise::builtins::register_globally(ctx, &framework.options().commands).await?;
348                Ok(Data {})
349            })
350        })
351        .build();
352
353    let client = serenity::ClientBuilder::new(token, intents)
354        .framework(framework)
355        .await;
356
357    client.unwrap().start().await.unwrap();
358}

examples/basic_structure/main.rs (line 100)

40async fn main() {
41    tracing_subscriber::fmt::init();
42
43    // FrameworkOptions contains all of poise's configuration option in one struct
44    // Every option can be omitted to use its default value
45    let options = poise::FrameworkOptions {
46        commands: vec![commands::help(), commands::vote(), commands::getvotes()],
47        prefix_options: poise::PrefixFrameworkOptions {
48            prefix: Some("~".into()),
49            edit_tracker: Some(Arc::new(poise::EditTracker::for_timespan(
50                Duration::from_secs(3600),
51            ))),
52            additional_prefixes: vec![
53                poise::Prefix::Literal("hey bot,"),
54                poise::Prefix::Literal("hey bot"),
55            ],
56            ..Default::default()
57        },
58        // The global error handler for all error cases that may occur
59        on_error: |error| Box::pin(on_error(error)),
60        // This code is run before every command
61        pre_command: |ctx| {
62            Box::pin(async move {
63                println!("Executing command {}...", ctx.command().qualified_name);
64            })
65        },
66        // This code is run after a command if it was successful (returned Ok)
67        post_command: |ctx| {
68            Box::pin(async move {
69                println!("Executed command {}!", ctx.command().qualified_name);
70            })
71        },
72        // Every command invocation must pass this check to continue execution
73        command_check: Some(|ctx| {
74            Box::pin(async move {
75                if ctx.author().id == 123456789 {
76                    return Ok(false);
77                }
78                Ok(true)
79            })
80        }),
81        // Enforce command checks even for owners (enforced by default)
82        // Set to true to bypass checks, which is useful for testing
83        skip_checks_for_owners: false,
84        event_handler: |_ctx, event, _framework, _data| {
85            Box::pin(async move {
86                println!(
87                    "Got an event in event handler: {:?}",
88                    event.snake_case_name()
89                );
90                Ok(())
91            })
92        },
93        ..Default::default()
94    };
95
96    let framework = poise::Framework::builder()
97        .setup(move |ctx, ready, framework| {
98            Box::pin(async move {
99                println!("Logged in as {}", ready.user.name);
100                poise::builtins::register_globally(ctx, &framework.options().commands).await?;
101                Ok(Data {
102                    votes: Mutex::new(HashMap::new()),
103                })
104            })
105        })
106        .options(options)
107        .build();
108
109    let token = var("DISCORD_TOKEN")
110        .expect("Missing `DISCORD_TOKEN` env var, see README for more information.");
111    let intents =
112        serenity::GatewayIntents::non_privileged() | serenity::GatewayIntents::MESSAGE_CONTENT;
113
114    let client = serenity::ClientBuilder::new(token, intents)
115        .framework(framework)
116        .await;
117
118    client.unwrap().start().await.unwrap()
119}

examples/feature_showcase/main.rs (line 114)

31async fn main() {
32    tracing_subscriber::fmt::init();
33
34    let framework = poise::Framework::builder()
35        .options(poise::FrameworkOptions {
36            commands: vec![
37                attachment_parameter::file_details(),
38                attachment_parameter::totalsize(),
39                autocomplete::greet(),
40                bool_parameter::oracle(),
41                #[cfg(feature = "cache")]
42                builtins::servers(),
43                builtins::help(),
44                builtins::pretty_help(),
45                checks::shutdown(),
46                checks::modonly(),
47                checks::delete(),
48                checks::ferrisparty(),
49                checks::cooldowns(),
50                checks::minmax(),
51                checks::get_guild_name(),
52                checks::only_in_dms(),
53                checks::lennyface(),
54                checks::permissions_v2(),
55                choice_parameter::choice(),
56                choice_parameter::inline_choice(),
57                choice_parameter::inline_choice_int(),
58                code_block_parameter::code(),
59                collector::boop(),
60                context_menu::user_info(),
61                context_menu::echo(),
62                inherit_checks::parent_checks(),
63                localization::welcome(),
64                modal::modal(),
65                modal::component_modal(),
66                paginate::paginate(),
67                panic_handler::div(),
68                parameter_attributes::addmultiple(),
69                parameter_attributes::voiceinfo(),
70                parameter_attributes::say(),
71                parameter_attributes::punish(),
72                parameter_attributes::stringlen(),
73                raw_identifiers::r#move(),
74                response_with_reply::reply(),
75                subcommands::parent(),
76                subcommand_required::parent_subcommand_required(),
77                track_edits::test_reuse_response(),
78                track_edits::add(),
79                user_apps::everywhere(),
80                user_apps::everywhere_context(),
81                user_apps::user_install(),
82                user_apps::not_in_guilds(),
83                user_apps::user_install_guild(),
84            ],
85            prefix_options: poise::PrefixFrameworkOptions {
86                prefix: Some("~".into()),
87                non_command_message: Some(|_, _, msg| {
88                    Box::pin(async move {
89                        println!("non command message!: {}", msg.content);
90                        Ok(())
91                    })
92                }),
93                ..Default::default()
94            },
95            on_error: |error| {
96                Box::pin(async move {
97                    println!("what the hell");
98                    match error {
99                        poise::FrameworkError::ArgumentParse { error, .. } => {
100                            if let Some(error) = error.downcast_ref::<serenity::RoleParseError>() {
101                                println!("Found a RoleParseError: {:?}", error);
102                            } else {
103                                println!("Not a RoleParseError :(");
104                            }
105                        }
106                        other => poise::builtins::on_error(other).await.unwrap(),
107                    }
108                })
109            },
110            ..Default::default()
111        })
112        .setup(move |ctx, _ready, framework| {
113            Box::pin(async move {
114                poise::builtins::register_globally(ctx, &framework.options().commands).await?;
115                Ok(Data {})
116            })
117        })
118        .build();
119
120    let token = std::env::var("DISCORD_TOKEN").unwrap();
121    let intents =
122        serenity::GatewayIntents::non_privileged() | serenity::GatewayIntents::MESSAGE_CONTENT;
123
124    let client = serenity::ClientBuilder::new(token, intents)
125        .framework(framework)
126        .await;
127
128    client.unwrap().start().await.unwrap()
129}