Derive Macro poise::SlashChoiceParameter

#[derive(SlashChoiceParameter)]
{
    // Attributes available to this derive:
    #[name]
}

Use this derive macro on an enum to easily generate a choice parameter type. A choice parameter is mainly useful in slash commands. It allows you to constrain input to a fixed set of choices.

Example

#[derive(Debug, poise::SlashChoiceParameter)]
pub enum MyStringChoice {
    #[name = "The first choice"]
    ChoiceA,
    #[name = "The second choice"]
    ChoiceB,
}

/// Dummy command to test slash command choice parameters
#[poise::command(prefix_command, slash_command)]
pub async fn choice(
    ctx: Context<'_>,
    #[description = "The choice you want to choose"] choice: MyStringChoice,
) -> Result<(), Error> {
    ctx.say(format!("You entered {:?}", choice)).await?;
    Ok(())
}