Function poise::dispatch::parse_invocation

pub async fn parse_invocation<'a, U: Send + Sync, E>(
    framework: FrameworkContext<'a, U, E>,
    ctx: &'a Context,
    msg: &'a Message,
    trigger: MessageDispatchTrigger,
    invocation_data: &'a Mutex<Box<dyn Any + Send + Sync>>,
    parent_commands: &'a mut Vec<&'a Command<U, E>>
) -> Result<Option<PrefixContext<'a, U, E>>, FrameworkError<'a, U, E>>

Given a Message and some context data, parses prefix, command etc. out of the message and returns the resulting crate::PrefixContext. To run the command, see run_invocation.

Examples found in repository?

src/dispatch/prefix.rs (lines 196-203)

188
189
190
191
192
193
194
195
196
197
198
199
200
201
202
203
204
205
206
207
208
209

pub async fn dispatch_message<'a, U: Send + Sync, E>(
    framework: crate::FrameworkContext<'a, U, E>,
    ctx: &'a serenity::Context,
    msg: &'a serenity::Message,
    trigger: crate::MessageDispatchTrigger,
    invocation_data: &'a tokio::sync::Mutex<Box<dyn std::any::Any + Send + Sync>>,
    parent_commands: &'a mut Vec<&'a crate::Command<U, E>>,
) -> Result<(), crate::FrameworkError<'a, U, E>> {
    if let Some(ctx) = parse_invocation(
        framework,
        ctx,
        msg,
        trigger,
        invocation_data,
        parent_commands,
    )
    .await?
    {
        run_invocation(ctx).await?;
    }
    Ok(())
}