Function poise::dispatch::run_invocation

pub async fn run_invocation<U, E>(
    ctx: PrefixContext<'_, U, E>
) -> Result<(), FrameworkError<'_, U, E>>

Given an existing parsed command invocation from parse_invocation, run it, including all the before and after code like checks and built in filters from edit tracking

Examples found in repository?

src/dispatch/prefix.rs (line 206)

188
189
190
191
192
193
194
195
196
197
198
199
200
201
202
203
204
205
206
207
208
209

pub async fn dispatch_message<'a, U: Send + Sync, E>(
    framework: crate::FrameworkContext<'a, U, E>,
    ctx: &'a serenity::Context,
    msg: &'a serenity::Message,
    trigger: crate::MessageDispatchTrigger,
    invocation_data: &'a tokio::sync::Mutex<Box<dyn std::any::Any + Send + Sync>>,
    parent_commands: &'a mut Vec<&'a crate::Command<U, E>>,
) -> Result<(), crate::FrameworkError<'a, U, E>> {
    if let Some(ctx) = parse_invocation(
        framework,
        ctx,
        msg,
        trigger,
        invocation_data,
        parent_commands,
    )
    .await?
    {
        run_invocation(ctx).await?;
    }
    Ok(())
}