Enum poise::CommandErrorContext

pub enum CommandErrorContext<'a, U, E> {
    Prefix(PrefixCommandErrorContext<'a, U, E>),
    Application(ApplicationCommandErrorContext<'a, U, E>),
}

Context of an error in user code

Contains slightly different data depending on where the error was raised

Variants

Prefix(PrefixCommandErrorContext<'a, U, E>)

Prefix command specific error context

Tuple Fields of Prefix

0: PrefixCommandErrorContext<'a, U, E>

Application(ApplicationCommandErrorContext<'a, U, E>)

Application command specific error context

Tuple Fields of Application

0: ApplicationCommandErrorContext<'a, U, E>

Implementations

impl<'a, U, E> CommandErrorContext<'a, U, E>

pub fn command(&self) -> CommandRef<'_, U, E>

Returns a reference to the command during whose execution the error occured.

Examples found in repository

examples/framework_usage/main.rs (line 48)

44
45
46
47
48
49
50
51
52

async fn on_error(error: Error, ctx: poise::ErrorContext<'_, Data, Error>) {
    match ctx {
        poise::ErrorContext::Setup => panic!("Failed to start bot: {:?}", error),
        poise::ErrorContext::Command(ctx) => {
            println!("Error in command `{}`: {:?}", ctx.command().name(), error)
        }
        _ => println!("Other error: {:?}", error),
    }
}

examples/serenity_example_port/main.rs (line 168)

↕

163
164
165
166
167
168
169
170
171
172
173
174
175
176
177
178
179
180
181

async fn on_error(error: Error, ctx: poise::ErrorContext<'_, Data, Error>) {
    match ctx {
        poise::ErrorContext::Command(ctx) => {
            println!(
                "Command '{}' returned error {:?}",
                ctx.command().name(),
                error
            );
        }
        poise::ErrorContext::Listener(event) => {
            println!(
                "Listener returned error during {} event: {:?}",
                event.name(),
                error
            );
        }
        _ => {}
    }
}

pub fn while_checking(&self) -> bool

Whether the error occured in a pre-command check or during execution

pub fn ctx(&self) -> Context<'a, U, E>

Further command context