Enum poise::Event

pub enum Event<'a> {

    CacheReady {
        guilds: Vec<GuildId>,
    },
    ChannelCreate {
        channel: &'a GuildChannel,
    },
    CategoryCreate {
        category: &'a ChannelCategory,
    },
    CategoryDelete {
        category: &'a ChannelCategory,
    },
    ChannelDelete {
        channel: &'a GuildChannel,
    },
    ChannelPinsUpdate {
        pin: ChannelPinsUpdateEvent,
    },
    ChannelUpdate {
        old: Option<Channel>,
        new: Channel,
    },
    GuildBanAddition {
        guild_id: GuildId,
        banned_user: User,
    },
    GuildBanRemoval {
        guild_id: GuildId,
        unbanned_user: User,
    },
    GuildCreate {
        guild: Guild,
        is_new: bool,
    },
    GuildDelete {
        incomplete: UnavailableGuild,
        full: Option<Guild>,
    },
    GuildEmojisUpdate {
        guild_id: GuildId,
        current_state: HashMap<EmojiId, Emoji>,
    },
    GuildIntegrationsUpdate {
        guild_id: GuildId,
    },
    GuildMemberAddition {
        new_member: Member,
    },
    GuildMemberRemoval {
        guild_id: GuildId,
        user: User,
        member_data_if_available: Option<Member>,
    },
    GuildMemberUpdate {
        old_if_available: Option<Member>,
        new: Member,
    },
    GuildMembersChunk {
        chunk: GuildMembersChunkEvent,
    },
    GuildRoleCreate {
        new: Role,
    },
    GuildRoleDelete {
        guild_id: GuildId,
        removed_role_id: RoleId,
        removed_role_data_if_available: Option<Role>,
    },
    GuildRoleUpdate {
        old_data_if_available: Option<Role>,
        new: Role,
    },
    GuildStickersUpdate {
        guild_id: GuildId,
        current_state: HashMap<StickerId, Sticker>,
    },
    GuildUnavailable {
        guild_id: GuildId,
    },
    GuildUpdate {
        old_data_if_available: Option<Guild>,
        new_but_incomplete: PartialGuild,
    },
    IntegrationCreate {
        integration: Integration,
    },
    IntegrationUpdate {
        integration: Integration,
    },
    IntegrationDelete {
        integration_id: IntegrationId,
        guild_id: GuildId,
        application_id: Option<ApplicationId>,
    },
    InviteCreate {
        data: InviteCreateEvent,
    },
    InviteDelete {
        data: InviteDeleteEvent,
    },
    Message {
        new_message: Message,
    },
    MessageDelete {
        channel_id: ChannelId,
        deleted_message_id: MessageId,
        guild_id: Option<GuildId>,
    },
    MessageDeleteBulk {
        channel_id: ChannelId,
        multiple_deleted_messages_ids: Vec<MessageId>,
        guild_id: Option<GuildId>,
    },
    MessageUpdate {
        old_if_available: Option<Message>,
        new: Option<Message>,
        event: MessageUpdateEvent,
    },
    ReactionAdd {
        add_reaction: Reaction,
    },
    ReactionRemove {
        removed_reaction: Reaction,
    },
    ReactionRemoveAll {
        channel_id: ChannelId,
        removed_from_message_id: MessageId,
    },
    PresenceReplace {
        new_presences: Vec<Presence>,
    },
    PresenceUpdate {
        new_data: Presence,
    },
    Ready {
        data_about_bot: Ready,
    },
    Resume {
        event: ResumedEvent,
    },
    ShardStageUpdate {
        update: ShardStageUpdateEvent,
    },
    StageInstanceCreate {
        stage_instance: StageInstance,
    },
    StageInstanceDelete {
        stage_instance: StageInstance,
    },
    StageInstanceUpdate {
        stage_instance: StageInstance,
    },
    ThreadCreate {
        thread: GuildChannel,
    },
    ThreadDelete {
        thread: PartialGuildChannel,
    },
    ThreadListSync {
        thread_list_sync: ThreadListSyncEvent,
    },
    ThreadMemberUpdate {
        thread_member: ThreadMember,
    },
    ThreadMembersUpdate {
        thread_members_update: ThreadMembersUpdateEvent,
    },
    ThreadUpdate {
        thread: GuildChannel,
    },
    TypingStart {
        event: TypingStartEvent,
    },
    Unknown {
        name: String,
        raw: Value,
    },
    UserUpdate {
        old_data: CurrentUser,
        new: CurrentUser,
    },
    VoiceServerUpdate {
        update: VoiceServerUpdateEvent,
    },
    VoiceStateUpdate {
        old: Option<VoiceState>,
        new: VoiceState,
    },
    WebhookUpdate {
        guild_id: GuildId,
        belongs_to_channel_id: ChannelId,
    },
    InteractionCreate {
        interaction: Interaction,
    },
    // some variants omitted

}

This enum stores every possible event that a serenity::EventHandler can receive.

Passed to the stored callback by EventWrapper.

Variants

CacheReady

Fields

guilds: Vec<GuildId>

Available on crate feature cache only.

ChannelCreate

Fields

channel: &'a GuildChannel

CategoryCreate

Fields

category: &'a ChannelCategory

CategoryDelete

Fields

category: &'a ChannelCategory

ChannelDelete

Fields

channel: &'a GuildChannel

ChannelPinsUpdate

Fields

pin: ChannelPinsUpdateEvent

ChannelUpdate

Fields

old: Option<Channel>

new: Channel

Available on crate feature cache only.

GuildBanAddition

Fields

guild_id: GuildId

banned_user: User

GuildBanRemoval

Fields

guild_id: GuildId

unbanned_user: User

GuildCreate

Fields

guild: Guild

is_new: bool

Available on crate feature cache only.

GuildDelete

Fields

incomplete: UnavailableGuild

full: Option<Guild>

Available on crate feature cache only.

GuildEmojisUpdate

Fields

guild_id: GuildId

current_state: HashMap<EmojiId, Emoji>

GuildIntegrationsUpdate

Fields

guild_id: GuildId

GuildMemberAddition

Fields

new_member: Member

GuildMemberRemoval

Fields

guild_id: GuildId

user: User

member_data_if_available: Option<Member>

Available on crate feature cache only.

GuildMemberUpdate

Fields

old_if_available: Option<Member>

new: Member

Available on crate feature cache only.

GuildMembersChunk

Fields

chunk: GuildMembersChunkEvent

GuildRoleCreate

Fields

new: Role

GuildRoleDelete

Fields

guild_id: GuildId

removed_role_id: RoleId

removed_role_data_if_available: Option<Role>

Available on crate feature cache only.

GuildRoleUpdate

Fields

old_data_if_available: Option<Role>

new: Role

Available on crate feature cache only.

GuildStickersUpdate

Fields

guild_id: GuildId

current_state: HashMap<StickerId, Sticker>

GuildUnavailable

Fields

guild_id: GuildId

GuildUpdate

Fields

old_data_if_available: Option<Guild>

new_but_incomplete: PartialGuild

Available on crate feature cache only.

IntegrationCreate

Fields

integration: Integration

IntegrationUpdate

Fields

integration: Integration

IntegrationDelete

Fields

integration_id: IntegrationId

guild_id: GuildId

application_id: Option<ApplicationId>

InviteCreate

Fields

data: InviteCreateEvent

InviteDelete

Fields

data: InviteDeleteEvent

Message

Fields

new_message: Message

MessageDelete

Fields

channel_id: ChannelId

deleted_message_id: MessageId

guild_id: Option<GuildId>

MessageDeleteBulk

Fields

channel_id: ChannelId

multiple_deleted_messages_ids: Vec<MessageId>

guild_id: Option<GuildId>

MessageUpdate

Fields

old_if_available: Option<Message>

new: Option<Message>

event: MessageUpdateEvent

Available on crate feature cache only.

ReactionAdd

Fields

add_reaction: Reaction

ReactionRemove

Fields

removed_reaction: Reaction

ReactionRemoveAll

Fields

channel_id: ChannelId

removed_from_message_id: MessageId

PresenceReplace

Fields

new_presences: Vec<Presence>

PresenceUpdate

Fields

new_data: Presence

Ready

Fields

data_about_bot: Ready

Resume

Fields

event: ResumedEvent

ShardStageUpdate

Fields

update: ShardStageUpdateEvent

StageInstanceCreate

Fields

stage_instance: StageInstance

StageInstanceDelete

Fields

stage_instance: StageInstance

StageInstanceUpdate

Fields

stage_instance: StageInstance

ThreadCreate

Fields

thread: GuildChannel

ThreadDelete

Fields

thread: PartialGuildChannel

ThreadListSync

Fields

thread_list_sync: ThreadListSyncEvent

ThreadMemberUpdate

Fields

thread_member: ThreadMember

ThreadMembersUpdate

Fields

thread_members_update: ThreadMembersUpdateEvent

ThreadUpdate

Fields

thread: GuildChannel

TypingStart

Fields

event: TypingStartEvent

Unknown

Fields

name: String

raw: Value

UserUpdate

Fields

old_data: CurrentUser

new: CurrentUser

Available on crate feature cache only.

VoiceServerUpdate

Fields

update: VoiceServerUpdateEvent

VoiceStateUpdate

Fields

old: Option<VoiceState>

new: VoiceState

Available on crate feature cache only.

WebhookUpdate

Fields

guild_id: GuildId

belongs_to_channel_id: ChannelId

InteractionCreate

Fields

interaction: Interaction

Implementations

impl Event<'_>

pub fn name(&self) -> &'static str

Return the name of the event type

Examples found in repository?

examples/serenity_example_port/main.rs (line 170)

↕

158
159
160
161
162
163
164
165
166
167
168
169
170
171
172
173
174
175
176
177
178
179
180

async fn on_error(error: poise::FrameworkError<'_, Data, Error>) {
    match error {
        poise::FrameworkError::Command { error, ctx } => {
            println!(
                "Command '{}' returned error {:?}",
                ctx.command().name,
                error
            );
        }
        poise::FrameworkError::Listener { error, event, .. } => {
            println!(
                "Listener returned error during {:?} event: {:?}",
                event.name(),
                error
            );
        }
        error => {
            if let Err(e) = poise::builtins::on_error(error).await {
                println!("Error while handling error: {}", e)
            }
        }
    }
}

Hide additional examples

examples/framework_usage/main.rs (line 136)

↕

66
67
68
69
70
71
72
73
74
75
76
77
78
79
80
81
82
83
84
85
86
87
88
89
90
91
92
93
94
95
96
97
98
99
100
101
102
103
104
105
106
107
108
109
110
111
112
113
114
115
116
117
118
119
120
121
122
123
124
125
126
127
128
129
130
131
132
133
134
135
136
137
138
139
140
141
142
143
144
145
146
147
148
149
150
151
152
153
154
155
156
157
158
159

async fn main() {
    env_logger::init();

    let options = poise::FrameworkOptions {
        commands: vec![
            help(),
            register(),
            general::vote(),
            general::getvotes(),
            general::addmultiple(),
            general::choice(),
            general::boop(),
            general::voiceinfo(),
            general::test_reuse_response(),
            general::oracle(),
            general::code(),
            general::say(),
            general::file_details(),
            general::totalsize(),
            general::modal(),
            general::punish(),
            context_menu::user_info(),
            context_menu::echo(),
            autocomplete::greet(),
            checks::shutdown(),
            checks::modonly(),
            checks::delete(),
            checks::ferrisparty(),
            checks::add(),
            checks::get_guild_name(),
            checks::only_in_dms(),
            checks::lennyface(),
            checks::permissions_v2(),
            subcommands::parent(),
            localization::welcome(),
        ],
        prefix_options: poise::PrefixFrameworkOptions {
            prefix: Some("~".into()),
            edit_tracker: Some(poise::EditTracker::for_timespan(Duration::from_secs(3600))),
            additional_prefixes: vec![
                poise::Prefix::Literal("hey bot"),
                poise::Prefix::Literal("hey bot,"),
            ],
            ..Default::default()
        },
        /// The global error handler for all error cases that may occur
        on_error: |error| Box::pin(on_error(error)),
        /// This code is run before every command
        pre_command: |ctx| {
            Box::pin(async move {
                println!("Executing command {}...", ctx.command().qualified_name);
            })
        },
        /// This code is run after a command if it was successful (returned Ok)
        post_command: |ctx| {
            Box::pin(async move {
                println!("Executed command {}!", ctx.command().qualified_name);
            })
        },
        /// Every command invocation must pass this check to continue execution
        command_check: Some(|ctx| {
            Box::pin(async move {
                if ctx.author().id == 123456789 {
                    return Ok(false);
                }
                Ok(true)
            })
        }),
        listener: |_ctx, event, _framework, _data| {
            Box::pin(async move {
                println!("Got an event in listener: {:?}", event.name());
                Ok(())
            })
        },
        ..Default::default()
    };

    poise::Framework::builder()
        .token(var("TOKEN").expect("Missing `TOKEN` env var, see README for more information."))
        .user_data_setup(move |_ctx, _ready, _framework| {
            Box::pin(async move {
                Ok(Data {
                    votes: Mutex::new(HashMap::new()),
                })
            })
        })
        .options(options)
        .intents(
            serenity::GatewayIntents::non_privileged() | serenity::GatewayIntents::MESSAGE_CONTENT,
        )
        .run()
        .await
        .unwrap();
}

pub async fn dispatch(self, ctx: Context, handler: &dyn EventHandler)

Runs this event in the given serenity::EventHandler