Enum poise::FrameworkError

pub enum FrameworkError<'a, U, E> {

    Setup {
        error: E,
    },
    Listener {
        error: E,
        ctx: Context,
        event: &'a Event<'a>,
        framework: FrameworkContext<'a, U, E>,
    },
    Command {
        error: E,
        ctx: Context<'a, U, E>,
    },
    ArgumentParse {
        error: Box<dyn Error + Send + Sync>,
        input: Option<String>,
        ctx: Context<'a, U, E>,
    },
    CommandStructureMismatch {
        description: &'static str,
        ctx: ApplicationContext<'a, U, E>,
    },
    CooldownHit {
        remaining_cooldown: Duration,
        ctx: Context<'a, U, E>,
    },
    MissingBotPermissions {
        missing_permissions: Permissions,
        ctx: Context<'a, U, E>,
    },
    MissingUserPermissions {
        missing_permissions: Option<Permissions>,
        ctx: Context<'a, U, E>,
    },
    NotAnOwner {
        ctx: Context<'a, U, E>,
    },
    GuildOnly {
        ctx: Context<'a, U, E>,
    },
    DmOnly {
        ctx: Context<'a, U, E>,
    },
    NsfwOnly {
        ctx: Context<'a, U, E>,
    },
    CommandCheckFailed {
        error: Option<E>,
        ctx: Context<'a, U, E>,
    },
    DynamicPrefix {
        error: E,
    },
    // some variants omitted

}

Any error that can occur while the bot runs. Either thrown by user code (those variants will have an error field with your error type E in it), or originating from within the framework.

These errors are handled with the crate::FrameworkOptions::on_error callback

Variants

Setup

Fields

error: E

Error which was thrown in the setup code

User code threw an error in user data setup

Listener

Fields

error: E

Error which was thrown in the listener code

ctx: Context

The serenity Context passed to the event

event: &'a Event<'a>

Which event was being processed when the error occurred

framework: FrameworkContext<'a, U, E>

The Framework passed to the event

User code threw an error in generic event listener

Command

Fields

error: E

Error which was thrown in the command code

ctx: Context<'a, U, E>

General context

Error occured during command execution

ArgumentParse

Fields

error: Box<dyn Error + Send + Sync>

Error which was thrown by the parameter type’s parsing routine

input: Option<String>

If applicable, the input on which parsing failed

ctx: Context<'a, U, E>

General context

A command argument failed to parse from the Discord message or interaction content

CommandStructureMismatch

Fields

description: &'static str

Developer-readable description of the type mismatch

ctx: ApplicationContext<'a, U, E>

General context

Expected a certain argument type at a certain position in the unstructured list of arguments, but found something else.

Most often the result of the bot not having registered the command in Discord, so Discord stores an outdated version of the command and its parameters.

CooldownHit

Fields

remaining_cooldown: Duration

Time until the command may be invoked for the next time in the given context

ctx: Context<'a, U, E>

General context

Command was invoked before its cooldown expired

MissingBotPermissions

Fields

missing_permissions: Permissions

Which permissions in particular the bot is lacking for this command

ctx: Context<'a, U, E>

General context

Command was invoked but the bot is lacking the permissions specified in crate::Command::required_bot_permissions

MissingUserPermissions

Fields

missing_permissions: Option<Permissions>

List of permissions that the user is lacking. May be None if retrieving the user’s permissions failed

ctx: Context<'a, U, E>

General context

Command was invoked but the user is lacking the permissions specified in crate::Command::required_bot_permissions

NotAnOwner

Fields

ctx: Context<'a, U, E>

General context

A non-owner tried to invoke an owners-only command

GuildOnly

Fields

ctx: Context<'a, U, E>

General context

Command was invoked but the channel was a DM channel

DmOnly

Fields

ctx: Context<'a, U, E>

General context

Command was invoked but the channel was a non-DM channel

NsfwOnly

Fields

ctx: Context<'a, U, E>

General context

Command was invoked but the channel wasn’t a NSFW channel

CommandCheckFailed

Fields

error: Option<E>

If execution wasn’t aborted because of an error but because it successfully returned false, this field is None

ctx: Context<'a, U, E>

General context

Provided pre-command check either errored, or returned false, so command execution aborted

DynamicPrefix

Fields

error: E

Error which was thrown in the dynamic prefix code

crate::PrefixFrameworkOptions::dynamic_prefix or crate::PrefixFrameworkOptions::stripped_dynamic_prefix returned an error