Enum poise::event::Event

pub enum Event<'a> {

    CacheReady {
        guilds: Vec<GuildId>,
    },
    ChannelCreate {
        channel: &'a GuildChannel,
    },
    CategoryCreate {
        category: &'a ChannelCategory,
    },
    CategoryDelete {
        category: &'a ChannelCategory,
    },
    ChannelDelete {
        channel: &'a GuildChannel,
    },
    ChannelPinsUpdate {
        pin: ChannelPinsUpdateEvent,
    },
    ChannelUpdate {
        old: Option<Channel>,
        new: Channel,
    },
    GuildBanAddition {
        guild_id: GuildId,
        banned_user: User,
    },
    GuildBanRemoval {
        guild_id: GuildId,
        unbanned_user: User,
    },
    GuildCreate {
        guild: Guild,
        is_new: bool,
    },
    GuildDelete {
        incomplete: UnavailableGuild,
        full: Option<Guild>,
    },
    GuildEmojisUpdate {
        guild_id: GuildId,
        current_state: HashMap<EmojiId, Emoji>,
    },
    GuildIntegrationsUpdate {
        guild_id: GuildId,
    },
    GuildMemberAddition {
        new_member: Member,
    },
    GuildMemberRemoval {
        guild_id: GuildId,
        user: User,
        member_data_if_available: Option<Member>,
    },
    GuildMemberUpdate {
        old_if_available: Option<Member>,
        new: Member,
    },
    GuildMembersChunk {
        chunk: GuildMembersChunkEvent,
    },
    GuildRoleCreate {
        new: Role,
    },
    GuildRoleDelete {
        guild_id: GuildId,
        removed_role_id: RoleId,
        removed_role_data_if_available: Option<Role>,
    },
    GuildRoleUpdate {
        old_data_if_available: Option<Role>,
        new: Role,
    },
    GuildStickersUpdate {
        guild_id: GuildId,
        current_state: HashMap<StickerId, Sticker>,
    },
    GuildUnavailable {
        guild_id: GuildId,
    },
    GuildUpdate {
        old_data_if_available: Option<Guild>,
        new_but_incomplete: PartialGuild,
    },
    IntegrationCreate {
        integration: Integration,
    },
    IntegrationUpdate {
        integration: Integration,
    },
    IntegrationDelete {
        integration_id: IntegrationId,
        guild_id: GuildId,
        application_id: Option<ApplicationId>,
    },
    InviteCreate {
        data: InviteCreateEvent,
    },
    InviteDelete {
        data: InviteDeleteEvent,
    },
    Message {
        new_message: Message,
    },
    MessageDelete {
        channel_id: ChannelId,
        deleted_message_id: MessageId,
        guild_id: Option<GuildId>,
    },
    MessageDeleteBulk {
        channel_id: ChannelId,
        multiple_deleted_messages_ids: Vec<MessageId>,
        guild_id: Option<GuildId>,
    },
    MessageUpdate {
        old_if_available: Option<Message>,
        new: Option<Message>,
        event: MessageUpdateEvent,
    },
    ReactionAdd {
        add_reaction: Reaction,
    },
    ReactionRemove {
        removed_reaction: Reaction,
    },
    ReactionRemoveAll {
        channel_id: ChannelId,
        removed_from_message_id: MessageId,
    },
    PresenceReplace {
        new_presences: Vec<Presence>,
    },
    PresenceUpdate {
        new_data: Presence,
    },
    Ready {
        data_about_bot: Ready,
    },
    Resume {
        event: ResumedEvent,
    },
    ShardStageUpdate {
        update: ShardStageUpdateEvent,
    },
    StageInstanceCreate {
        stage_instance: StageInstance,
    },
    StageInstanceDelete {
        stage_instance: StageInstance,
    },
    StageInstanceUpdate {
        stage_instance: StageInstance,
    },
    ThreadCreate {
        thread: GuildChannel,
    },
    ThreadDelete {
        thread: PartialGuildChannel,
    },
    ThreadListSync {
        thread_list_sync: ThreadListSyncEvent,
    },
    ThreadMemberUpdate {
        thread_member: ThreadMember,
    },
    ThreadMembersUpdate {
        thread_members_update: ThreadMembersUpdateEvent,
    },
    ThreadUpdate {
        thread: GuildChannel,
    },
    TypingStart {
        event: TypingStartEvent,
    },
    Unknown {
        name: String,
        raw: Value,
    },
    UserUpdate {
        old_data: CurrentUser,
        new: CurrentUser,
    },
    VoiceServerUpdate {
        update: VoiceServerUpdateEvent,
    },
    VoiceStateUpdate {
        old: Option<VoiceState>,
        new: VoiceState,
    },
    WebhookUpdate {
        guild_id: GuildId,
        belongs_to_channel_id: ChannelId,
    },
    InteractionCreate {
        interaction: Interaction,
    },
    // some variants omitted

}

This enum stores every possible event that a serenity::EventHandler can receive.

Passed to the stored callback by EventWrapper.

Variants

CacheReady

Fields

guilds: Vec<GuildId>

Available on crate feature cache only.

ChannelCreate

Fields

channel: &'a GuildChannel

CategoryCreate

Fields

category: &'a ChannelCategory

CategoryDelete

Fields

category: &'a ChannelCategory

ChannelDelete

Fields

channel: &'a GuildChannel

ChannelPinsUpdate

Fields

pin: ChannelPinsUpdateEvent

ChannelUpdate

Fields

old: Option<Channel>

new: Channel

Available on crate feature cache only.

GuildBanAddition

Fields

guild_id: GuildId

banned_user: User

GuildBanRemoval

Fields

guild_id: GuildId

unbanned_user: User

GuildCreate

Fields

guild: Guild

is_new: bool

Available on crate feature cache only.

GuildDelete

Fields

incomplete: UnavailableGuild

full: Option<Guild>

Available on crate feature cache only.

GuildEmojisUpdate

Fields

guild_id: GuildId

current_state: HashMap<EmojiId, Emoji>

GuildIntegrationsUpdate

Fields

guild_id: GuildId

GuildMemberAddition

Fields

new_member: Member

GuildMemberRemoval

Fields

guild_id: GuildId

user: User

member_data_if_available: Option<Member>

Available on crate feature cache only.

GuildMemberUpdate

Fields

old_if_available: Option<Member>

new: Member

Available on crate feature cache only.

GuildMembersChunk

Fields

chunk: GuildMembersChunkEvent

GuildRoleCreate

Fields

new: Role

GuildRoleDelete

Fields

guild_id: GuildId

removed_role_id: RoleId

removed_role_data_if_available: Option<Role>

Available on crate feature cache only.

GuildRoleUpdate

Fields

old_data_if_available: Option<Role>

new: Role

Available on crate feature cache only.

GuildStickersUpdate

Fields

guild_id: GuildId

current_state: HashMap<StickerId, Sticker>

GuildUnavailable

Fields

guild_id: GuildId

GuildUpdate

Fields

old_data_if_available: Option<Guild>

new_but_incomplete: PartialGuild

Available on crate feature cache only.

IntegrationCreate

Fields

integration: Integration

IntegrationUpdate

Fields

integration: Integration

IntegrationDelete

Fields

integration_id: IntegrationId

guild_id: GuildId

application_id: Option<ApplicationId>

InviteCreate

Fields

data: InviteCreateEvent

InviteDelete

Fields

data: InviteDeleteEvent

Message

Fields

new_message: Message

MessageDelete

Fields

channel_id: ChannelId

deleted_message_id: MessageId

guild_id: Option<GuildId>

MessageDeleteBulk

Fields

channel_id: ChannelId

multiple_deleted_messages_ids: Vec<MessageId>

guild_id: Option<GuildId>

MessageUpdate

Fields

old_if_available: Option<Message>

new: Option<Message>

event: MessageUpdateEvent

Available on crate feature cache only.

ReactionAdd

Fields

add_reaction: Reaction

ReactionRemove

Fields

removed_reaction: Reaction

ReactionRemoveAll

Fields

channel_id: ChannelId

removed_from_message_id: MessageId

PresenceReplace

Fields

new_presences: Vec<Presence>

PresenceUpdate

Fields

new_data: Presence

Ready

Fields

data_about_bot: Ready

Resume

Fields

event: ResumedEvent

ShardStageUpdate

Fields

update: ShardStageUpdateEvent

StageInstanceCreate

Fields

stage_instance: StageInstance

StageInstanceDelete

Fields

stage_instance: StageInstance

StageInstanceUpdate

Fields

stage_instance: StageInstance

ThreadCreate

Fields

thread: GuildChannel

ThreadDelete

Fields

thread: PartialGuildChannel

ThreadListSync

Fields

thread_list_sync: ThreadListSyncEvent

ThreadMemberUpdate

Fields

thread_member: ThreadMember

ThreadMembersUpdate

Fields

thread_members_update: ThreadMembersUpdateEvent

ThreadUpdate

Fields

thread: GuildChannel

TypingStart

Fields

event: TypingStartEvent

Unknown

Fields

name: String

raw: Value

UserUpdate

Fields

old_data: CurrentUser

new: CurrentUser

Available on crate feature cache only.

VoiceServerUpdate

Fields

update: VoiceServerUpdateEvent

VoiceStateUpdate

Fields

old: Option<VoiceState>

new: VoiceState

Available on crate feature cache only.

WebhookUpdate

Fields

guild_id: GuildId

belongs_to_channel_id: ChannelId

InteractionCreate

Fields

interaction: Interaction

Implementations

impl Event<'_>

pub fn name(&self) -> &'static str

Return the name of the event type

Examples found in repository?

src/structs/framework_error.rs (line 244)

231
232
233
234
235
236
237
238
239
240
241
242
243
244
245
246
247
248
249
250
251
252
253
254
255
256
257
258
259
260
261
262
263
264
265
266
267
268
269
270
271
272
273
274
275
276
277
278
279
280
281
282
283
284
285
286
287
288
289
290
291
292
293
294
295
296
297
298
299
300
301
302
303
304
305
306
307
308
309
310
311
312
313
314
315
316
317
318
319
320
321
322
323
324
325
326
327
328
329
330
331
332
333
334
335

    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
        match self {
            Self::Setup {
                error: _,
                framework: _,
                data_about_bot: _,
                ctx: _,
            } => write!(f, "poise setup error"),
            Self::EventHandler {
                error: _,
                ctx: _,
                event,
                framework: _,
            } => write!(f, "error in {} event event handler", event.name()),
            Self::Command { error: _, ctx } => {
                write!(f, "error in command `{}`", full_command_name!(ctx))
            }
            Self::ArgumentParse {
                error: _,
                input,
                ctx,
            } => write!(
                f,
                "failed to parse argument in command `{}` on input {:?}",
                full_command_name!(ctx),
                input
            ),
            Self::CommandStructureMismatch { description, ctx } => write!(
                f,
                "unexpected application command structure in command `{}`: {}",
                full_command_name!(crate::Context::Application(*ctx)),
                description
            ),
            Self::CooldownHit {
                remaining_cooldown,
                ctx,
            } => write!(
                f,
                "cooldown hit in command `{}` ({:?} remaining)",
                full_command_name!(ctx),
                remaining_cooldown
            ),
            Self::MissingBotPermissions {
                missing_permissions,
                ctx,
            } => write!(
                f,
                "bot is missing permisions ({}) to execute command `{}`",
                missing_permissions,
                full_command_name!(ctx),
            ),
            Self::MissingUserPermissions {
                missing_permissions,
                ctx,
            } => write!(
                f,
                "user is or may be missing permisions ({:?}) to execute command `{}`",
                missing_permissions,
                full_command_name!(ctx),
            ),
            Self::NotAnOwner { ctx } => write!(
                f,
                "owner-only command `{}` cannot be run by non-owners",
                full_command_name!(ctx)
            ),
            Self::GuildOnly { ctx } => write!(
                f,
                "guild-only command `{}` cannot run in DMs",
                full_command_name!(ctx)
            ),
            Self::DmOnly { ctx } => write!(
                f,
                "DM-only command `{}` cannot run in guilds",
                full_command_name!(ctx)
            ),
            Self::NsfwOnly { ctx } => write!(
                f,
                "nsfw-only command `{}` cannot run in non-nsfw channels",
                full_command_name!(ctx)
            ),
            Self::CommandCheckFailed { error: _, ctx } => write!(
                f,
                "pre-command check for command `{}` either denied access or errored",
                full_command_name!(ctx)
            ),
            Self::DynamicPrefix {
                error: _,
                ctx: _,
                msg,
            } => {
                write!(
                    f,
                    "dynamic prefix callback errored on message {:?}",
                    msg.content
                )
            }
            Self::UnknownCommand { msg_content, .. } => {
                write!(f, "unknown command `{}`", msg_content)
            }
            Self::UnknownInteraction { interaction, .. } => {
                write!(f, "unknown interaction `{}`", interaction.data().name)
            }
            Self::__NonExhaustive => unreachable!(),
        }
    }

Hide additional examples

src/builtins/mod.rs (line 38)

29
30
31
32
33
34
35
36
37
38
39
40
41
42
43
44
45
46
47
48
49
50
51
52
53
54
55
56
57
58
59
60
61
62
63
64
65
66
67
68
69
70
71
72
73
74
75
76
77
78
79
80
81
82
83
84
85
86
87
88
89
90
91
92
93
94
95
96
97
98
99
100
101
102
103
104
105
106
107
108
109
110
111
112
113
114
115
116
117
118
119
120
121
122
123
124
125
126
127
128
129
130
131
132
133
134
135
136
137
138
139
140
141
142
143
144
145
146
147
148
149
150
151
152
153
154
155
156
157
158
159
160
161

pub async fn on_error<U, E: std::fmt::Display + std::fmt::Debug>(
    error: crate::FrameworkError<'_, U, E>,
) -> Result<(), serenity::Error> {
    match error {
        crate::FrameworkError::Setup { error, .. } => {
            log::error!("Error in user data setup: {}", error);
        }
        crate::FrameworkError::EventHandler { error, event, .. } => log::error!(
            "User event event handler encountered an error on {} event: {}",
            event.name(),
            error
        ),
        crate::FrameworkError::Command { ctx, error } => {
            let error = error.to_string();
            ctx.say(error).await?;
        }
        crate::FrameworkError::ArgumentParse { ctx, input, error } => {
            // If we caught an argument parse error, give a helpful error message with the
            // command explanation if available
            let usage = match ctx.command().help_text {
                Some(help_text) => help_text(),
                None => "Please check the help menu for usage information".into(),
            };
            let response = if let Some(input) = input {
                format!(
                    "**Cannot parse `{}` as argument: {}**\n{}",
                    input, error, usage
                )
            } else {
                format!("**{}**\n{}", error, usage)
            };
            ctx.say(response).await?;
        }
        crate::FrameworkError::CommandStructureMismatch { ctx, description } => {
            log::error!(
                "Error: failed to deserialize interaction arguments for `/{}`: {}",
                ctx.command.name,
                description,
            );
        }
        crate::FrameworkError::CommandCheckFailed { ctx, error } => {
            log::error!(
                "A command check failed in command {} for user {}: {:?}",
                ctx.command().name,
                ctx.author().name,
                error,
            );
        }
        crate::FrameworkError::CooldownHit {
            remaining_cooldown,
            ctx,
        } => {
            let msg = format!(
                "You're too fast. Please wait {} seconds before retrying",
                remaining_cooldown.as_secs()
            );
            ctx.send(|b| b.content(msg).ephemeral(true)).await?;
        }
        crate::FrameworkError::MissingBotPermissions {
            missing_permissions,
            ctx,
        } => {
            let msg = format!(
                "Command cannot be executed because the bot is lacking permissions: {}",
                missing_permissions,
            );
            ctx.send(|b| b.content(msg).ephemeral(true)).await?;
        }
        crate::FrameworkError::MissingUserPermissions {
            missing_permissions,
            ctx,
        } => {
            let response = if let Some(missing_permissions) = missing_permissions {
                format!(
                    "You're lacking permissions for `{}{}`: {}",
                    ctx.prefix(),
                    ctx.command().name,
                    missing_permissions,
                )
            } else {
                format!(
                    "You may be lacking permissions for `{}{}`. Not executing for safety",
                    ctx.prefix(),
                    ctx.command().name,
                )
            };
            ctx.send(|b| b.content(response).ephemeral(true)).await?;
        }
        crate::FrameworkError::NotAnOwner { ctx } => {
            let response = "Only bot owners can call this command";
            ctx.send(|b| b.content(response).ephemeral(true)).await?;
        }
        crate::FrameworkError::GuildOnly { ctx } => {
            let response = "You cannot run this command in DMs.";
            ctx.send(|b| b.content(response).ephemeral(true)).await?;
        }
        crate::FrameworkError::DmOnly { ctx } => {
            let response = "You cannot run this command outside DMs.";
            ctx.send(|b| b.content(response).ephemeral(true)).await?;
        }
        crate::FrameworkError::NsfwOnly { ctx } => {
            let response = "You cannot run this command outside NSFW channels.";
            ctx.send(|b| b.content(response).ephemeral(true)).await?;
        }
        crate::FrameworkError::DynamicPrefix { error, msg, .. } => {
            log::error!(
                "Dynamic prefix failed for message {:?}: {}",
                msg.content,
                error
            );
        }
        crate::FrameworkError::UnknownCommand {
            msg_content,
            prefix,
            ..
        } => {
            log::warn!(
                "Recognized prefix `{}`, but didn't recognize command name in `{}`",
                prefix,
                msg_content,
            );
        }
        crate::FrameworkError::UnknownInteraction { interaction, .. } => {
            log::warn!(
                "received unknown interaction \"{}\"",
                interaction.data().name
            );
        }
        crate::FrameworkError::__NonExhaustive => panic!(),
    }

    Ok(())
}

pub async fn dispatch(self, ctx: Context, handler: &dyn EventHandler)

Runs this event in the given serenity::EventHandler

Examples found in repository?

src/framework/mod.rs (line 106)

71
72
73
74
75
76
77
78
79
80
81
82
83
84
85
86
87
88
89
90
91
92
93
94
95
96
97
98
99
100
101
102
103
104
105
106
107
108
109
110
111
112
113
114
115
116
117
118
119
120
121
122
123

    pub async fn new<F>(
        client_builder: serenity::ClientBuilder,
        setup: F,
        mut options: crate::FrameworkOptions<U, E>,
    ) -> Result<std::sync::Arc<Self>, serenity::Error>
    where
        F: Send
            + Sync
            + 'static
            + for<'a> FnOnce(
                &'a serenity::Context,
                &'a serenity::Ready,
                &'a Self,
            ) -> BoxFuture<'a, Result<U, E>>,
        U: Send + Sync + 'static,
        E: Send + 'static,
    {
        use std::sync::{Arc, Mutex};

        set_qualified_names(&mut options.commands);
        message_content_intent_sanity_check(&options.prefix_options, client_builder.get_intents());

        let framework_cell = Arc::new(once_cell::sync::OnceCell::<Arc<Self>>::new());
        let framework_cell_2 = framework_cell.clone();
        let existing_event_handler = client_builder.get_event_handler();
        let event_handler = crate::EventWrapper(move |ctx, event| {
            // unwrap_used: we will only receive events once the client has been started, by which
            // point framework_cell has been initialized
            #[allow(clippy::unwrap_used)]
            let framework = framework_cell_2.get().unwrap().clone();
            let existing_event_handler = existing_event_handler.clone();

            Box::pin(async move {
                raw_dispatch_event(&*framework, &ctx, &event).await;
                if let Some(handler) = existing_event_handler {
                    event.dispatch(ctx, &*handler).await;
                }
            }) as _
        });

        let client: serenity::Client = client_builder.event_handler(event_handler).await?;

        let framework = Arc::new(Self {
            user_data: once_cell::sync::OnceCell::new(),
            bot_id: once_cell::sync::OnceCell::new(),
            setup: Mutex::new(Some(Box::new(setup))),
            options,
            shard_manager: client.shard_manager.clone(),
            client: parking_lot::Mutex::new(Some(client)),
        });
        let _: Result<_, _> = framework_cell.set(framework.clone());
        Ok(framework)
    }