Struct poise::event::EventWrapper

pub struct EventWrapper<F>(pub F)
where
    F: Send + Sync + for<'a> Fn(Context, Event<'a>) -> BoxFuture<'a, ()>;

A serenity::EventHandler implementation that wraps every received event into the Event enum and propagates it to a callback.

Packaging every event into a singular type can make it easier to pass around and process.

use serenity::prelude::*;

let handler = poise::EventWrapper(|ctx, event| Box::pin(async move {
    println!("Received an event: {:?}", event);
}));

Client::builder(token, intents)
    .event_handler(handler)
    // ...

Tuple Fields
0: F