Function poise::find_command

pub fn find_command<'a, U, E>(
    commands: &'a [Command<U, E>],
    remaining_message: &'a str,
    case_insensitive: bool
) -> Option<(&'a Command<U, E>, &'a str, &'a str)>where
    U: Send + Sync,

Find a command or subcommand within &[Command], given a command invocation without a prefix. Returns the verbatim command name string as well as the command arguments (i.e. the remaining string).

The API must be like this (as opposed to just taking the command name upfront) because of subcommands.

#[poise::command(prefix_command)]
async fn command1(ctx: poise::Context<'_, (), ()>) -> Result<(), ()> { Ok(()) }
#[poise::command(prefix_command, subcommands("command3"))]
async fn command2(ctx: poise::Context<'_, (), ()>) -> Result<(), ()> { Ok(()) }
#[poise::command(prefix_command)]
async fn command3(ctx: poise::Context<'_, (), ()>) -> Result<(), ()> { Ok(()) }
let commands = vec![command1(), command2()];

assert_eq!(
    poise::find_command(&commands, "command1 my arguments", false),
    Some((&commands[0], "command1", "my arguments")),
);
assert_eq!(
    poise::find_command(&commands, "command2 command3 my arguments", false),
    Some((&commands[1].subcommands[0], "command3", "my arguments")),
);
assert_eq!(
    poise::find_command(&commands, "CoMmAnD2 cOmMaNd99 my arguments", true),
    Some((&commands[1], "CoMmAnD2", "cOmMaNd99 my arguments")),
);