Function poise::send_reply

pub async fn send_reply<'a, U, E>(
    ctx: Context<'_, U, E>, 
    builder: impl for<'b> FnOnce(&'b mut CreateReply<'a>) -> &'b mut CreateReply<'a>
) -> Result<Option<ReplyHandle<'_>>, Error>

Send a message in the given context: normal message if prefix command, interaction response if application command.

If you just want to send a string, use say_reply.

ctx.send(|f| f
    .content("Works for slash and prefix commands")
    .embed(|f| f
        .title("Much versatile, very wow")
        .description("I need more documentation ok?")
    )
    .ephemeral(true) // this one only applies in application commands though
).await?;