Function poise::framework::insert_owners_from_http

pub async fn insert_owners_from_http(
    token: &str,
    owners: &mut HashSet<UserId>
) -> Result<(), Error>

Runs serenity::Http::get_current_application_info and inserts owner data into crate::FrameworkOptions::owners

Examples found in repository?

src/framework/builder.rs (line 213)

187
188
189
190
191
192
193
194
195
196
197
198
199
200
201
202
203
204
205
206
207
208
209
210
211
212
213
214
215
216
217
218
219
220
221
222
223
224
225
226

    pub async fn build(self) -> Result<std::sync::Arc<crate::Framework<U, E>>, serenity::Error>
    where
        U: Send + Sync + 'static,
        E: Send + 'static,
    {
        // Aggregate required values or panic if not provided
        // WHEN CHANGING THESE, UPDATE FrameworkBuilder docs!
        let token = self.token.expect("No token was provided to the framework");
        let intents = self.intents.expect(
            "

No gateway intents were provided to the framework via `FrameworkBuilder::intents()`. If you're \
unsure, use
`serenity::GatewayIntents::non_privileged() | serenity::GatewayIntents::MESSAGE_CONTENT`
and enable MESSAGE_CONTENT in your Discord bot dashboard

",
        );
        let setup = self
            .setup
            .expect("No user data setup function was provided to the framework");
        let mut options = self.options.expect("No framework options provided");

        // Build framework options by concatenating user-set options with commands and owners
        options.commands.extend(self.commands);
        if self.initialize_owners {
            if let Err(e) = super::insert_owners_from_http(&token, &mut options.owners).await {
                log::warn!("Failed to insert owners from HTTP: {}", e);
            }
        }

        // Create serenity client
        let mut client_builder = serenity::ClientBuilder::new(token, intents);
        if let Some(client_settings) = self.client_settings {
            client_builder = client_settings(client_builder);
        }

        // Create framework with specified settings
        crate::Framework::new(client_builder, setup, options).await
    }