poise

Module modal

Modal trait and utility items for implementing it (mainly for the derive macro)

Traits

Modal

Derivable trait for modal interactions, Discords version of interactive forms

Functions

execute_modal

Convenience function for showing the modal and waiting for a response.

execute_modal_on_component_interaction

Convenience function for showing the modal on a message interaction and waiting for a response.