Function poise::modal::execute_modal

pub async fn execute_modal<U: Send + Sync, E, M: Modal>(
    ctx: ApplicationContext<'_, U, E>,
    defaults: Option<M>,
    timeout: Option<Duration>
) -> Result<Option<M>, Error>

Convenience function for showing the modal and waiting for a response.

If the user doesn’t submit before the timeout expires, None is returned.

Note: a modal must be the first response to a command. You cannot send any messages before, or the modal will fail.

This function:

sends the modal via Modal::create()

waits for the user to submit via serenity::CollectModalInteraction

acknowledges the submitted data so that Discord closes the pop-up for the user

parses the submitted data via Modal::parse(), wrapping errors in serenity::Error::Other

If you need more specialized behavior, you can copy paste the implementation of this function and adjust to your needs. The code of this function is just a starting point.

Examples found in repository?

src/modal.rs (line 149)

146
147
148
149
150
151
152
153
154
155
156
157
158
159

    async fn execute<U: Send + Sync, E>(
        ctx: crate::ApplicationContext<'_, U, E>,
    ) -> Result<Option<Self>, serenity::Error> {
        execute_modal(ctx, None::<Self>, None).await
    }

    /// Calls `execute_modal(ctx, Some(defaults), None)`. See [`execute_modal`]
    // TODO: deprecate this in favor of execute_modal()?
    async fn execute_with_defaults<U: Send + Sync, E>(
        ctx: crate::ApplicationContext<'_, U, E>,
        defaults: Self,
    ) -> Result<Option<Self>, serenity::Error> {
        execute_modal(ctx, Some(defaults), None).await
    }