poise::modal

Function execute_modal_on_component_interaction

pub async fn execute_modal_on_component_interaction<M: Modal>(
    ctx: impl AsRef<Context>,
    interaction: ComponentInteraction,
    defaults: Option<M>,
    timeout: Option<Duration>,
) -> Result<Option<M>, Error>

Convenience function for showing the modal on a message interaction and waiting for a response.

If the user doesn’t submit before the timeout expires, None is returned.

This function:

sends the modal via Modal::create() as a mci interaction response

waits for the user to submit via serenity::ModalInteractionCollector

acknowledges the submitted data so that Discord closes the pop-up for the user

parses the submitted data via Modal::parse(), wrapping errors in serenity::Error::Other

If you need more specialized behavior, you can copy paste the implementation of this function and adjust to your needs. The code of this function is just a starting point.

Examples found in repository?

examples/feature_showcase/modal.rs (line 46)

25pub async fn component_modal(ctx: crate::Context<'_>) -> Result<(), Error> {
26    let reply = {
27        let components = vec![serenity::CreateActionRow::Buttons(vec![
28            serenity::CreateButton::new("open_modal")
29                .label("Open modal")
30                .style(poise::serenity_prelude::ButtonStyle::Success),
31        ])];
32
33        poise::CreateReply::default()
34            .content("Click the button below to open the modal")
35            .components(components)
36    };
37
38    ctx.send(reply).await?;
39
40    while let Some(mci) = serenity::ComponentInteractionCollector::new(ctx.serenity_context())
41        .timeout(std::time::Duration::from_secs(120))
42        .filter(move |mci| mci.data.custom_id == "open_modal")
43        .await
44    {
45        let data =
46            poise::execute_modal_on_component_interaction::<MyModal>(ctx, mci, None, None).await?;
47        println!("Got data: {:?}", data);
48    }
49    Ok(())
50}