Trait poise::modal::Modal

pub trait Modal: Sized {
    fn create(
        defaults: Option<Self>,
        custom_id: String
    ) -> CreateInteractionResponse<'static>;
    fn parse(data: ModalSubmitInteractionData) -> Result<Self, &'static str>;

    fn execute<'life0, 'async_trait, U, E>(
        ctx: ApplicationContext<'life0, U, E>
    ) -> Pin<Box<dyn Future<Output = Result<Option<Self>, Error>> + Send + 'async_trait>>
    where
        U: 'async_trait + Send + Sync,
        E: 'async_trait,
        'life0: 'async_trait,
        Self: Send + 'async_trait,
    { ... }
    fn execute_with_defaults<'life0, 'async_trait, U, E>(
        ctx: ApplicationContext<'life0, U, E>,
        defaults: Self
    ) -> Pin<Box<dyn Future<Output = Result<Option<Self>, Error>> + Send + 'async_trait>>
    where
        U: 'async_trait + Send + Sync,
        E: 'async_trait,
        'life0: 'async_trait,
        Self: Send + 'async_trait,
    { ... }
}

Derivable trait for modal interactions, Discords version of interactive forms

You don’t need to implement this trait manually; use #[derive(poise::Modal)] instead

Example

use poise::Modal;
type ApplicationContext<'a> = poise::ApplicationContext<'a, Data, Error>;

#[derive(Debug, Modal)]
#[name = "Modal title"] // Struct name by default
struct MyModal {
    #[name = "First input label"] // Field name by default
    #[placeholder = "Your first input goes here"] // No placeholder by default
    #[min_length = 5] // No length restriction by default (so, 1-4000 chars)
    #[max_length = 500]
    first_input: String,
    #[name = "Second input label"]
    #[paragraph] // Switches from single-line input to multiline text box
    second_input: Option<String>, // Option means optional input
}

#[poise::command(slash_command)]
pub async fn modal(ctx: ApplicationContext<'_>) -> Result<(), Error> {
    let data = MyModal::execute(ctx).await?;
    println!("Got data: {:?}", data);

    Ok(())
}

Required Methods

fn create(
defaults: Option<Self>,
custom_id: String
) -> CreateInteractionResponse<'static>

Returns an interaction response builder which creates the modal for this type

Optionally takes an initialized instance as pre-filled values of this modal (see Self::execute_with_defaults() for more info)

fn parse(data: ModalSubmitInteractionData) -> Result<Self, &'static str>

Parses a received modal submit interaction into this type

Returns an error if a field was missing. This should never happen, because Discord will only let users submit when all required fields are filled properly

Provided Methods

fn execute<'life0, 'async_trait, U, E>(
ctx: ApplicationContext<'life0, U, E>
) -> Pin<Box<dyn Future<Output = Result<Option<Self>, Error>> + Send + 'async_trait>>where
U: 'async_trait + Send + Sync,
E: 'async_trait,
'life0: 'async_trait,
Self: Send + 'async_trait,

Calls execute_modal(ctx, None, None). See execute_modal

Examples found in repository?

examples/framework_usage/commands/general.rs (line 290)

285
286
287
288
289
290
291
292
293
294

pub async fn modal(
    ctx: poise::ApplicationContext<'_, crate::Data, crate::Error>,
) -> Result<(), Error> {
    use poise::Modal as _;

    let data = MyModal::execute(ctx).await?;
    println!("Got data: {:?}", data);

    Ok(())
}

fn execute_with_defaults<'life0, 'async_trait, U, E>(
ctx: ApplicationContext<'life0, U, E>,
defaults: Self
) -> Pin<Box<dyn Future<Output = Result<Option<Self>, Error>> + Send + 'async_trait>>where
U: 'async_trait + Send + Sync,
E: 'async_trait,
'life0: 'async_trait,
Self: Send + 'async_trait,

Calls execute_modal(ctx, Some(defaults), None). See execute_modal