poise::prefix_argument

Trait PopArgument

pub trait PopArgument<'a>: Sized {
    // Required method
    fn pop_from<'life0, 'life1, 'async_trait>(
        args: &'a str,
        attachment_index: usize,
        ctx: &'life0 Context,
        msg: &'life1 Message,
    ) -> Pin<Box<dyn Future<Output = Result<(&'a str, usize, Self), (Box<dyn Error + Send + Sync>, Option<String>)>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'a: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait;
}

Parse a value out of a string by popping off the front of the string. Discord message context is available for parsing, and IO may be done as part of the parsing.

Implementors should assume that a string never starts with whitespace, and fail to parse if it does. This is for consistency’s sake and also because it keeps open the possibility of parsing whitespace.

Similar in spirit to std::str::FromStr.

Required Methods

fn pop_from<'life0, 'life1, 'async_trait>( args: &'a str, attachment_index: usize, ctx: &'life0 Context, msg: &'life1 Message, ) -> Pin<Box<dyn Future<Output = Result<(&'a str, usize, Self), (Box<dyn Error + Send + Sync>, Option<String>)>> + Send + 'async_trait>>
where Self: 'async_trait, 'a: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait,

Pops an argument from the args string, and parses it as Self.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety".