Function poise::reply::send_reply

pub async fn send_reply<'att, U, E>(
    ctx: Context<'_, U, E>,
    builder: impl for<'a> FnOnce(&'a mut CreateReply<'att>) -> &'a mut CreateReply<'att>
) -> Result<ReplyHandle<'_>, Error>

Send a message in the given context: normal message if prefix command, interaction response if application command.

If you just want to send a string, use say_reply.

Note: panics when called in an autocomplete context!

ctx.send(|f| f
    .content("Works for slash and prefix commands")
    .embed(|f| f
        .title("Much versatile, very wow")
        .description("I need more documentation ok?")
    )
    .ephemeral(true) // this one only applies in application commands though
).await?;

Examples found in repository?

src/reply/send_reply.rs (line 44)

40
41
42
43
44
45

pub async fn say_reply<U, E>(
    ctx: crate::Context<'_, U, E>,
    text: impl Into<String>,
) -> Result<crate::ReplyHandle<'_>, serenity::Error> {
    send_reply(ctx, |m| m.content(text.into())).await
}

Hide additional examples

src/structs/context.rs (line 107)

101
102
103
104
105
106
107
108

    pub async fn send<'att>(
        self,
        builder: impl for<'b> FnOnce(
            &'b mut crate::CreateReply<'att>,
        ) -> &'b mut crate::CreateReply<'att>,
    ) -> Result<crate::ReplyHandle<'a>, serenity::Error> {
        crate::send_reply(self, builder).await
    }