poise::reply

Struct CreateReply

pub struct CreateReply { /* private fields */ }

Message builder that abstracts over prefix and application command responses

Implementations

impl CreateReply

pub fn new() -> Self

Creates a blank CreateReply. Equivalent to Self::default.

pub fn content(self, content: impl Into<String>) -> Self

Set the content of the message.

Examples found in repository?

examples/feature_showcase/modal.rs (line 34)

25pub async fn component_modal(ctx: crate::Context<'_>) -> Result<(), Error> {
26    let reply = {
27        let components = vec![serenity::CreateActionRow::Buttons(vec![
28            serenity::CreateButton::new("open_modal")
29                .label("Open modal")
30                .style(poise::serenity_prelude::ButtonStyle::Success),
31        ])];
32
33        poise::CreateReply::default()
34            .content("Click the button below to open the modal")
35            .components(components)
36    };
37
38    ctx.send(reply).await?;
39
40    let serenity_ctx = ctx.serenity_context();
41    while let Some(mci) = serenity::ComponentInteractionCollector::new(serenity_ctx)
42        .timeout(std::time::Duration::from_secs(120))
43        .filter(move |mci| mci.data.custom_id == "open_modal")
44        .await
45    {
46        let data =
47            poise::execute_modal_on_component_interaction::<MyModal>(serenity_ctx, mci, None, None)
48                .await?;
49        println!("Got data: {:?}", data);
50    }
51    Ok(())
52}

Hide additional examples

examples/feature_showcase/collector.rs (line 17)

6pub async fn boop(ctx: Context<'_>) -> Result<(), Error> {
7    let uuid_boop = ctx.id();
8
9    let reply = {
10        let components = vec![serenity::CreateActionRow::Buttons(vec![
11            serenity::CreateButton::new(format!("{uuid_boop}"))
12                .style(serenity::ButtonStyle::Primary)
13                .label("Boop me!"),
14        ])];
15
16        CreateReply::default()
17            .content("I want some boops!")
18            .components(components)
19    };
20
21    ctx.send(reply).await?;
22
23    let mut boop_count = 0;
24    while let Some(mci) = serenity::ComponentInteractionCollector::new(ctx)
25        .author_id(ctx.author().id)
26        .channel_id(ctx.channel_id())
27        .timeout(std::time::Duration::from_secs(120))
28        .filter(move |mci| mci.data.custom_id == uuid_boop.to_string())
29        .await
30    {
31        boop_count += 1;
32
33        let mut msg = mci.message.clone();
34        msg.edit(
35            ctx,
36            serenity::EditMessage::new().content(format!("Boop count: {boop_count}")),
37        )
38        .await?;
39
40        mci.create_response(ctx, serenity::CreateInteractionResponse::Acknowledge)
41            .await?;
42    }
43
44    Ok(())
45}

examples/feature_showcase/track_edits.rs (line 20)

5pub async fn test_reuse_response(ctx: Context<'_>) -> Result<(), Error> {
6    let image_url = "https://raw.githubusercontent.com/serenity-rs/serenity/current/logo.png";
7
8    let reply = {
9        let embed = serenity::CreateEmbed::default()
10            .description("embed 1")
11            .image(image_url);
12
13        let components = vec![serenity::CreateActionRow::Buttons(vec![
14            serenity::CreateButton::new("1")
15                .label("button 1")
16                .style(serenity::ButtonStyle::Primary),
17        ])];
18
19        poise::CreateReply::default()
20            .content("message 1")
21            .embed(embed)
22            .components(components)
23    };
24
25    ctx.send(reply).await?;
26
27    tokio::time::sleep(std::time::Duration::from_secs(2)).await;
28
29    let image_url = "https://raw.githubusercontent.com/serenity-rs/serenity/current/examples/e09_create_message_builder/ferris_eyes.png";
30    let reply = {
31        let embed = serenity::CreateEmbed::default()
32            .description("embed 2")
33            .image(image_url);
34
35        let components = vec![serenity::CreateActionRow::Buttons(vec![
36            serenity::CreateButton::new("2")
37                .label("button 2")
38                .style(serenity::ButtonStyle::Danger),
39        ])];
40
41        poise::CreateReply::default()
42            .content("message 2")
43            .embed(embed)
44            .components(components)
45    };
46
47    ctx.send(reply).await?;
48    Ok(())
49}

pub fn embed(self, embed: CreateEmbed) -> Self

Adds an embed to the message.

Existing embeds are kept.

Examples found in repository?

examples/feature_showcase/track_edits.rs (line 21)

5pub async fn test_reuse_response(ctx: Context<'_>) -> Result<(), Error> {
6    let image_url = "https://raw.githubusercontent.com/serenity-rs/serenity/current/logo.png";
7
8    let reply = {
9        let embed = serenity::CreateEmbed::default()
10            .description("embed 1")
11            .image(image_url);
12
13        let components = vec![serenity::CreateActionRow::Buttons(vec![
14            serenity::CreateButton::new("1")
15                .label("button 1")
16                .style(serenity::ButtonStyle::Primary),
17        ])];
18
19        poise::CreateReply::default()
20            .content("message 1")
21            .embed(embed)
22            .components(components)
23    };
24
25    ctx.send(reply).await?;
26
27    tokio::time::sleep(std::time::Duration::from_secs(2)).await;
28
29    let image_url = "https://raw.githubusercontent.com/serenity-rs/serenity/current/examples/e09_create_message_builder/ferris_eyes.png";
30    let reply = {
31        let embed = serenity::CreateEmbed::default()
32            .description("embed 2")
33            .image(image_url);
34
35        let components = vec![serenity::CreateActionRow::Buttons(vec![
36            serenity::CreateButton::new("2")
37                .label("button 2")
38                .style(serenity::ButtonStyle::Danger),
39        ])];
40
41        poise::CreateReply::default()
42            .content("message 2")
43            .embed(embed)
44            .components(components)
45    };
46
47    ctx.send(reply).await?;
48    Ok(())
49}

pub fn components(self, components: Vec<CreateActionRow>) -> Self

Set components (buttons and select menus) for this message.

Any previously set components will be overwritten.

Examples found in repository?

examples/feature_showcase/modal.rs (line 35)

25pub async fn component_modal(ctx: crate::Context<'_>) -> Result<(), Error> {
26    let reply = {
27        let components = vec![serenity::CreateActionRow::Buttons(vec![
28            serenity::CreateButton::new("open_modal")
29                .label("Open modal")
30                .style(poise::serenity_prelude::ButtonStyle::Success),
31        ])];
32
33        poise::CreateReply::default()
34            .content("Click the button below to open the modal")
35            .components(components)
36    };
37
38    ctx.send(reply).await?;
39
40    let serenity_ctx = ctx.serenity_context();
41    while let Some(mci) = serenity::ComponentInteractionCollector::new(serenity_ctx)
42        .timeout(std::time::Duration::from_secs(120))
43        .filter(move |mci| mci.data.custom_id == "open_modal")
44        .await
45    {
46        let data =
47            poise::execute_modal_on_component_interaction::<MyModal>(serenity_ctx, mci, None, None)
48                .await?;
49        println!("Got data: {:?}", data);
50    }
51    Ok(())
52}

Hide additional examples

examples/feature_showcase/collector.rs (line 18)

6pub async fn boop(ctx: Context<'_>) -> Result<(), Error> {
7    let uuid_boop = ctx.id();
8
9    let reply = {
10        let components = vec![serenity::CreateActionRow::Buttons(vec![
11            serenity::CreateButton::new(format!("{uuid_boop}"))
12                .style(serenity::ButtonStyle::Primary)
13                .label("Boop me!"),
14        ])];
15
16        CreateReply::default()
17            .content("I want some boops!")
18            .components(components)
19    };
20
21    ctx.send(reply).await?;
22
23    let mut boop_count = 0;
24    while let Some(mci) = serenity::ComponentInteractionCollector::new(ctx)
25        .author_id(ctx.author().id)
26        .channel_id(ctx.channel_id())
27        .timeout(std::time::Duration::from_secs(120))
28        .filter(move |mci| mci.data.custom_id == uuid_boop.to_string())
29        .await
30    {
31        boop_count += 1;
32
33        let mut msg = mci.message.clone();
34        msg.edit(
35            ctx,
36            serenity::EditMessage::new().content(format!("Boop count: {boop_count}")),
37        )
38        .await?;
39
40        mci.create_response(ctx, serenity::CreateInteractionResponse::Acknowledge)
41            .await?;
42    }
43
44    Ok(())
45}

examples/feature_showcase/track_edits.rs (line 22)

5pub async fn test_reuse_response(ctx: Context<'_>) -> Result<(), Error> {
6    let image_url = "https://raw.githubusercontent.com/serenity-rs/serenity/current/logo.png";
7
8    let reply = {
9        let embed = serenity::CreateEmbed::default()
10            .description("embed 1")
11            .image(image_url);
12
13        let components = vec![serenity::CreateActionRow::Buttons(vec![
14            serenity::CreateButton::new("1")
15                .label("button 1")
16                .style(serenity::ButtonStyle::Primary),
17        ])];
18
19        poise::CreateReply::default()
20            .content("message 1")
21            .embed(embed)
22            .components(components)
23    };
24
25    ctx.send(reply).await?;
26
27    tokio::time::sleep(std::time::Duration::from_secs(2)).await;
28
29    let image_url = "https://raw.githubusercontent.com/serenity-rs/serenity/current/examples/e09_create_message_builder/ferris_eyes.png";
30    let reply = {
31        let embed = serenity::CreateEmbed::default()
32            .description("embed 2")
33            .image(image_url);
34
35        let components = vec![serenity::CreateActionRow::Buttons(vec![
36            serenity::CreateButton::new("2")
37                .label("button 2")
38                .style(serenity::ButtonStyle::Danger),
39        ])];
40
41        poise::CreateReply::default()
42            .content("message 2")
43            .embed(embed)
44            .components(components)
45    };
46
47    ctx.send(reply).await?;
48    Ok(())
49}

pub fn attachment(self, attachment: CreateAttachment) -> Self

Add an attachment.

pub fn ephemeral(self, ephemeral: bool) -> Self

Toggles whether the message is an ephemeral response (only invoking user can see it).

This only has an effect in slash commands!

pub fn allowed_mentions(self, allowed_mentions: CreateAllowedMentions) -> Self

Set the allowed mentions for the message.

See serenity::CreateAllowedMentions for more information.

pub fn poll(self, poll: CreatePoll<Ready>) -> Self

Adds a poll to the message. Only one poll can be added per message.

See serenity::CreatePoll for more information on creating and configuring a poll.

pub fn reply(self, reply: bool) -> Self

Makes this message an inline reply to another message like serenity::Message::reply (prefix-only, because slash commands are always inline replies anyways).

To disable the ping, set Self::allowed_mentions with serenity::CreateAllowedMentions::replied_user set to false.

impl CreateReply

Methods to create a message builder from any type from this CreateReply. Used by poise internally to actually send a response to Discord

pub fn to_slash_initial_response( self, builder: CreateInteractionResponseMessage, ) -> CreateInteractionResponseMessage

Serialize this response builder to a serenity::CreateInteractionResponseMessage

pub fn to_slash_followup_response( self, builder: CreateInteractionResponseFollowup, ) -> CreateInteractionResponseFollowup

Serialize this response builder to a serenity::CreateInteractionResponseFollowup

pub fn to_slash_initial_response_edit( self, builder: EditInteractionResponse, ) -> EditInteractionResponse

Serialize this response builder to a serenity::EditInteractionResponse

pub fn to_prefix_edit(self, builder: EditMessage) -> EditMessage

Serialize this response builder to a serenity::EditMessage

pub fn to_prefix(self, invocation_message: MessageReference) -> CreateMessage

Serialize this response builder to a serenity::CreateMessage