Struct poise::reply::ReplyHandle

pub struct ReplyHandle<'a>(_);

Returned from send_reply() to operate on the sent message

Discord sometimes returns the serenity::Message object directly, but sometimes you have to request it manually. This enum abstracts over the two cases

Implementations

impl ReplyHandle<'_>

pub async fn into_message(self) -> Result<Message, Error>

Retrieve the message object of the sent reply.

Note: to delete or edit, use ReplyHandle::delete() and ReplyHandle::edit() directly! Doing it via the methods from a Message object will fail for ephemeral messages

If you don’t need ownership of Message, you can use ReplyHandle::message

Only needs to do an HTTP request in the application command response case

pub async fn message(&self) -> Result<Cow<'_, Message>, Error>

Retrieve the message object of the sent reply.

Note: to delete or edit, use ReplyHandle::delete() and ReplyHandle::edit() directly! Doing it via the methods from a Message object will fail for ephemeral messages

Returns a reference to the known Message object, or fetches the message from the discord API.

To get an owned serenity::Message, use Self::into_message()

Examples found in repository?

src/builtins/register.rs (line 221)

172
173
174
175
176
177
178
179
180
181
182
183
184
185
186
187
188
189
190
191
192
193
194
195
196
197
198
199
200
201
202
203
204
205
206
207
208
209
210
211
212
213
214
215
216
217
218
219
220
221
222
223
224
225
226
227
228
229
230
231
232
233
234
235
236
237
238
239
240
241
242
243
244
245
246
247
248
249
250
251
252
253
254
255
256
257
258
259
260
261
262
263
264
265
266
267
268
269
270
271
272
273
274
275
276
277
278
279
280
281
282
283
284
285
286
287
288
289
290
291
292
293
294
295
296
297
298
299
300
301
302
303
304
305

pub async fn register_application_commands_buttons<U, E>(
    ctx: crate::Context<'_, U, E>,
) -> Result<(), serenity::Error> {
    let create_commands = create_application_commands(&ctx.framework().options().commands);
    let num_commands = create_commands.0.len();

    let is_bot_owner = ctx.framework().options().owners.contains(&ctx.author().id);
    if !is_bot_owner {
        ctx.say("Can only be used by bot owner").await?;
        return Ok(());
    }

    let reply = ctx
        .send(|m| {
            m.content("Choose what to do with the commands:")
                .components(|c| {
                    c.create_action_row(|r| {
                        r.create_button(|b| {
                            b.custom_id("register.guild")
                                .label("Register in guild")
                                .style(serenity::ButtonStyle::Primary)
                                .emoji('📋')
                        })
                        .create_button(|b| {
                            b.custom_id("unregister.guild")
                                .label("Delete in guild")
                                .style(serenity::ButtonStyle::Danger)
                                .emoji('🗑')
                        })
                    })
                    .create_action_row(|r| {
                        r.create_button(|b| {
                            b.custom_id("register.global")
                                .label("Register globally")
                                .style(serenity::ButtonStyle::Primary)
                                .emoji('📋')
                        })
                        .create_button(|b| {
                            b.custom_id("unregister.global")
                                .label("Delete globally")
                                .style(serenity::ButtonStyle::Danger)
                                .emoji('🗑')
                        })
                    })
                })
        })
        .await?;

    let interaction = reply
        .message()
        .await?
        .await_component_interaction(ctx)
        .author_id(ctx.author().id)
        .await;

    reply
        .edit(ctx, |b| {
            b.components(|b| b).content("Processing... Please wait.")
        })
        .await?; // remove buttons after button press and edit message
    let pressed_button_id = match &interaction {
        Some(m) => &m.data.custom_id,
        None => {
            ctx.say(":warning: You didn't interact in time - please run the command again.")
                .await?;
            return Ok(());
        }
    };

    let (register, global) = match &**pressed_button_id {
        "register.global" => (true, true),
        "unregister.global" => (false, true),
        "register.guild" => (true, false),
        "unregister.guild" => (false, false),
        other => {
            log::warn!("unknown register button ID: {:?}", other);
            return Ok(());
        }
    };

    let start_time = std::time::Instant::now();

    if global {
        if register {
            ctx.say(format!(
                ":gear: Registering {} global commands...",
                num_commands
            ))
            .await?;
            serenity::Command::set_global_application_commands(ctx, |b| {
                *b = create_commands;
                b
            })
            .await?;
        } else {
            ctx.say(":gear: Unregistering global commands...").await?;
            serenity::Command::set_global_application_commands(ctx, |b| b).await?;
        }
    } else {
        let guild_id = match ctx.guild_id() {
            Some(x) => x,
            None => {
                ctx.say(":x: Must be called in guild").await?;
                return Ok(());
            }
        };
        if register {
            ctx.say(format!(
                ":gear: Registering {} guild commands...",
                num_commands
            ))
            .await?;
            guild_id
                .set_application_commands(ctx, |b| {
                    *b = create_commands;
                    b
                })
                .await?;
        } else {
            ctx.say(":gear: Unregistering guild commands...").await?;
            guild_id.set_application_commands(ctx, |b| b).await?;
        }
    }

    // Calulate time taken and send message
    let time_taken = start_time.elapsed();
    ctx.say(format!(
        ":white_check_mark: Done! Took {}ms",
        time_taken.as_millis()
    ))
    .await?;

    Ok(())
}

pub async fn edit<'att, U, E>(
&self,
ctx: Context<'_, U, E>,
builder: impl for<'a> FnOnce(&'a mut CreateReply<'att>) -> &'a mut CreateReply<'att>
) -> Result<(), Error>

Edits the message that this ReplyHandle points to

Examples found in repository?

src/builtins/register.rs (lines 228-230)

172
173
174
175
176
177
178
179
180
181
182
183
184
185
186
187
188
189
190
191
192
193
194
195
196
197
198
199
200
201
202
203
204
205
206
207
208
209
210
211
212
213
214
215
216
217
218
219
220
221
222
223
224
225
226
227
228
229
230
231
232
233
234
235
236
237
238
239
240
241
242
243
244
245
246
247
248
249
250
251
252
253
254
255
256
257
258
259
260
261
262
263
264
265
266
267
268
269
270
271
272
273
274
275
276
277
278
279
280
281
282
283
284
285
286
287
288
289
290
291
292
293
294
295
296
297
298
299
300
301
302
303
304
305

pub async fn register_application_commands_buttons<U, E>(
    ctx: crate::Context<'_, U, E>,
) -> Result<(), serenity::Error> {
    let create_commands = create_application_commands(&ctx.framework().options().commands);
    let num_commands = create_commands.0.len();

    let is_bot_owner = ctx.framework().options().owners.contains(&ctx.author().id);
    if !is_bot_owner {
        ctx.say("Can only be used by bot owner").await?;
        return Ok(());
    }

    let reply = ctx
        .send(|m| {
            m.content("Choose what to do with the commands:")
                .components(|c| {
                    c.create_action_row(|r| {
                        r.create_button(|b| {
                            b.custom_id("register.guild")
                                .label("Register in guild")
                                .style(serenity::ButtonStyle::Primary)
                                .emoji('📋')
                        })
                        .create_button(|b| {
                            b.custom_id("unregister.guild")
                                .label("Delete in guild")
                                .style(serenity::ButtonStyle::Danger)
                                .emoji('🗑')
                        })
                    })
                    .create_action_row(|r| {
                        r.create_button(|b| {
                            b.custom_id("register.global")
                                .label("Register globally")
                                .style(serenity::ButtonStyle::Primary)
                                .emoji('📋')
                        })
                        .create_button(|b| {
                            b.custom_id("unregister.global")
                                .label("Delete globally")
                                .style(serenity::ButtonStyle::Danger)
                                .emoji('🗑')
                        })
                    })
                })
        })
        .await?;

    let interaction = reply
        .message()
        .await?
        .await_component_interaction(ctx)
        .author_id(ctx.author().id)
        .await;

    reply
        .edit(ctx, |b| {
            b.components(|b| b).content("Processing... Please wait.")
        })
        .await?; // remove buttons after button press and edit message
    let pressed_button_id = match &interaction {
        Some(m) => &m.data.custom_id,
        None => {
            ctx.say(":warning: You didn't interact in time - please run the command again.")
                .await?;
            return Ok(());
        }
    };

    let (register, global) = match &**pressed_button_id {
        "register.global" => (true, true),
        "unregister.global" => (false, true),
        "register.guild" => (true, false),
        "unregister.guild" => (false, false),
        other => {
            log::warn!("unknown register button ID: {:?}", other);
            return Ok(());
        }
    };

    let start_time = std::time::Instant::now();

    if global {
        if register {
            ctx.say(format!(
                ":gear: Registering {} global commands...",
                num_commands
            ))
            .await?;
            serenity::Command::set_global_application_commands(ctx, |b| {
                *b = create_commands;
                b
            })
            .await?;
        } else {
            ctx.say(":gear: Unregistering global commands...").await?;
            serenity::Command::set_global_application_commands(ctx, |b| b).await?;
        }
    } else {
        let guild_id = match ctx.guild_id() {
            Some(x) => x,
            None => {
                ctx.say(":x: Must be called in guild").await?;
                return Ok(());
            }
        };
        if register {
            ctx.say(format!(
                ":gear: Registering {} guild commands...",
                num_commands
            ))
            .await?;
            guild_id
                .set_application_commands(ctx, |b| {
                    *b = create_commands;
                    b
                })
                .await?;
        } else {
            ctx.say(":gear: Unregistering guild commands...").await?;
            guild_id.set_application_commands(ctx, |b| b).await?;
        }
    }

    // Calulate time taken and send message
    let time_taken = start_time.elapsed();
    ctx.say(format!(
        ":white_check_mark: Done! Took {}ms",
        time_taken.as_millis()
    ))
    .await?;

    Ok(())
}

pub async fn delete<U, E>(&self, ctx: Context<'_, U, E>) -> Result<(), Error>

Deletes this message