Function poise::samples::help

pub async fn help<D, E>(
    ctx: Context<'_, D, E>, 
    command: Option<&str>, 
    extra_text_at_bottom: &str, 
    response_mode: HelpResponseMode
) -> Result<(), Error>

A help command that outputs text in a code block, groups commands by categories, and annotates commands with a slash if they exist as slash commands.

Example usage from Ferris, the Discord bot running in the Rust community server:

/// Show this menu
#[poise::command(prefix_command, track_edits, slash_command)]
pub async fn help(
    ctx: Context<'_>,
    #[description = "Specific command to show help about"] command: Option<String>,
) -> Result<(), Error> {
    let bottom_text = "Type ?help command for more info on a command.
You can edit your message to the bot and the bot will edit its response.";
    poise::samples::help(
        ctx,
        command.as_deref(),
        bottom_text,
        poise::samples::HelpResponseMode::Ephemeral,
    )
    .await?;
    Ok(())
}

Output:

Playground:
  ?play        Compile and run Rust code in a playground
  ?eval        Evaluate a single Rust expression
  ?miri        Run code and detect undefined behavior using Miri
  ?expand      Expand macros to their raw desugared form
  ?clippy      Catch common mistakes using the Clippy linter
  ?fmt         Format code using rustfmt
  ?microbench  Benchmark small snippets of code
  ?procmacro   Compile and use a procedural macro
  ?godbolt     View assembly using Godbolt
  ?mca         Run performance analysis using llvm-mca
  ?llvmir      View LLVM IR using Godbolt
Crates:
  /crate       Lookup crates on crates.io
  /doc         Lookup documentation
Moderation:
  /cleanup     Deletes the bot's messages for cleanup
  /ban         Bans another person
  ?move        Move a discussion to another channel
  /rustify     Adds the Rustacean role to members
Miscellaneous:
  ?go          Evaluates Go code
  /source      Links to the bot GitHub repo
  /help        Show this menu

Type ?help command for more info on a command.
You can edit your message to the bot and the bot will edit its response.