poise

Module serenity_prelude

This module re-exports a bunch of items from all over serenity. Useful if you can’t remember the full paths of serenity items.

One way to use this prelude module in your project is

use poise::serenity_prelude as serenity;

Re-exports

pub use all::validate as validate_token;

pub use all::InvalidToken;

pub use all::async_trait;

pub use all::Result;

pub use all::AttachmentType;

pub use all::Color;

pub use all::colours;

pub use all::audit_log;

pub use all::automod;

pub use all::Mentionable;

pub use all::EntitlementOwner;

pub use all::PRESET_GENERAL;

pub use all::PRESET_TEXT;

pub use all::PRESET_VOICE;

pub use all::CacheUpdate;

pub use all::ShardManagerOptions;

pub use all::ShardQueuer;

pub use all::ShardRunner;

pub use all::ShardRunnerOptions;

pub use all::CollectorCallback;

pub use all::Shard;

pub use all::WsClient;

pub use all::StatusCode;

pub use all::HttpBuilder;

pub use all::deserialize_errors;

pub use all::Ratelimiter;

pub use all::RequestBuilder;

pub use all::parse_message_id_pair;

pub use all::parse_message_url;

pub use all::content_safe;

pub use all::QuickModalResponse;

pub use all::CreateQuickModal;

pub use all::ContentModifier;

pub use all::EmbedMessageBuilding;

pub use all::futures;

pub use all::Client;

pub use all::create_poll;

pub use all::CacheRef;

pub use all::MemberRef;

pub use all::GuildRoleRef;

pub use all::UserRef;

pub use all::GuildRef;

pub use all::SettingsRef;

pub use all::GuildChannelRef;

pub use all::CurrentUserRef;

pub use all::GuildRolesRef;

pub use all::GuildChannelsRef;

pub use all::MessageRef;

pub use all::ChannelMessagesRef;

pub use all::ClientBuilder;

pub use all::collect;

pub use all::UNKNOWN_ERROR;

pub use all::UNKNOWN_OPCODE;

pub use all::DECODE_ERROR;

pub use all::NOT_AUTHENTICATED;

pub use all::AUTHENTICATION_FAILED;

pub use all::ALREADY_AUTHENTICATED;

pub use all::INVALID_SEQUENCE;

pub use all::RATE_LIMITED;

pub use all::SESSION_TIMEOUT;

pub use all::INVALID_SHARD;

pub use all::SHARDING_REQUIRED;

pub use all::INVALID_GATEWAY_INTENTS;

pub use all::DISALLOWED_GATEWAY_INTENTS;

pub use all::EMBED_MAX_LENGTH;

pub use all::EMBED_MAX_COUNT;

pub use all::STICKER_MAX_COUNT;

pub use all::GATEWAY_VERSION;

pub use all::LARGE_THRESHOLD;

pub use all::MESSAGE_CODE_LIMIT;

pub use all::MEMBER_FETCH_LIMIT;

pub use all::USER_AGENT;

pub use all::close_codes;

pub use all::CacheHttp;

pub use all::GuildPagination;

pub use all::UserPagination;

pub use all::token;

pub use all::parse_invite;

pub use all::parse_user_tag;

pub use all::parse_user_mention;

pub use all::parse_username;

pub use all::parse_role_mention;

pub use all::parse_role;

pub use all::parse_channel_mention;

pub use all::parse_channel;

pub use all::parse_emoji;

pub use all::parse_quotes;

pub use all::parse_webhook;

pub use all::shard_id;

pub use all::constants;

pub use all::json;

pub use all::model;

pub use all::prelude;

pub use all::builder;

pub use all::cache;

pub use all::client;

pub use all::collector;

pub use all::framework;

pub use all::gateway;

pub use all::http;

pub use all::utils;

pub use all::FutureExt;

pub use all::Opcode;

pub use all::Command;

pub use all::CommandType;

pub use all::EntryPointHandlerType;

pub use all::CommandOption;

pub use all::CommandOptionType;

pub use all::CommandOptionChoice;

pub use all::CommandPermissions;

pub use all::CommandPermission;

pub use all::CommandPermissionType;

pub use all::CommandInteraction;

pub use all::CommandData;

pub use all::AutocompleteOption;

pub use all::ResolvedOption;

pub use all::ResolvedValue;

pub use all::Unresolved;

pub use all::ResolvedTarget;

pub use all::CommandDataResolved;

pub use all::CommandDataOption;

pub use all::CommandDataOptionValue;

pub use all::ComponentType;

pub use all::ActionRow;

pub use all::ActionRowComponent;

pub use all::ButtonKind;

pub use all::Button;

pub use all::ButtonStyle;

pub use all::SelectMenu;

pub use all::SelectMenuOption;

pub use all::InputText;

pub use all::InputTextStyle;

pub use all::ComponentInteraction;

pub use all::ComponentInteractionDataKind;

pub use all::ComponentInteractionData;

pub use all::Interaction;

pub use all::InteractionType;

pub use all::InteractionResponseFlags;

pub use all::AuthorizingIntegrationOwner;

pub use all::AuthorizingIntegrationOwners;

pub use all::MessageInteraction;

pub use all::MessageCommandInteractionMetadata;

pub use all::MessageComponentInteractionMetadata;

pub use all::MessageModalSubmitInteractionMetadata;

pub use all::MessageInteractionMetadata;

pub use all::ModalInteraction;

pub use all::ModalInteractionData;

pub use all::Scope;

pub use all::PingInteraction;

pub use all::PartialCurrentApplicationInfo;

pub use all::CurrentApplicationInfo;

pub use all::InstallationContext;

pub use all::InteractionContext;

pub use all::InstallationContextConfig;

pub use all::Team;

pub use all::TeamMember;

pub use all::MembershipState;

pub use all::TeamMemberRole;

pub use all::ApplicationFlags;

pub use all::InstallParams;

pub use all::Attachment;

pub use all::MessagesIter;

pub use all::Embed;

pub use all::EmbedAuthor;

pub use all::EmbedField;

pub use all::EmbedFooter;

pub use all::EmbedImage;

pub use all::EmbedProvider;

pub use all::EmbedThumbnail;

pub use all::EmbedVideo;

pub use all::GuildChannel;

pub use all::ForumLayoutType;

pub use all::PartialGuildChannel;

pub use all::Message;

pub use all::MessageReaction;

pub use all::CountDetails;

pub use all::MessageType;

pub use all::MessageActivityKind;

pub use all::MessageApplication;

pub use all::MessageActivity;

pub use all::MessageReferenceKind;

pub use all::MessageReference;

pub use all::ChannelMention;

pub use all::MessageSnapshot;

pub use all::MessageFlags;

pub use all::Nonce;

pub use all::RoleSubscriptionData;

pub use all::Poll;

pub use all::PollMedia;

pub use all::PollMediaEmoji;

pub use all::PollAnswer;

pub use all::PollLayoutType;

pub use all::PollResults;

pub use all::PollAnswerCount;

pub use all::PartialChannel;

pub use all::FollowedChannel;

pub use all::PrivateChannel;

pub use all::Reaction;

pub use all::ReactionTypes;

pub use all::ReactionType;

pub use all::ReactionConversionError;

pub use all::Channel;

pub use all::ChannelType;

pub use all::PermissionOverwrite;

pub use all::PermissionOverwriteType;

pub use all::VideoQualityMode;

pub use all::StageInstancePrivacyLevel;

pub use all::AutoArchiveDuration;

pub use all::StageInstance;

pub use all::ThreadMetadata;

pub use all::ThreadsData;

pub use all::ForumEmoji;

pub use all::ForumTag;

pub use all::SortOrder;

pub use all::ChannelFlags;

pub use all::Colour;

pub use all::Connection;

pub use all::ConnectionVisibility;

pub use all::CommandPermissionsUpdateEvent;

pub use all::AutoModRuleCreateEvent;

pub use all::AutoModRuleUpdateEvent;

pub use all::AutoModRuleDeleteEvent;

pub use all::AutoModActionExecutionEvent;

pub use all::ChannelCreateEvent;

pub use all::ChannelDeleteEvent;

pub use all::ChannelPinsUpdateEvent;

pub use all::ChannelUpdateEvent;

pub use all::GuildAuditLogEntryCreateEvent;

pub use all::GuildBanAddEvent;

pub use all::GuildBanRemoveEvent;

pub use all::GuildCreateEvent;

pub use all::GuildDeleteEvent;

pub use all::GuildEmojisUpdateEvent;

pub use all::GuildIntegrationsUpdateEvent;

pub use all::GuildMemberAddEvent;

pub use all::GuildMemberRemoveEvent;

pub use all::GuildMemberUpdateEvent;

pub use all::GuildMembersChunkEvent;

pub use all::SoundboardSoundsEvent;

pub use all::SoundboardSoundCreateEvent;

pub use all::SoundboardSoundUpdateEvent;

pub use all::SoundboardSoundsUpdateEvent;

pub use all::SoundboardSoundDeleteEvent;

pub use all::GuildRoleCreateEvent;

pub use all::GuildRoleDeleteEvent;

pub use all::GuildRoleUpdateEvent;

pub use all::GuildStickersUpdateEvent;

pub use all::InviteCreateEvent;

pub use all::InviteDeleteEvent;

pub use all::GuildUpdateEvent;

pub use all::MessageCreateEvent;

pub use all::MessageDeleteBulkEvent;

pub use all::MessageDeleteEvent;

pub use all::MessageUpdateEvent;

pub use all::PresenceUpdateEvent;

pub use all::PresencesReplaceEvent;

pub use all::ReactionAddEvent;

pub use all::ReactionRemoveEvent;

pub use all::ReactionRemoveAllEvent;

pub use all::ReactionRemoveEmojiEvent;

pub use all::ReadyEvent;

pub use all::ResumedEvent;

pub use all::TypingStartEvent;

pub use all::UnknownEvent;

pub use all::UserUpdateEvent;

pub use all::VoiceServerUpdateEvent;

pub use all::VoiceStateUpdateEvent;

pub use all::VoiceChannelStatusUpdateEvent;

pub use all::WebhookUpdateEvent;

pub use all::InteractionCreateEvent;

pub use all::IntegrationCreateEvent;

pub use all::IntegrationUpdateEvent;

pub use all::IntegrationDeleteEvent;

pub use all::StageInstanceCreateEvent;

pub use all::StageInstanceUpdateEvent;

pub use all::StageInstanceDeleteEvent;

pub use all::ThreadCreateEvent;

pub use all::ThreadUpdateEvent;

pub use all::ThreadDeleteEvent;

pub use all::ThreadListSyncEvent;

pub use all::ThreadMemberUpdateEvent;

pub use all::ThreadMembersUpdateEvent;

pub use all::GuildScheduledEventCreateEvent;

pub use all::GuildScheduledEventUpdateEvent;

pub use all::GuildScheduledEventDeleteEvent;

pub use all::GuildScheduledEventUserAddEvent;

pub use all::GuildScheduledEventUserRemoveEvent;

pub use all::EntitlementCreateEvent;

pub use all::EntitlementUpdateEvent;

pub use all::EntitlementDeleteEvent;

pub use all::MessagePollVoteAddEvent;

pub use all::MessagePollVoteRemoveEvent;

pub use all::GatewayEvent;

pub use all::Event;

pub use all::BotGateway;

pub use all::Activity;

pub use all::ActivityButton;

pub use all::ActivityAssets;

pub use all::ActivityFlags;

pub use all::ActivityParty;

pub use all::ActivitySecrets;

pub use all::ActivityEmoji;

pub use all::ActivityType;

pub use all::Gateway;

pub use all::ClientStatus;

pub use all::PresenceUser;

pub use all::Presence;

pub use all::Ready;

pub use all::SessionStartLimit;

pub use all::ShardInfo;

pub use all::ActivityTimestamps;

pub use all::GatewayIntents;

pub use all::ChannelAction;

pub use all::ChannelOverwriteAction;

pub use all::MemberAction;

pub use all::RoleAction;

pub use all::InviteAction;

pub use all::WebhookAction;

pub use all::EmojiAction;

pub use all::MessageAction;

pub use all::IntegrationAction;

pub use all::StageInstanceAction;

pub use all::StickerAction;

pub use all::ScheduledEventAction;

pub use all::ThreadAction;

pub use all::AutoModAction;

pub use all::CreatorMonetizationAction;

pub use all::VoiceChannelStatusAction;

pub use all::AuditLogs;

pub use all::PartialIntegration;

pub use all::AuditLogEntry;

pub use all::Options;

pub use all::Emoji;

pub use all::MembersIter;

pub use all::GuildWidgetStyle;

pub use all::GuildPreview;

pub use all::Integration;

pub use all::IntegrationExpireBehaviour;

pub use all::IntegrationAccount;

pub use all::IntegrationApplication;

pub use all::Member;

pub use all::GuildMemberFlags;

pub use all::PartialMember;

pub use all::PartialThreadMember;

pub use all::ThreadMember;

pub use all::ThreadMemberFlags;

pub use all::PartialGuild;

pub use all::PremiumTier;

pub use all::Role;

pub use all::RoleColours;

pub use all::RoleTags;

pub use all::ScheduledEvent;

pub use all::ScheduledEventStatus;

pub use all::ScheduledEventType;

pub use all::ScheduledEventMetadata;

pub use all::ScheduledEventUser;

pub use all::ScheduledEventPrivacyLevel;

pub use all::SystemChannelFlags;

pub use all::GuildWelcomeScreen;

pub use all::GuildWelcomeChannel;

pub use all::GuildWelcomeChannelEmoji;

pub use all::Ban;

pub use all::BulkBanResponse;

pub use all::AfkMetadata;

pub use all::Guild;

pub use all::GuildWidget;

pub use all::GuildPrune;

pub use all::GuildInfo;

pub use all::UnavailableGuild;

pub use all::DefaultMessageNotificationLevel;

pub use all::ExplicitContentFilter;

pub use all::MfaLevel;

pub use all::VerificationLevel;

pub use all::NsfwLevel;

pub use all::AfkTimeout;

pub use all::IncidentsData;

pub use all::AttachmentId;

pub use all::ApplicationId;

pub use all::ChannelId;

pub use all::EmojiId;

pub use all::GenericId;

pub use all::GuildId;

pub use all::IntegrationId;

pub use all::MessageId;

pub use all::RoleId;

pub use all::ScheduledEventId;

pub use all::SoundId;

pub use all::StickerId;

pub use all::StickerPackId;

pub use all::StickerPackBannerId;

pub use all::SkuId;

pub use all::UserId;

pub use all::WebhookId;

pub use all::AuditLogEntryId;

pub use all::InteractionId;

pub use all::CommandId;

pub use all::CommandPermissionId;

pub use all::CommandVersionId;

pub use all::TargetId;

pub use all::StageInstanceId;

pub use all::RuleId;

pub use all::ForumTagId;

pub use all::EntitlementId;

pub use all::ShardId;

pub use all::AnswerId;

pub use all::Invite;

pub use all::InviteChannel;

pub use all::InviteGuild;

pub use all::RichInvite;

pub use all::InviteStageInstance;

pub use all::InviteTargetType;

pub use all::Mention;

pub use all::MentionParseError;

pub use all::ImageHash;

pub use all::ImageHashParseError;

pub use all::EmojiIdentifier;

pub use all::EmojiIdentifierParseError;

pub use all::Incident;

pub use all::IncidentUpdate;

pub use all::Maintenance;

pub use all::Sku;

pub use all::SkuKind;

pub use all::SkuFlags;

pub use all::Entitlement;

pub use all::EntitlementKind;

pub use all::Permissions;

pub use all::Soundboard;

pub use all::StickerItem;

pub use all::StickerPack;

pub use all::Sticker;

pub use all::StickerType;

pub use all::StickerFormatType;

pub use all::CurrentUser;

pub use all::OnlineStatus;

pub use all::User;

pub use all::PremiumType;

pub use all::UserPublicFlags;

pub use all::PrimaryGuild;

pub use all::AvatarDecorationData;

pub use all::Collectibles;

pub use all::Nameplate;

pub use all::VoiceRegion;

pub use all::VoiceState;

pub use all::WebhookType;

pub use all::Webhook;

pub use all::WebhookGuild;

pub use all::WebhookChannel;

pub use all::Builder;

pub use all::AddMember;

pub use all::CreateBotAuthParameters;

pub use all::CreateAllowedMentions;

pub use all::CreateAttachment;

pub use all::EditAttachments;

pub use all::CreateChannel;

pub use all::CreateCommandOption;

pub use all::CreateCommand;

pub use all::EditCommandPermissions;

pub use all::CreateCommandPermission;

pub use all::CreateActionRow;

pub use all::CreateButton;

pub use all::CreateSelectMenuKind;

pub use all::CreateSelectMenu;

pub use all::CreateSelectMenuOption;

pub use all::CreateInputText;

pub use all::CreateEmbed;

pub use all::CreateEmbedAuthor;

pub use all::CreateEmbedFooter;

pub use all::CreateForumPost;

pub use all::CreateForumTag;

pub use all::CreateInteractionResponse;

pub use all::CreateInteractionResponseMessage;

pub use all::AutocompleteChoice;

pub use all::CreateAutocompleteResponse;

pub use all::CreateModal;

pub use all::CreateInteractionResponseFollowup;

pub use all::CreateInvite;

pub use all::CreateMessage;

pub use all::CreateScheduledEvent;

pub use all::CreateSoundboard;

pub use all::CreateStageInstance;

pub use all::CreateSticker;

pub use all::CreateThread;

pub use all::CreateWebhook;

pub use all::EditAutoModRule;

pub use all::EditChannel;

pub use all::EditGuild;

pub use all::EditGuildIncidentActions;

pub use all::EditGuildWelcomeScreen;

pub use all::CreateGuildWelcomeChannel;

pub use all::EditGuildWidget;

pub use all::EditInteractionResponse;

pub use all::EditMember;

pub use all::EditMessage;

pub use all::EditProfile;

pub use all::EditRole;

pub use all::CreateRoleColours;

pub use all::EditScheduledEvent;

pub use all::EditSoundboard;

pub use all::EditStageInstance;

pub use all::EditSticker;

pub use all::EditThread;

pub use all::EditVoiceState;

pub use all::EditWebhook;

pub use all::EditWebhookMessage;

pub use all::ExecuteWebhook;

pub use all::GetEntitlements;

pub use all::GetMessages;

pub use all::Cache;

pub use all::ComponentInteractionCollector;

pub use all::ModalInteractionCollector;

pub use all::ReactionCollector;

pub use all::MessageCollector;

pub use all::EventCollector;

pub use all::Framework;

pub use all::ShardQueuerMessage;

pub use all::ShardRunnerInfo;

pub use all::PresenceData;

pub use all::ActivityData;

pub use all::ConnectionStage;

pub use all::ShardAction;

pub use all::ReconnectType;

pub use all::ChunkGuildFilter;

pub use all::Http;

pub use all::DiscordJsonError;

pub use all::DiscordJsonSingleError;

pub use all::ErrorResponse;

pub use all::HttpError;

pub use all::MultipartUpload;

pub use all::Multipart;

pub use all::RatelimitInfo;

pub use all::Ratelimit;

pub use all::Request;

pub use all::RatelimitingBucket;

pub use all::Route;

pub use all::Typing;

pub use all::LightMethod;

pub use all::MessagePagination;

pub use all::MemberParseError;

pub use all::MessageParseError;

pub use all::UserParseError;

pub use all::ChannelParseError;

pub use all::GuildChannelParseError;

pub use all::GuildParseError;

pub use all::RoleParseError;

pub use all::EmojiParseError;

pub use all::ArgumentConvert;

pub use all::ContentSafeOptions;

pub use all::FormattedTimestamp;

pub use all::FormattedTimestampStyle;

pub use all::FormattedTimestampParseError;

pub use all::Error;

pub use all::AffectedRole;

pub use all::Change;

pub use all::EntityType;

pub use all::EventType as AutomodEventType;

pub use all::Action;

pub use all::ActionExecution;

pub use all::ActionType;

pub use all::KeywordPresetType;

pub use all::Rule;

pub use all::Trigger;

pub use all::TriggerMetadata;

pub use all::TriggerType;

pub use all::CreatePoll;

pub use all::CreatePollAnswer;

pub use all::Settings;

pub use all::Context;

pub use all::Error as ClientError;

pub use all::EventHandler;

pub use all::FullEvent;

pub use all::RawEventHandler;

pub use all::ShardStageUpdateEvent;

pub use all::ShardManager;

pub use all::ShardMessenger;

pub use all::ShardRunnerMessage;

pub use all::Error as GatewayError;

pub use all::CustomMessage;

pub use all::Content;

pub use all::MessageBuilder;

pub use all::Error as ModelError;

pub use all::Timestamp;