Struct poise::serenity_prelude::builder::EditAutoModRule

pub struct EditAutoModRule<'a> { /* private fields */ }

Available on crate feature builder only.

A builder for creating or editing guild AutoMod rules.

Examples

See GuildId::create_automod_rule for details.

Discord docs

Implementations

impl<'a> EditAutoModRule<'a>

pub fn new() -> EditAutoModRule<'a>

Equivalent to Self::default.

pub fn name(self, name: impl Into<String>) -> EditAutoModRule<'a>

The display name of the rule.

pub fn event_type(self, event_type: EventType) -> EditAutoModRule<'a>

Set the event context the rule should be checked.

pub fn trigger(self, trigger: Trigger) -> EditAutoModRule<'a>

Set the type of content which can trigger the rule.

None: The trigger type can’t be edited after creation. Only its values.

pub fn actions(self, actions: Vec<Action>) -> EditAutoModRule<'a>

Set the actions which will execute when the rule is triggered.

pub fn enabled(self, enabled: bool) -> EditAutoModRule<'a>

Set whether the rule is enabled.

pub fn exempt_roles( self, roles: impl IntoIterator<Item = impl Into<RoleId>> ) -> EditAutoModRule<'a>

Set roles that should not be affected by the rule.

Maximum of 20.

pub fn exempt_channels( self, channels: impl IntoIterator<Item = impl Into<ChannelId>> ) -> EditAutoModRule<'a>

Set channels that should not be affected by the rule.

Maximum of 50.

pub fn audit_log_reason(self, reason: &'a str) -> EditAutoModRule<'a>

Sets the request’s audit log reason.