Enum poise::serenity_prelude::Event

#[non_exhaustive]
pub enum Event {

    CommandPermissionsUpdate(CommandPermissionsUpdateEvent),
    AutoModRuleCreate(AutoModRuleCreateEvent),
    AutoModRuleUpdate(AutoModRuleUpdateEvent),
    AutoModRuleDelete(AutoModRuleDeleteEvent),
    AutoModActionExecution(AutoModActionExecutionEvent),
    ChannelCreate(ChannelCreateEvent),
    ChannelDelete(ChannelDeleteEvent),
    ChannelPinsUpdate(ChannelPinsUpdateEvent),
    ChannelUpdate(ChannelUpdateEvent),
    GuildAuditLogEntryCreate(GuildAuditLogEntryCreateEvent),
    GuildBanAdd(GuildBanAddEvent),
    GuildBanRemove(GuildBanRemoveEvent),
    GuildCreate(GuildCreateEvent),
    GuildDelete(GuildDeleteEvent),
    GuildEmojisUpdate(GuildEmojisUpdateEvent),
    GuildIntegrationsUpdate(GuildIntegrationsUpdateEvent),
    GuildMemberAdd(GuildMemberAddEvent),
    GuildMemberRemove(GuildMemberRemoveEvent),
    GuildMemberUpdate(GuildMemberUpdateEvent),
    GuildMembersChunk(GuildMembersChunkEvent),
    GuildRoleCreate(GuildRoleCreateEvent),
    GuildRoleDelete(GuildRoleDeleteEvent),
    GuildRoleUpdate(GuildRoleUpdateEvent),
    GuildStickersUpdate(GuildStickersUpdateEvent),
    GuildUpdate(GuildUpdateEvent),
    InviteCreate(InviteCreateEvent),
    InviteDelete(InviteDeleteEvent),
    MessageCreate(MessageCreateEvent),
    MessageDelete(MessageDeleteEvent),
    MessageDeleteBulk(MessageDeleteBulkEvent),
    MessageUpdate(MessageUpdateEvent),
    PresenceUpdate(PresenceUpdateEvent),
    PresencesReplace(PresencesReplaceEvent),
    ReactionAdd(ReactionAddEvent),
    ReactionRemove(ReactionRemoveEvent),
    ReactionRemoveAll(ReactionRemoveAllEvent),
    ReactionRemoveEmoji(ReactionRemoveEmojiEvent),
    Ready(ReadyEvent),
    Resumed(ResumedEvent),
    TypingStart(TypingStartEvent),
    UserUpdate(UserUpdateEvent),
    VoiceStateUpdate(VoiceStateUpdateEvent),
    VoiceServerUpdate(VoiceServerUpdateEvent),
    VoiceChannelStatusUpdate(VoiceChannelStatusUpdateEvent),
    WebhookUpdate(WebhookUpdateEvent),
    InteractionCreate(InteractionCreateEvent),
    IntegrationCreate(IntegrationCreateEvent),
    IntegrationUpdate(IntegrationUpdateEvent),
    IntegrationDelete(IntegrationDeleteEvent),
    StageInstanceCreate(StageInstanceCreateEvent),
    StageInstanceUpdate(StageInstanceUpdateEvent),
    StageInstanceDelete(StageInstanceDeleteEvent),
    ThreadCreate(ThreadCreateEvent),
    ThreadUpdate(ThreadUpdateEvent),
    ThreadDelete(ThreadDeleteEvent),
    ThreadListSync(ThreadListSyncEvent),
    ThreadMemberUpdate(ThreadMemberUpdateEvent),
    ThreadMembersUpdate(ThreadMembersUpdateEvent),
    GuildScheduledEventCreate(GuildScheduledEventCreateEvent),
    GuildScheduledEventUpdate(GuildScheduledEventUpdateEvent),
    GuildScheduledEventDelete(GuildScheduledEventDeleteEvent),
    GuildScheduledEventUserAdd(GuildScheduledEventUserAddEvent),
    GuildScheduledEventUserRemove(GuildScheduledEventUserRemoveEvent),
    EntitlementCreate(EntitlementCreateEvent),
    EntitlementUpdate(EntitlementUpdateEvent),
    EntitlementDelete(EntitlementDeleteEvent),
    Unknown(UnknownEvent),

}

Event received over a websocket connection

Discord docs.

Variants (Non-exhaustive)

Non-exhaustive enums could have additional variants added in future. Therefore, when matching against variants of non-exhaustive enums, an extra wildcard arm must be added to account for any future variants.

CommandPermissionsUpdate(CommandPermissionsUpdateEvent)

The permissions of an Command was changed.

Fires the EventHandler::command_permissions_update event.

AutoModRuleCreate(AutoModRuleCreateEvent)

A Rule was created.

Fires the EventHandler::auto_moderation_rule_create event.

AutoModRuleUpdate(AutoModRuleUpdateEvent)

A Rule has been updated.

Fires the EventHandler::auto_moderation_rule_update event.

AutoModRuleDelete(AutoModRuleDeleteEvent)

A Rule was deleted.

Fires the EventHandler::auto_moderation_rule_delete event.

AutoModActionExecution(AutoModActionExecutionEvent)

A Rule was triggered and an action was executed.

Fires the EventHandler::auto_moderation_action_execution event.

ChannelCreate(ChannelCreateEvent)

A Channel was created.

Fires the EventHandler::channel_create event.

ChannelDelete(ChannelDeleteEvent)

A Channel has been deleted.

Fires the EventHandler::channel_delete event.

ChannelPinsUpdate(ChannelPinsUpdateEvent)

The pins for a Channel have been updated.

Fires the EventHandler::channel_pins_update event.

ChannelUpdate(ChannelUpdateEvent)

A Channel has been updated.

Fires the EventHandler::channel_update event.

GuildAuditLogEntryCreate(GuildAuditLogEntryCreateEvent)

GuildBanAdd(GuildBanAddEvent)

GuildBanRemove(GuildBanRemoveEvent)

GuildCreate(GuildCreateEvent)

GuildDelete(GuildDeleteEvent)

GuildEmojisUpdate(GuildEmojisUpdateEvent)

GuildIntegrationsUpdate(GuildIntegrationsUpdateEvent)

GuildMemberAdd(GuildMemberAddEvent)

GuildMemberRemove(GuildMemberRemoveEvent)

GuildMemberUpdate(GuildMemberUpdateEvent)

A member’s roles have changed

GuildMembersChunk(GuildMembersChunkEvent)

GuildRoleCreate(GuildRoleCreateEvent)

GuildRoleDelete(GuildRoleDeleteEvent)

GuildRoleUpdate(GuildRoleUpdateEvent)

GuildStickersUpdate(GuildStickersUpdateEvent)

A Sticker was created, updated, or deleted

GuildUpdate(GuildUpdateEvent)

InviteCreate(InviteCreateEvent)

An Invite was created.

Fires the EventHandler::invite_create event handler.

InviteDelete(InviteDeleteEvent)

An Invite was deleted.

Fires the EventHandler::invite_delete event handler.

MessageCreate(MessageCreateEvent)

MessageDelete(MessageDeleteEvent)

MessageDeleteBulk(MessageDeleteBulkEvent)

MessageUpdate(MessageUpdateEvent)

A message has been edited, either by the user or the system

PresenceUpdate(PresenceUpdateEvent)

A member’s presence state (or username or avatar) has changed

PresencesReplace(PresencesReplaceEvent)

The presence list of the user’s friends should be replaced entirely

ReactionAdd(ReactionAddEvent)

A reaction was added to a message.

Fires the EventHandler::reaction_add event handler.

ReactionRemove(ReactionRemoveEvent)

A reaction was removed to a message.

Fires the EventHandler::reaction_remove event handler.

ReactionRemoveAll(ReactionRemoveAllEvent)

A request was issued to remove all Reactions from a Message.

Fires the EventHandler::reaction_remove_all event handler.

ReactionRemoveEmoji(ReactionRemoveEmojiEvent)

Sent when a bot removes all instances of a given emoji from the reactions of a message.

Fires the EventHandler::reaction_remove_emoji event handler.

Ready(ReadyEvent)

The first event in a connection, containing the initial ready cache.

May also be received at a later time in the event of a reconnect.

Resumed(ResumedEvent)

The connection has successfully resumed after a disconnect.

TypingStart(TypingStartEvent)

A user is typing; considered to last 5 seconds

UserUpdate(UserUpdateEvent)

Update to the logged-in user’s information

VoiceStateUpdate(VoiceStateUpdateEvent)

A member’s voice state has changed

VoiceServerUpdate(VoiceServerUpdateEvent)

Voice server information is available

VoiceChannelStatusUpdate(VoiceChannelStatusUpdateEvent)

Fired when the status of a Voice Channel changes.

WebhookUpdate(WebhookUpdateEvent)

A webhook for a channel was updated in a Guild.

InteractionCreate(InteractionCreateEvent)

An interaction was created.

IntegrationCreate(IntegrationCreateEvent)

A guild integration was created

IntegrationUpdate(IntegrationUpdateEvent)

A guild integration was updated

IntegrationDelete(IntegrationDeleteEvent)

A guild integration was deleted

StageInstanceCreate(StageInstanceCreateEvent)

A stage instance was created.

StageInstanceUpdate(StageInstanceUpdateEvent)

A stage instance was updated.

StageInstanceDelete(StageInstanceDeleteEvent)

A stage instance was deleted.

ThreadCreate(ThreadCreateEvent)

A thread was created or the current user was added to a private thread.

ThreadUpdate(ThreadUpdateEvent)

A thread was updated.

ThreadDelete(ThreadDeleteEvent)

A thread was deleted.

ThreadListSync(ThreadListSyncEvent)

The current user gains access to a channel.

ThreadMemberUpdate(ThreadMemberUpdateEvent)

The ThreadMember object for the current user is updated.

ThreadMembersUpdate(ThreadMembersUpdateEvent)

Anyone is added to or removed from a thread.

GuildScheduledEventCreate(GuildScheduledEventCreateEvent)

A scheduled event was created.

GuildScheduledEventUpdate(GuildScheduledEventUpdateEvent)

A scheduled event was updated.

GuildScheduledEventDelete(GuildScheduledEventDeleteEvent)

A scheduled event was deleted.

GuildScheduledEventUserAdd(GuildScheduledEventUserAddEvent)

A guild member has subscribed to a scheduled event.

GuildScheduledEventUserRemove(GuildScheduledEventUserRemoveEvent)

A guild member has unsubscribed from a scheduled event.

EntitlementCreate(EntitlementCreateEvent)

A user subscribed to a SKU.

EntitlementUpdate(EntitlementUpdateEvent)

A user’s entitlement was updated or renewed.

EntitlementDelete(EntitlementDeleteEvent)

A user’s entitlement was deleted by Discord, or refunded.

Unknown(UnknownEvent)

An event type not covered by the above

Implementations

impl Event

pub fn name(&self) -> Option<String>

Return the event name of this event. Returns None if the event is Unknown.