Module poise::serenity_prelude::gateway

Available on crate feature gateway only.

Contains the necessary plumping for maintaining a connection with Discord. The primary building blocks are the Client and the Shard.

The Client is a high-level interface that takes care of communicating with Discord’s REST API as well as receiving and dispatching events from the gateway using a WebSocket client.

On the other hand, the Shard is a low-level receiver and sender representing a single connection to Discord. The client will handle shard management automatically for you, so you should only care about using it directly if you really need to. See the sharding module for details and documentation.

Structs

ActivityData

Activity data of the current user.

CollectorCallbackcollector

Newtype around a callback that will be called on every incoming request. As long as this collector should still receive events, it should return true. Once it returns false, it is removed.

PresenceData

Presence data of the current user.

Shard

A Shard is an abstract handler for a websocket connection to Discord’s gateway. The shard allows for sending and receiving messages over the websocket, such as setting the active activity, reconnecting, syncing guilds, and more.

ShardManager

A manager for handling the status of shards by starting them, restarting them, and stopping them when required.

ShardManagerOptions

ShardMessenger

A handle to a ShardRunner.

ShardQueue

A queue of ShardIds that is split up into multiple buckets according to the value of max_concurrency.

ShardQueuer

The shard queuer is a simple loop that runs indefinitely to manage the startup of shards.

ShardRunner

A runner for managing a Shard and its respective WebSocket client.

ShardRunnerInfo

Information about a ShardRunner.

ShardRunnerOptions

Options to be passed to ShardRunner::new.

ShardStageUpdateEvent

An event denoting that a shard’s connection stage was changed.

WsClient

Enums

ChunkGuildFilter

Discord docs.

ConnectionStage

Indicates the current connection stage of a Shard.

GatewayError

An error that occurred while attempting to deal with the gateway.

ReconnectType

The type of reconnection that should be performed.

ShardAction

ShardQueuerMessage

A message to be sent to the ShardQueuer.

ShardRunnerMessage

A message to send from a shard over a WebSocket.