poise::serenity_prelude::model
Module channel

Models relating to channels and types within channels.

Structs

Attachment

A file uploaded with a message. Not to be confused with Embeds.

ChannelFlags

Describes extra features of the channel.

ChannelMention

Discord docs.

CountDetails

A representation of reaction count details.

Embed

Represents a rich embed which allows using richer markdown, multiple fields and more. This was heavily inspired by slack’s attachments.

EmbedAuthor

An author object in an embed.

EmbedField

A field object in an embed.

EmbedFooter

Footer information for an embed.

EmbedImage

An image object in an embed.

EmbedProvider

The provider of an embed.

EmbedThumbnail

The dimensions and URL of an embed thumbnail.

EmbedVideo

Video information for an embed.

FollowedChannel

A container for the IDs returned by following a news channel.

ForumTag

An object that represents a tag able to be applied to a thread in a GUILD_FORUM channel.

GuildChannel

Represents a guild’s text, news, or voice channel. Some methods are available only for voice channels and some are only available for text channels. News channels are a subset of text channels and lack slow mode hence Self::rate_limit_per_user will be None.

Message

A representation of a message over a guild’s text channel, a group, or a private channel.

MessageActivity

Rich Presence activity information.

MessageApplication

Rich Presence application information.

MessageFlags

Describes extra features of the message.

MessageReaction

A representation of a reaction to a message.

MessageReference

Reference data sent with crossposted messages.

MessagesItermodel

A helper class returned by ChannelId::messages_iter

PartialChannel

A container for any partial channel.

PartialGuildChannel

A partial guild channel.

PermissionOverwrite

A channel-specific permission overwrite for a member or role.

Poll

A poll that has been attached to a Message.

PollAnswer

A possible answer for a Poll.

PollAnswerCount

The count of a single PollAnswer’s results.

PollMedia

A piece of data used in mutliple parts of the Poll structure.

PollResults

The model for the results of a Poll.

PrivateChannel

A Direct Message text channel with another user.

Reaction

An emoji reaction to a message.

ReactionConversionError

RoleSubscriptionData

Discord docs

StageInstance

Discord docs.

ThreadMetadata

A thread data.

ThreadsData

A response to getting several threads channels.

Enums

AutoArchiveDuration

See ThreadMetadata::auto_archive_duration.

Channel

A container for any channel.

ChannelType

A representation of a type of channel.

ForumEmoji

An object that specifies the emoji to use for Forum related emoji parameters.

ForumLayoutType

See GuildChannel::default_forum_layout.

MessageActivityKind

Discord docs.

MessageType

Differentiates between regular and different types of system messages.

Nonce

PermissionOverwriteType

The type of edit being made to a Channel’s permissions.

PollLayoutType

Represents the different layouts that a Poll may have.

PollMediaEmoji

The “Partial Emoji” attached to a PollMedia model.

ReactionType

The type of a Reaction sent.

ReactionTypes

A list of types a reaction can be.

SortOrder

The sort order for threads in a forum.

StageInstancePrivacyLevel

See StageInstance::privacy_level.

VideoQualityMode

The video quality mode for a voice channel.

Type Aliases

AttachmentTypeDeprecatedmodel