Struct poise::serenity_prelude::nonmax::NonMaxI128

pub struct NonMaxI128(/* private fields */);

An integer that is known not to equal its maximum value.

Implementations

impl NonMaxI128

pub const ZERO: NonMaxI128 = _

Gets non-max with the value zero (0)

pub const ONE: NonMaxI128 = _

Gets non-max with the value one (1)

pub const MAX: NonMaxI128 = _

Gets non-max with maximum possible value (which is maximum of the underlying primitive minus one)

pub const fn new(value: i128) -> Option<NonMaxI128>

Creates a new non-max if the given value is not the maximum value.

pub const unsafe fn new_unchecked(value: i128) -> NonMaxI128

Creates a new non-max without checking the value.

Safety

The value must not equal the maximum representable value for the primitive type.

pub const fn get(&self) -> i128

Returns the value as a primitive type.