Crate poise::serenity_prelude::small_fixed_array

A crate for FixedArray and FixedString, types to provide a smaller memory footprint in exchange for:

Immutablity, FixedArray and FixedString cannot be mutated without converting back to their expanded forms.

Maximum length, FixedArray and FixedString have a length cap of LenT::MAX elements.

These types provide cheap conversions to Vec and String, to make up for most of these downsides, but it is still not recommended to use these collections for mutated values as you will see a performance downside.

These can be thought of as Box<[T]> and Box<str>, except the length is denoted as LenT, by default u32.

Features

nightly: Speeds up FixedString::len for small strings, using portable_simd.

serde: Provides serde implementations for FixedArray and FixedString.

typesize: Provides [typesize] implementations for FixedArray and FixedString.

MSRV

The Minimum Supported Rust Version of this crate is 1.70.

It is considered a breaking change to raise this.

Structs

FixedArray

A fixed size array with length provided at creation denoted in a ValidLength, by default u32.

FixedString

A fixed size String with length provided at creation denoted in ValidLength, by default u32.

Traits

TruncatingInto

A sealed helper trait for calling FixedArray<T>::from_vec_trunc or FixedString::from_string_trunc.

ValidLength

A sealed trait to represent valid lengths for a FixedArray.