Trait poise::serenity_prelude::small_fixed_array::TruncatingInto

pub trait TruncatingInto<T>: Sealed {
    // Required method
    fn trunc_into(self) -> T;
}

A sealed helper trait for calling FixedArray<T>::from_vec_trunc or FixedString::from_string_trunc.

Both of these functions may truncate the input in order to fit it into the provided ValidLength, therefore this trait must be imported in order to make possible truncation made obvious in user code.

Required Methods

fn trunc_into(self) -> T