Struct poise::serenity_prelude::CustomMessage

pub struct CustomMessage { /* private fields */ }

A builder for constructing a personal Message instance. This can be useful for emitting a manual dispatch to the framework, but you don’t have a message in hand, or just have a fragment of its data.

Implementations

impl CustomMessage

pub fn new() -> CustomMessage

Available on crate feature utils only.

Constructs a new instance of this builder, alongside a message with dummy data. Use the methods to replace the individual bits of this message with valid data.

pub fn id(&mut self, id: MessageId) -> &mut CustomMessage

Available on crate feature utils only.

Assign the dummy message a proper ID for identification.

If not used, the default value is MessageId::new(1).

pub fn attachments( &mut self, attachments: impl IntoIterator<Item = Attachment> ) -> &mut CustomMessage

Available on crate feature utils only.

Assign the dummy message files attached to it.

If not used, the default value is an empty vector (Vec::default()).

pub fn author(&mut self, user: User) -> &mut CustomMessage

Available on crate feature utils only.

Assign the dummy message its author.

If not used, the default value is a dummy User.

pub fn channel_id(&mut self, channel_id: ChannelId) -> &mut CustomMessage

Available on crate feature utils only.

Assign the dummy message its origin channel’s ID.

If not used, the default value is ChannelId::new(1).

pub fn content(&mut self, s: impl Into<String>) -> &mut CustomMessage

Available on crate feature utils only.

Assign the dummy message its contents.

If not used, the default value is an empty string (String::default()).

pub fn edited_timestamp<T>(&mut self, timestamp: T) -> &mut CustomMessage
where T: Into<Timestamp>,

Available on crate feature utils only.

Assign the dummy message the timestamp it was edited.

If not used, the default value is None (not all messages are edited).

pub fn embeds( &mut self, embeds: impl IntoIterator<Item = Embed> ) -> &mut CustomMessage

Available on crate feature utils only.

Assign the dummy message embeds.

If not used, the default value is an empty vector (Vec::default()).

pub fn guild_id(&mut self, guild_id: GuildId) -> &mut CustomMessage

Available on crate feature utils only.

Assign the dummy message its origin guild’s ID.

If not used, the default value is None (not all messages are sent in guilds).

pub fn kind(&mut self, kind: MessageType) -> &mut CustomMessage

Available on crate feature utils only.

Assign the dummy message its type.

If not used, the default value is MessageType::Regular.

pub fn member(&mut self, member: PartialMember) -> &mut CustomMessage

Available on crate feature utils only.

Assign the dummy message member data pertaining to its author.

If not used, the default value is None (not all messages are sent in guilds).

pub fn mention_everyone(&mut self, mentions: bool) -> &mut CustomMessage

Available on crate feature utils only.

Assign the dummy message a flag whether it mentions everyone (@everyone).

If not used, the default value is false.

pub fn mention_roles( &mut self, roles: impl IntoIterator<Item = RoleId> ) -> &mut CustomMessage

Available on crate feature utils only.

Assign the dummy message a list of roles it mentions.

If not used, the default value is an empty vector (Vec::default()).

pub fn mentions( &mut self, mentions: impl IntoIterator<Item = User> ) -> &mut CustomMessage

Available on crate feature utils only.

Assign the dummy message a list of mentions.

If not used, the default value is an empty vector (Vec::default()).

pub fn pinned(&mut self, pinned: bool) -> &mut CustomMessage

Available on crate feature utils only.

Assign the dummy message a flag whether it’s been pinned.

If not used, the default value is false.

pub fn reactions( &mut self, reactions: impl IntoIterator<Item = MessageReaction> ) -> &mut CustomMessage

Available on crate feature utils only.

Assign the dummy message a list of emojis it was reacted with.

If not used, the default value is an empty vector (Vec::default()).

pub fn timestamp<T>(&mut self, timestamp: T) -> &mut CustomMessage
where T: Into<Timestamp>,

Available on crate feature utils only.

Assign the dummy message the timestamp it was created at.

If not used, the default value is the current local time.

pub fn tts(&mut self, tts: bool) -> &mut CustomMessage

Available on crate feature utils only.

Assign the dummy message a flag whether it’ll be read by a Text-To-Speech program.

If not used, the default value is false.

pub fn webhook_id(&mut self, id: WebhookId) -> &mut CustomMessage

Available on crate feature utils only.

Assign the dummy message the webhook author’s ID.

If not used, the default value is None (not all messages are sent by webhooks).

pub fn build(self) -> Message

Available on crate feature utils only.

Consume this builder and return the constructed message.