poise::serenity_prelude
Struct EditMember

pub struct EditMember<'a> { /* private fields */ }

A builder which edits the properties of a Member, to be used in conjunction with Member::edit.

Discord docs

Implementations

impl<'a> EditMember<'a>

pub fn new() -> EditMember<'a>

Available on crate feature builder only.

Equivalent to Self::default.

pub fn deafen(self, deafen: bool) -> EditMember<'a>

Available on crate feature builder only.

Whether to deafen the member.

Note: Requires the Deafen Members permission.

pub fn mute(self, mute: bool) -> EditMember<'a>

Available on crate feature builder only.

Whether to mute the member.

Note: Requires the Mute Members permission.

pub fn nickname(self, nickname: impl Into<String>) -> EditMember<'a>

Available on crate feature builder only.

Changes the memberโs nickname. Pass an empty string to reset the nickname.

Note: Requires the Manage Nicknames permission.

pub fn roles( self, roles: impl IntoIterator<Item = impl Into<RoleId>>, ) -> EditMember<'a>

Available on crate feature builder only.

Set the list of roles that the member should have.

Note: Requires the Manage Roles permission to modify.

pub fn voice_channel(self, channel_id: impl Into<ChannelId>) -> EditMember<'a>

Available on crate feature builder only.

Move the member into a voice channel.

Note: Requires the Move Members permission.

pub fn disconnect_member(self) -> EditMember<'a>

Available on crate feature builder only.

Disconnects the user from their voice channel, if any.

Note: Requires the Move Members permission.

pub fn disable_communication_until(self, time: String) -> EditMember<'a>

Available on crate feature builder only.

Times the user out until time, an ISO8601-formatted datetime string.

time is considered invalid if it is not a valid ISO8601 timestamp or if it is greater than 28 days from the current time.

Note: Requires the Moderate Members permission.

pub fn disable_communication_until_datetime( self, time: Timestamp, ) -> EditMember<'a>

Available on crate feature builder only.

Times the user out until time.

time is considered invalid if it is greater than 28 days from the current time.

Note: Requires the Moderate Members permission.

pub fn enable_communication(self) -> EditMember<'a>

Available on crate feature builder only.

Allow a user to communicate, removing their timeout, if there is one.

Note: Requires the Moderate Members permission.

pub fn flags(self, flags: GuildMemberFlags) -> EditMember<'a>

Available on crate feature builder only.

pub fn audit_log_reason(self, reason: &'a str) -> EditMember<'a>

Available on crate feature builder only.

Sets the requestโs audit log reason.