Struct poise::serenity_prelude::EditVoiceState

pub struct EditVoiceState { /* private fields */ }

A builder which edits a user’s voice state, to be used in conjunction with GuildChannel::edit_voice_state.

Discord docs:

current user

other users

Implementations

impl EditVoiceState

pub fn new() -> EditVoiceState

Available on crate feature builder only.

Equivalent to Self::default.

pub fn suppress(self, deafen: bool) -> EditVoiceState

Available on crate feature builder only.

Whether to suppress the user. Setting this to false will invite a user to speak.

Note: Requires the Mute Members permission to suppress another user or unsuppress the current user. This is not required if suppressing the current user.

pub fn request_to_speak(self, request: bool) -> EditVoiceState

Available on crate feature builder only.

Requests or clears a request to speak. Passing true is equivalent to passing the current time to Self::request_to_speak_timestamp.

Note: Requires the Request to Speak permission.

pub fn request_to_speak_timestamp( self, timestamp: impl Into<Timestamp> ) -> EditVoiceState

Available on crate feature builder only.

Sets the current bot user’s request to speak timestamp. This can be any present or future time.

Note: Requires the Request to Speak permission.