Struct poise::serenity_prelude::PrivateChannel

#[non_exhaustive]
pub struct PrivateChannel {
    pub id: ChannelId,
    pub last_message_id: Option<MessageId>,
    pub last_pin_timestamp: Option<Timestamp>,
    pub kind: ChannelType,
    pub recipient: User,
}

A Direct Message text channel with another user.

Discord docs.

Fields (Non-exhaustive)

Non-exhaustive structs could have additional fields added in future. Therefore, non-exhaustive structs cannot be constructed in external crates using the traditional Struct { .. } syntax; cannot be matched against without a wildcard ..; and struct update syntax will not work.

id: ChannelId

The unique Id of the private channel.

Can be used to calculate the first message’s creation date.

last_message_id: Option<MessageId>

The Id of the last message sent.

last_pin_timestamp: Option<Timestamp>

Timestamp of the last time a Message was pinned.

kind: ChannelType

Indicator of the type of channel this is.

This should always be ChannelType::Private.

recipient: User

The recipient to the private channel.

Implementations

impl PrivateChannel

pub async fn broadcast_typing( &self, http: impl AsRef<Http> ) -> Result<(), Error>

Available on crate feature model only.

Broadcasts that the current user is typing to the recipient.

See ChannelId::broadcast_typing for more details.

pub async fn create_reaction( &self, http: impl AsRef<Http>, message_id: impl Into<MessageId>, reaction_type: impl Into<ReactionType> ) -> Result<(), Error>

Available on crate feature model only.

React to a Message with a custom Emoji or unicode character.

Message::react may be a more suited method of reacting in most cases.

Errors

Returns Error::Http if the reaction cannot be added, or if a message with that Id does not exist.

pub async fn delete( &self, http: impl AsRef<Http> ) -> Result<PrivateChannel, Error>

Available on crate feature model only.

Deletes the channel. This does not delete the contents of the channel, and is equivalent to closing a private channel on the client, which can be re-opened.

pub async fn delete_messages<T>( &self, http: impl AsRef<Http>, message_ids: impl IntoIterator<Item = T> ) -> Result<(), Error>
where T: AsRef<MessageId>,

Available on crate feature model only.

Deletes all messages by Ids from the given vector in the channel.

The minimum amount of messages is 2 and the maximum amount is 100.

Requires the Manage Messages permission.

Note: Messages that are older than 2 weeks can’t be deleted using this method.

Errors

Returns ModelError::BulkDeleteAmount if an attempt was made to delete either 0 or more than 100 messages.

pub async fn delete_permission( &self, http: impl AsRef<Http>, permission_type: PermissionOverwriteType ) -> Result<(), Error>

Available on crate feature model only.

Deletes all permission overrides in the channel from a member or role.

Note: Requires the Manage Channel permission.

pub async fn delete_reaction( &self, http: impl AsRef<Http>, message_id: impl Into<MessageId>, user_id: Option<UserId>, reaction_type: impl Into<ReactionType> ) -> Result<(), Error>

Available on crate feature model only.

Deletes the given Reaction from the channel.

Note: In private channels, the current user may only delete it’s own reactions.

Errors

Returns Error::Http if the reaction is not from the current user.

pub async fn edit_message( &self, cache_http: impl CacheHttp, message_id: impl Into<MessageId>, builder: EditMessage ) -> Result<Message, Error>

Available on crate feature model only.

Edits a Message in the channel given its Id.

Message editing preserves all unchanged message data, with some exceptions for embeds and attachments.

Note: In most cases requires that the current user be the author of the message.

Refer to the documentation for EditMessage for information regarding content restrictions and requirements.

Errors

See EditMessage::execute for a list of possible errors, and their corresponding reasons.

pub fn is_nsfw(&self) -> bool

Available on crate feature model only.

Determines if the channel is NSFW.

Note: This method is for consistency. This will always return false, due to DMs not being considered NSFW.

pub async fn message( &self, cache_http: impl CacheHttp, message_id: impl Into<MessageId> ) -> Result<Message, Error>

Available on crate feature model only.

Gets a message from the channel.

Errors

Returns Error::Http if a message with that Id does not exist in this channel.

pub async fn messages( &self, cache_http: impl CacheHttp, builder: GetMessages ) -> Result<Vec<Message>, Error>

Available on crate feature model only.

Gets messages from the channel.

Note: If the user does not have the Read Message History permission, returns an empty Vec.

Errors

Returns Error::Http if the current user lacks permission.

pub fn name(&self) -> String

Available on crate feature model only.

Returns “DM with $username#discriminator”.

pub async fn reaction_users( &self, http: impl AsRef<Http>, message_id: impl Into<MessageId>, reaction_type: impl Into<ReactionType>, limit: Option<u8>, after: impl Into<Option<UserId>> ) -> Result<Vec<User>, Error>

Available on crate feature model only.

Gets the list of Users who have reacted to a Message with a certain Emoji.

The default limit is 50 - specify otherwise to receive a different maximum number of users. The maximum that may be retrieve at a time is 100, if a greater number is provided then it is automatically reduced.

The optional after attribute is to retrieve the users after a certain user. This is useful for pagination.

Errors

Returns Error::Http if a message with the given Id does not exist in the channel.

pub async fn pin( &self, http: impl AsRef<Http>, message_id: impl Into<MessageId> ) -> Result<(), Error>

Available on crate feature model only.

Pins a Message to the channel.

Errors

Returns Error::Http if the number of pinned messages would exceed the 50 message limit.

pub async fn pins(&self, http: impl AsRef<Http>) -> Result<Vec<Message>, Error>

Available on crate feature model only.

Retrieves the list of messages that have been pinned in the private channel.

pub async fn say( &self, cache_http: impl CacheHttp, content: impl Into<String> ) -> Result<Message, Error>

Available on crate feature model only.

Sends a message with just the given message content in the channel.

Note: Message content must be under 2000 unicode code points.

Errors

Returns a ModelError::MessageTooLong if the content length is over the above limit. See CreateMessage::execute for more details.

pub async fn send_files( self, cache_http: impl CacheHttp, files: impl IntoIterator<Item = CreateAttachment>, builder: CreateMessage ) -> Result<Message, Error>

Available on crate feature model only.

Sends file(s) along with optional message contents.

Refer to ChannelId::send_files for examples and more information.

Errors

See CreateMessage::execute for a list of possible errors, and their corresponding reasons.

pub async fn send_message( &self, cache_http: impl CacheHttp, builder: CreateMessage ) -> Result<Message, Error>

Available on crate feature model only.

Sends a message to the channel.

Refer to the documentation for CreateMessage for information regarding content restrictions and requirements.

Errors

See CreateMessage::execute for a list of possible errors, and their corresponding reasons.

pub fn start_typing(self, http: &Arc<Http>) -> Typing

Available on crate feature model only.

Starts typing in the channel for an indefinite period of time.

Returns Typing that is used to trigger the typing. Typing::stop must be called on the returned struct to stop typing. Note that on some clients, typing may persist for a few seconds after Typing::stop is called. Typing is also stopped when the struct is dropped.

If a message is sent while typing is triggered, the user will stop typing for a brief period of time and then resume again until either Typing::stop is called or the struct is dropped.

This should rarely be used for bots, although it is a good indicator that a long-running command is still being processed.

Examples

// Initiate typing (assuming http is `Arc<Http>` and `channel` is bound)
let typing = channel.start_typing(&http);

// Run some long-running process
long_process();

// Stop typing
typing.stop();

Errors

May return Error::Http if the current user cannot send a direct message to this user.

pub async fn unpin( &self, http: impl AsRef<Http>, message_id: impl Into<MessageId> ) -> Result<(), Error>

Available on crate feature model only.

Unpins a Message in the channel given by its Id.

Errors

Returns Error::Http if the current user lacks permission, if the message was deleted, or if the channel already has the limit of 50 pinned messages.