poise

Module slash_argument

Application command argument handling code

Macros

parse_slash_args

Macro for extracting and parsing slash command arguments out of an array of serenity::CommandDataOption.

Enums

SlashArgError

Possible errors when parsing slash command arguments

Traits

ContextMenuParameter

Implemented for all types that can be used in a context menu command

SlashArgument

Implement this trait on types that you want to use as a slash command parameter.