poise::slash_argument

Macro parse_slash_args

macro_rules! parse_slash_args {
    ($ctx:expr, $interaction:expr, $args:expr => $(
        ( $name:literal: $($type:tt)* )
    ),* $(,)? ) => { ... };
}

Macro for extracting and parsing slash command arguments out of an array of serenity::CommandDataOption.

An invocation of this macro is generated by crate::command, so you usually don’t need this macro directly.

let ctx: serenity::Context = todo!();
let interaction: serenity::CommandInteraction = todo!();
let args: &[serenity::ResolvedOption] = todo!();

let (arg1, arg2) = poise::parse_slash_args!(
    &ctx, &interaction,
    args => ("arg1": String), ("arg2": Option<u32>)
).await?;