Struct poise::Command

pub struct Command<U, E> {

    pub prefix_action: Option<for<'a> fn(_: PrefixContext<'a, U, E>) -> BoxFuture<'a, Result<(), FrameworkError<'a, U, E>>>>,
    pub slash_action: Option<for<'a> fn(_: ApplicationContext<'a, U, E>) -> BoxFuture<'a, Result<(), FrameworkError<'a, U, E>>>>,
    pub context_menu_action: Option<ContextMenuCommandAction<U, E>>,
    pub subcommands: Vec<Command<U, E>>,
    pub name: String,
    pub name_localizations: HashMap<String, String>,
    pub qualified_name: String,
    pub identifying_name: String,
    pub category: Option<&'static str>,
    pub hide_in_help: bool,
    pub description: Option<String>,
    pub description_localizations: HashMap<String, String>,
    pub help_text: Option<fn() -> String>,
    pub cooldowns: Mutex<Cooldowns>,
    pub reuse_response: bool,
    pub default_member_permissions: Permissions,
    pub required_permissions: Permissions,
    pub required_bot_permissions: Permissions,
    pub owners_only: bool,
    pub guild_only: bool,
    pub dm_only: bool,
    pub nsfw_only: bool,
    pub on_error: Option<fn(_: FrameworkError<'_, U, E>) -> BoxFuture<'_, ()>>,
    pub checks: Vec<fn(_: Context<'_, U, E>) -> BoxFuture<'_, Result<bool, E>>>,
    pub parameters: Vec<CommandParameter<U, E>>,
    pub custom_data: Box<dyn Any + Send + Sync>,
    pub aliases: &'static [&'static str],
    pub invoke_on_edit: bool,
    pub broadcast_typing: bool,
    pub context_menu_name: Option<&'static str>,
    pub ephemeral: bool,
    /* private fields */

}

Type returned from #[poise::command] annotated functions, which contains all of the generated prefix and application commands

Fields
prefix_action: Option<for<'a> fn(_: PrefixContext<'a, U, E>) -> BoxFuture<'a, Result<(), FrameworkError<'a, U, E>>>>

Callback to execute when this command is invoked in a prefix context

slash_action: Option<for<'a> fn(_: ApplicationContext<'a, U, E>) -> BoxFuture<'a, Result<(), FrameworkError<'a, U, E>>>>

Callback to execute when this command is invoked in a slash context

context_menu_action: Option<ContextMenuCommandAction<U, E>>

Callback to execute when this command is invoked in a context menu context

The enum variant shows which Discord item this context menu command works on

subcommands: Vec<Command<U, E>>

Subcommands of this command, if any

name: String

Main name of the command. Aliases (prefix-only) can be set in Self::aliases.

name_localizations: HashMap<String, String>

Localized names with locale string as the key (slash-only)

qualified_name: String

Full name including parent command names.

Initially set to just Self::name and properly populated when the framework is started.

identifying_name: String

A string to identify this particular command within a list of commands.

Can be configured via the crate::command macro (though it’s probably not needed for most bots). If not explicitly configured, it falls back to the command function name.

category: Option<&'static str>

Identifier for the category that this command will be displayed in for help commands.

hide_in_help: bool

Whether to hide this command in help menus.

description: Option<String>

Short description of the command. Displayed inline in help menus and similar.

description_localizations: HashMap<String, String>

Localized descriptions with locale string as the key (slash-only)

help_text: Option<fn() -> String>

Multiline description with detailed usage instructions. Displayed in the command specific help: ~help command_name

cooldowns: Mutex<Cooldowns>

Handles command cooldowns. Mainly for framework internal use

reuse_response: bool

After the first response, whether to post subsequent responses as edits to the initial message

Note: in prefix commands, this only has an effect if crate::PrefixFrameworkOptions::edit_tracker is set.

default_member_permissions: Permissions

Permissions which users must have to invoke this command. Used by Discord to set who can invoke this as a slash command. Not used on prefix commands or checked internally.

Set to serenity::Permissions::empty() by default

required_permissions: Permissions

Permissions which users must have to invoke this command. This is checked internally and works for both prefix commands and slash commands.

Set to serenity::Permissions::empty() by default

required_bot_permissions: Permissions

Permissions without which command execution will fail. You can set this to fail early and give a descriptive error message in case the bot hasn’t been assigned the minimum permissions by the guild admin.

Set to serenity::Permissions::empty() by default

owners_only: bool

If true, only users from the owners list may use this command.

guild_only: bool

If true, only people in guilds may use this command

dm_only: bool

If true, the command may only run in DMs

nsfw_only: bool

If true, the command may only run in NSFW channels

on_error: Option<fn(_: FrameworkError<'_, U, E>) -> BoxFuture<'_, ()>>

Command-specific override for crate::FrameworkOptions::on_error

checks: Vec<fn(_: Context<'_, U, E>) -> BoxFuture<'_, Result<bool, E>>>

If any of these functions returns false, this command will not be executed.

parameters: Vec<CommandParameter<U, E>>

List of parameters for this command

Used for registering and parsing slash commands. Can also be used in help commands

custom_data: Box<dyn Any + Send + Sync>

Arbitrary data, useful for storing custom metadata about your commands

aliases: &'static [&'static str]

Alternative triggers for the command (prefix-only)

invoke_on_edit: bool

Whether to rerun the command if an existing invocation message is edited (prefix-only)

broadcast_typing: bool

Whether to broadcast a typing indicator while executing this commmand (prefix-only)

context_menu_name: Option<&'static str>

Context menu specific name for this command, displayed in Discord’s context menu

ephemeral: bool

Whether responses to this command should be ephemeral by default (application-only)

Implementations

impl<U, E> Command<U, E>

pub fn create_as_slash_command(&self) -> Option<CreateApplicationCommand>

Generates a slash command builder from this Command instance. This can be used to register this command on Discord’s servers

pub fn create_as_context_menu_command(&self) -> Option<CreateApplicationCommand>

Generates a context menu command builder from this Command instance. This can be used to register this command on Discord’s servers

pub fn category(&mut self, category: &'static str) -> &mut Self

👎Deprecated:
Please use poise::Command { category: "...", ..command() } instead

Deprecated

pub fn subcommand(
&mut self,
subcommand: Command<U, E>,
meta_builder: impl FnOnce(&mut Self) -> &mut Self
) -> &mut Self

👎Deprecated:
Please use poise::Command { subcommands: vec![...], ..command() } instead

Insert a subcommand