Struct poise::CommandBuilder

pub struct CommandBuilder<U, E> { /* fields omitted */ }

Builder struct to add a command to the framework

Implementations

impl<U, E> CommandBuilder<U, E>

pub fn category(&mut self, _category: &'static str) -> &mut Self

👎 Deprecated:
Please use category = "..." on the command attribute instead

Deprecated

pub fn subcommand(
&mut self,
definition: CommandDefinition<U, E>,
meta_builder: impl FnOnce(&mut Self) -> &mut Self
) -> &mut Self

Insert a subcommand

Examples found in repository

examples/serenity_example_port/main.rs (line 297)

↕

236
237
238
239
240
241
242
243
244
245
246
247
248
249
250
251
252
253
254
255
256
257
258
259
260
261
262
263
264
265
266
267
268
269
270
271
272
273
274
275
276
277
278
279
280
281
282
283
284
285
286
287
288
289
290
291
292
293
294
295
296
297
298
299
300
301
302
303
304
305
306
307
308
309
310
311
312
313
314
315
316
317
318
319
320
321
322
323
324
325
326
327
328
329
330
331
332
333
334

fn main() {
    let options = poise::FrameworkOptions {
        listener: |ctx, event, framework, user_data| {
            Box::pin(event_listener(ctx, event, framework, user_data))
        },
        on_error: |error, ctx| Box::pin(on_error(error, ctx)),
        // Set a function to be called prior to each command execution. This
        // provides all context of the command that would also be passed to the actual command code
        pre_command: |ctx| Box::pin(pre_command(ctx)),
        // Similar to `pre_command`, except will be called directly _after_
        // command execution.
        post_command: |ctx| Box::pin(post_command(ctx)),

        // Options specific to prefix commands, i.e. commands invoked via chat messages
        prefix_options: poise::PrefixFrameworkOptions {
            prefix: Some(String::from("~")),

            mention_as_prefix: false,
            // An edit tracker needs to be supplied here to make edit tracking in commands work
            edit_tracker: Some(poise::EditTracker::for_timespan(
                std::time::Duration::from_secs(3600 * 3),
            )),
            ..Default::default()
        },

        ..Default::default()
    };

    // The Framework builder will automatically retrieve the bot owner and application ID via the
    // passed token, so that information need not be passed here
    poise::Framework::build()
        // Configure the client with your Discord bot token in the environment.
        .token(std::env::var("DISCORD_TOKEN").expect("Expected DISCORD_TOKEN in the environment"))
        .options(options)
        .user_data_setup(|_ctx, _data_about_bot, _framework| {
            Box::pin(async move {
                Ok(Data {
                    command_counter: std::sync::Mutex::new(std::collections::HashMap::new()),
                })
            })
        })
        // The `#[poise::command(prefix_command, slash_command)]` macro transforms the function into
        // `fn() -> poise::CommandDefinition`.
        // Therefore, you need to call the command function without any arguments to get the
        // command definition instance to pass to the framework
        .command(help(), |f| f)
        // This function registers slash commands on Discord. When you change something about a
        // command signature, for example by changing its name, adding or removing parameters, or
        // changing a parameter type, you should call this function.
        .command(register(), |f| f)
        .command(about(), |f| f)
        .command(am_i_admin(), |f| f)
        .command(say(), |f| f)
        .command(commands(), |f| f)
        .command(ping(), |f| f)
        .command(latency(), |f| f)
        .command(some_long_command(), |f| f)
        .command(upper_command(), |f| {
            // A command can have sub-commands, just like in command lines tools.
            // Imagine `cargo help` and `cargo help run`.
            // Subcommands are also specified here, inside the builder
            f.subcommand(sub(), |f| f)
        })
        .command(bird(), |f| f)
        .command(cat(), |f| f)
        .command(dog(), |f| f)
        .command(multiply(), |f| f)
        .command(slow_mode(), |f| f)
        .run()
        .await
        .expect("Client error");

    // INFO: currently not supported by poise
    /*
    // Set a function that's called whenever an attempted command-call's
    // command could not be found.
    .unrecognised_command(unknown_command)
    // Set a function that's called whenever a message is not a command.
    .normal_message(normal_message)
    // Set a function that's called whenever a command's execution didn't complete for one
    // reason or another. For example, when a user has exceeded a rate-limit or a command
    // can only be performed by the bot owner.
    .on_dispatch_error(dispatch_error)
    // Can't be used more than once per 5 seconds:
    .bucket("emoji", |b| b.delay(5)).await
    // Can't be used more than 2 times per 30 seconds, with a 5 second delay applying per channel.
    // Optionally `await_ratelimits` will delay until the command can be executed instead of
    // cancelling the command invocation.
    .bucket("complicated", |b| b.limit(2).time_span(30).delay(5)
        // The target each bucket will apply to.
        .limit_for(LimitedFor::Channel)
        // The maximum amount of command invocations that can be delayed per target.
        // Setting this to 0 (default) will never await/delay commands and cancel the invocation.
        .await_ratelimits(1)
        // A function to call when a rate limit leads to a delay.
        .delay_action(delay_action)
    ).await
    */
}