Struct poise::CommandId

pub struct CommandId {
    pub identifying_name: String,
    pub category: Option<&'static str>,
    pub hide_in_help: bool,
    pub inline_help: Option<&'static str>,
    pub multiline_help: Option<fn() -> String>,
    pub cooldowns: Mutex<Cooldowns>,
    pub required_permissions: Permissions,
    pub required_bot_permissions: Permissions,
    pub owners_only: bool,
}

This struct holds all data shared across different command types of the same implementation.

For example with a #[command(prefix_command, slash_command)], the generated crate::PrefixCommand and crate::SlashCommand will both contain an Arc<CommandId> pointing to the same CommandId instance.

Fields
identifying_name: String

A string to identify this particular command within a list of commands.

Can be configured via the crate::command macro (though it’s probably not needed for most bots). If not explicitly configured, it falls back to prefix command name, slash command name, or context menu command name (in that order).

category: Option<&'static str>

Identifier for the category that this command will be displayed in for help commands.

hide_in_help: bool

Whether to hide this command in help menus.

inline_help: Option<&'static str>

Short description of the command. Displayed inline in help menus and similar.

multiline_help: Option<fn() -> String>

Multiline description with detailed usage instructions. Displayed in the command specific help: ~help command_name

cooldowns: Mutex<Cooldowns>

Handles command cooldowns. Mainly for framework internal use

required_permissions: Permissions

Permissions which users must have to invoke this command.

Set to serenity::Permissions::empty() by default

required_bot_permissions: Permissions

Permissions without which command execution will fail. You can set this to fail early and give a descriptive error message in case the bot hasn’t been assigned the minimum permissions by the guild admin.

Set to serenity::Permissions::empty() by default

owners_only: bool

If true, only users from the owners list may use this command.