Struct poise::CreateReply

pub struct CreateReply<'a> {
    pub content: Option<String>,
    pub embed: Option<CreateEmbed>,
    pub attachments: Vec<AttachmentType<'a>>,
    pub ephemeral: bool,
    pub components: Option<CreateComponents>,
}

Message builder that abstracts over prefix and application command responses

Fields
content: Option<String>

Message content

embed: Option<CreateEmbed>

Embed, if present

attachments: Vec<AttachmentType<'a>>

Message attachments

ephemeral: bool

Whether the message is ephemeral (only has an effect in application commands)

components: Option<CreateComponents>

Message components, that is, buttons

Implementations

impl<'a> CreateReply<'a>

pub fn content(&mut self, content: impl Into<String>) -> &mut Self

Set the content of the message.

Examples found in repository

examples/framework_usage/commands.rs (line 107)

↕

103
104
105
106
107
108
109
110
111
112
113
114
115
116
117
118
119
120
121
122
123
124
125
126
127
128
129
130
131
132
133
134
135
136
137
138
139
140
141
142

pub async fn boop(ctx: Context<'_>) -> Result<(), Error> {
    let uuid_boop = ctx.id();

    ctx.send(|m| {
        m.content("I want some boops!").components(|c| {
            c.create_action_row(|ar| {
                ar.create_button(|b| {
                    b.style(serenity::ButtonStyle::Primary)
                        .label("Boop me!")
                        .custom_id(uuid_boop)
                })
            })
        })
    })
    .await?;

    let mut boop_count = 0;
    while let Some(mci) = serenity::CollectComponentInteraction::new(ctx.discord())
        .author_id(ctx.author().id)
        .channel_id(ctx.channel_id())
        .timeout(std::time::Duration::from_secs(120))
        .filter(move |mci| mci.data.custom_id == uuid_boop.to_string())
        .await
    {
        boop_count += 1;

        let mut msg = mci.message.clone();
        msg.edit(ctx.discord(), |m| {
            m.content(format!("Boop count: {}", boop_count))
        })
        .await?;

        mci.create_interaction_response(ctx.discord(), |ir| {
            ir.kind(serenity::InteractionResponseType::DeferredUpdateMessage)
        })
        .await?;
    }

    Ok(())
}

pub fn embed(
&mut self,
f: impl FnOnce(&mut CreateEmbed) -> &mut CreateEmbed
) -> &mut Self

Set an embed for the message.

Any previously set embed will be overwritten.

pub fn components(
&mut self,
f: impl FnOnce(&mut CreateComponents) -> &mut CreateComponents
) -> &mut Self

Set components (buttons) for this message.

Any previously set components will be overwritten.

Examples found in repository

examples/framework_usage/commands.rs (lines 107-115)

↕

103
104
105
106
107
108
109
110
111
112
113
114
115
116
117
118
119
120
121
122
123
124
125
126
127
128
129
130
131
132
133
134
135
136
137
138
139
140
141
142

pub async fn boop(ctx: Context<'_>) -> Result<(), Error> {
    let uuid_boop = ctx.id();

    ctx.send(|m| {
        m.content("I want some boops!").components(|c| {
            c.create_action_row(|ar| {
                ar.create_button(|b| {
                    b.style(serenity::ButtonStyle::Primary)
                        .label("Boop me!")
                        .custom_id(uuid_boop)
                })
            })
        })
    })
    .await?;

    let mut boop_count = 0;
    while let Some(mci) = serenity::CollectComponentInteraction::new(ctx.discord())
        .author_id(ctx.author().id)
        .channel_id(ctx.channel_id())
        .timeout(std::time::Duration::from_secs(120))
        .filter(move |mci| mci.data.custom_id == uuid_boop.to_string())
        .await
    {
        boop_count += 1;

        let mut msg = mci.message.clone();
        msg.edit(ctx.discord(), |m| {
            m.content(format!("Boop count: {}", boop_count))
        })
        .await?;

        mci.create_interaction_response(ctx.discord(), |ir| {
            ir.kind(serenity::InteractionResponseType::DeferredUpdateMessage)
        })
        .await?;
    }

    Ok(())
}

pub fn attachment(&mut self, attachment: AttachmentType<'a>) -> &mut Self

Add an attachment.

This will not have an effect in a slash command’s initial response!

pub fn ephemeral(&mut self, ephemeral: bool) -> &mut Self

Toggles whether the message is an ephemeral response (only invoking user can see it).

This only has an effect in slash commands!

If this is the initial response and this response has previously been deferred, the ephemerality is decided by the defer operation. I.e. if you deferred the response without enabling ephemeral, the initial response will not be ephemeral.