Struct poise::EditTracker

pub struct EditTracker { /* fields omitted */ }

Stores messages and the associated bot responses in order to implement poise’s edit tracking feature.

Implementations

impl EditTracker

pub fn for_timespan(duration: Duration) -> RwLock<Self>

Create an edit tracker which tracks messages for the specified duration.

Note: EditTracker will only purge messages outside the duration when Self::purge is called. If you supply the created EditTracker to crate::Framework, the framework will take care of that by calling Self::purge periodically.

Examples found in repository

examples/framework_usage/main.rs (line 59)

↕

55
56
57
58
59
60
61
62
63
64
65
66
67
68
69
70
71
72
73
74
75
76
77
78
79
80
81
82
83
84
85
86
87
88
89
90
91
92
93
94
95
96
97
98
99
100
101
102
103

fn main() {
    let options = poise::FrameworkOptions {
        prefix_options: poise::PrefixFrameworkOptions {
            prefix: Some("~".into()),
            edit_tracker: Some(poise::EditTracker::for_timespan(Duration::from_secs(3600))),
            additional_prefixes: vec![
                poise::Prefix::Literal("hey bot"),
                poise::Prefix::Literal("hey bot,"),
            ],
            ..Default::default()
        },
        on_error: |error, ctx| Box::pin(on_error(error, ctx)),
        pre_command: |ctx| {
            Box::pin(async move {
                println!("Executing command {}...", ctx.command().unwrap().name());
            })
        },
        post_command: |ctx| {
            Box::pin(async move {
                println!("Executed command {}!", ctx.command().unwrap().name());
            })
        },
        ..Default::default()
    };

    poise::Framework::build()
        .token(var("TOKEN").expect("Missing `TOKEN` env var, see README for more information."))
        .user_data_setup(move |_ctx, _ready, _framework| {
            Box::pin(async move {
                Ok(Data {
                    votes: Mutex::new(HashMap::new()),
                })
            })
        })
        .options(options)
        .command(help(), |f| f)
        .command(register(), |f| f)
        .command(commands::vote(), |f| f)
        .command(commands::getvotes(), |f| f)
        .command(commands::add(), |f| f)
        .command(commands::choice(), |f| f)
        .command(commands::boop(), |f| f)
        .command(context_menu::user_info(), |f| f)
        .command(context_menu::echo(), |f| f)
        .command(autocomplete::greet(), |f| f)
        .run()
        .await
        .unwrap();
}

pub fn process_message_update(
&mut self,
user_msg_update: &MessageUpdateEvent,
ignore_edit_tracker_cache: bool
) -> Option<Message>

Returns a copy of a newly up-to-date cached message, or a brand new generated message when not in cache

Returns None if the command shouldn’t be re-run, e.g. if the message content stayed the same

pub fn purge(&mut self)

Forget all of the messages that are older than the specified duration.

pub fn find_bot_response(
&mut self,
user_msg_id: MessageId
) -> Option<&mut Message>

Given a message by a user, find the corresponding bot response, if one exists and is cached.