Struct poise::Framework

pub struct Framework<U, E> { /* private fields */ }

The main framework struct which stores all data and handles message and interaction dispatch.

Technically, this is just an optional abstraction over crate::dispatch_event with some additional conveniences built-in:

fills in correct values for crate::Command::qualified_name: set_qualified_names

spawns a background task to periodically clear edit tracker cache

sets up user data on the first Ready event

keeps track of shard manager and bot ID automatically

You can build a bot without Framework: see the manual_dispatch example in the repository

Implementations

impl<U, E> Framework<U, E>

pub fn build() -> FrameworkBuilder<U, E>

👎Deprecated:
Please use Framework::builder instead

Create a framework builder to configure, create and run a framework.

For more information, see FrameworkBuilder

Examples found in repository?

examples/fluent_localization/main.rs (line 46)

↕

39
40
41
42
43
44
45
46
47
48
49
50
51
52
53
54
55
56
57

async fn main() {
    env_logger::init();

    let mut commands = vec![welcome(), register()];
    let translations = translation::read_ftl().expect("failed to read translation files");
    translation::apply_translations(&translations, &mut commands);

    poise::Framework::build()
        .token(std::env::var("TOKEN").unwrap())
        .intents(serenity::GatewayIntents::non_privileged())
        .options(poise::FrameworkOptions {
            commands,
            ..Default::default()
        })
        .user_data_setup(move |_, _, _| Box::pin(async move { Ok(Data { translations }) }))
        .run()
        .await
        .unwrap();
}

pub fn builder() -> FrameworkBuilder<U, E>

Create a framework builder to configure, create and run a framework.

For more information, see FrameworkBuilder

Examples found in repository?

examples/quickstart/main.rs (line 28)

↕

27
28
29
30
31
32
33
34
35
36
37
38

async fn main() {
    let framework = poise::Framework::builder()
        .options(poise::FrameworkOptions {
            commands: vec![age(), register()],
            ..Default::default()
        })
        .token(std::env::var("DISCORD_TOKEN").expect("missing DISCORD_TOKEN"))
        .intents(serenity::GatewayIntents::non_privileged())
        .user_data_setup(move |_ctx, _ready, _framework| Box::pin(async move { Ok(Data {}) }));

    framework.run().await.unwrap();
}

Hide additional examples

examples/invocation_data/main.rs (line 50)

↕

49
50
51
52
53
54
55
56
57
58
59
60
61
62
63
64
65
66
67
68
69
70
71
72
73
74
75
76
77
78
79
80
81
82
83
84
85
86
87
88
89
90
91
92
93
94
95
96
97
98
99
100
101
102
103
104
105
106
107
108
109
110
111
112
113
114
115
116
117
118
119
120
121
122

async fn main() {
    poise::Framework::builder()
        .token(std::env::var("TOKEN").unwrap())
        .user_data_setup(move |ctx, _, framework| {
            Box::pin(async move {
                poise::serenity_prelude::GuildId(703332075914264606)
                    .set_application_commands(ctx, |b| {
                        *b = poise::samples::create_application_commands(
                            &framework.options().commands,
                        );
                        b
                    })
                    .await
                    .unwrap();
                Ok(())
            })
        })
        .options(poise::FrameworkOptions {
            pre_command: |ctx| {
                Box::pin(async move {
                    println!(
                        "In pre_command: {:?}",
                        ctx.invocation_data::<&str>().await.as_deref()
                    );
                })
            },
            command_check: Some(|ctx| {
                Box::pin(async move {
                    // Global command check is the first callback that's invoked, so let's set the
                    // data here
                    println!("Writing invocation data!");
                    ctx.set_invocation_data("hello").await;

                    println!(
                        "In global check: {:?}",
                        ctx.invocation_data::<&str>().await.as_deref()
                    );

                    Ok(true)
                })
            }),
            post_command: |ctx| {
                Box::pin(async move {
                    println!(
                        "In post_command: {:?}",
                        ctx.invocation_data::<&str>().await.as_deref()
                    );
                })
            },
            on_error: |err| {
                Box::pin(async move {
                    match err {
                        poise::FrameworkError::Command { ctx, .. } => {
                            println!(
                                "In on_error: {:?}",
                                ctx.invocation_data::<&str>().await.as_deref()
                            );
                        }
                        err => poise::samples::on_error(err).await.unwrap(),
                    }
                })
            },

            commands: vec![invocation_data_test()],
            prefix_options: poise::PrefixFrameworkOptions {
                prefix: Some("~".into()),
                ..Default::default()
            },
            ..Default::default()
        })
        .run()
        .await
        .unwrap();
}

examples/framework_usage/main.rs (line 143)

↕

66
67
68
69
70
71
72
73
74
75
76
77
78
79
80
81
82
83
84
85
86
87
88
89
90
91
92
93
94
95
96
97
98
99
100
101
102
103
104
105
106
107
108
109
110
111
112
113
114
115
116
117
118
119
120
121
122
123
124
125
126
127
128
129
130
131
132
133
134
135
136
137
138
139
140
141
142
143
144
145
146
147
148
149
150
151
152
153
154
155
156
157
158
159

async fn main() {
    env_logger::init();

    let options = poise::FrameworkOptions {
        commands: vec![
            help(),
            register(),
            general::vote(),
            general::getvotes(),
            general::addmultiple(),
            general::choice(),
            general::boop(),
            general::voiceinfo(),
            general::test_reuse_response(),
            general::oracle(),
            general::code(),
            general::say(),
            general::file_details(),
            general::totalsize(),
            general::modal(),
            general::punish(),
            context_menu::user_info(),
            context_menu::echo(),
            autocomplete::greet(),
            checks::shutdown(),
            checks::modonly(),
            checks::delete(),
            checks::ferrisparty(),
            checks::add(),
            checks::get_guild_name(),
            checks::only_in_dms(),
            checks::lennyface(),
            checks::permissions_v2(),
            subcommands::parent(),
            localization::welcome(),
        ],
        prefix_options: poise::PrefixFrameworkOptions {
            prefix: Some("~".into()),
            edit_tracker: Some(poise::EditTracker::for_timespan(Duration::from_secs(3600))),
            additional_prefixes: vec![
                poise::Prefix::Literal("hey bot"),
                poise::Prefix::Literal("hey bot,"),
            ],
            ..Default::default()
        },
        /// The global error handler for all error cases that may occur
        on_error: |error| Box::pin(on_error(error)),
        /// This code is run before every command
        pre_command: |ctx| {
            Box::pin(async move {
                println!("Executing command {}...", ctx.command().qualified_name);
            })
        },
        /// This code is run after a command if it was successful (returned Ok)
        post_command: |ctx| {
            Box::pin(async move {
                println!("Executed command {}!", ctx.command().qualified_name);
            })
        },
        /// Every command invocation must pass this check to continue execution
        command_check: Some(|ctx| {
            Box::pin(async move {
                if ctx.author().id == 123456789 {
                    return Ok(false);
                }
                Ok(true)
            })
        }),
        listener: |_ctx, event, _framework, _data| {
            Box::pin(async move {
                println!("Got an event in listener: {:?}", event.name());
                Ok(())
            })
        },
        ..Default::default()
    };

    poise::Framework::builder()
        .token(var("TOKEN").expect("Missing `TOKEN` env var, see README for more information."))
        .user_data_setup(move |_ctx, _ready, _framework| {
            Box::pin(async move {
                Ok(Data {
                    votes: Mutex::new(HashMap::new()),
                })
            })
        })
        .options(options)
        .intents(
            serenity::GatewayIntents::non_privileged() | serenity::GatewayIntents::MESSAGE_CONTENT,
        )
        .run()
        .await
        .unwrap();
}

examples/serenity_example_port/main.rs (line 289)

↕

235
236
237
238
239
240
241
242
243
244
245
246
247
248
249
250
251
252
253
254
255
256
257
258
259
260
261
262
263
264
265
266
267
268
269
270
271
272
273
274
275
276
277
278
279
280
281
282
283
284
285
286
287
288
289
290
291
292
293
294
295
296
297
298
299
300
301
302
303
304
305
306
307
308
309
310
311
312
313
314
315
316
317
318
319
320
321
322
323
324
325
326
327
328
329
330

async fn main() {
    let options = poise::FrameworkOptions {
        commands: vec![
            // The `#[poise::command(prefix_command, slash_command)]` macro transforms the function into
            // `fn() -> poise::Command`.
            // Therefore, you need to call the command function without any arguments to get the
            // command definition instance to pass to the framework
            help(),
            // This function registers slash commands on Discord. When you change something about a
            // command signature, for example by changing its name, adding or removing parameters, or
            // changing a parameter type, you should call this function.
            register(),
            about(),
            am_i_admin(),
            say(),
            commands(),
            ping(),
            latency(),
            some_long_command(),
            upper_command(),
            bird(),
            cat(),
            dog(),
            multiply(),
            slow_mode(),
        ],
        listener: |ctx, event, framework, user_data| {
            Box::pin(event_listener(ctx, event, framework, user_data))
        },
        on_error: |error| Box::pin(on_error(error)),
        // Set a function to be called prior to each command execution. This
        // provides all context of the command that would also be passed to the actual command code
        pre_command: |ctx| Box::pin(pre_command(ctx)),
        // Similar to `pre_command`, except will be called directly _after_
        // command execution.
        post_command: |ctx| Box::pin(post_command(ctx)),

        // Options specific to prefix commands, i.e. commands invoked via chat messages
        prefix_options: poise::PrefixFrameworkOptions {
            prefix: Some(String::from("~")),

            mention_as_prefix: false,
            // An edit tracker needs to be supplied here to make edit tracking in commands work
            edit_tracker: Some(poise::EditTracker::for_timespan(
                std::time::Duration::from_secs(3600 * 3),
            )),
            ..Default::default()
        },

        ..Default::default()
    };

    // The Framework builder will automatically retrieve the bot owner and application ID via the
    // passed token, so that information need not be passed here
    poise::Framework::builder()
        // Configure the client with your Discord bot token in the environment.
        .token(std::env::var("DISCORD_TOKEN").expect("Expected DISCORD_TOKEN in the environment"))
        .options(options)
        .user_data_setup(|_ctx, _data_about_bot, _framework| {
            Box::pin(async move {
                Ok(Data {
                    command_counter: std::sync::Mutex::new(std::collections::HashMap::new()),
                })
            })
        })
        .run()
        .await
        .expect("Client error");

    // INFO: currently not supported by poise
    /*
    // Set a function that's called whenever an attempted command-call's
    // command could not be found.
    .unrecognised_command(unknown_command)
    // Set a function that's called whenever a message is not a command.
    .normal_message(normal_message)
    // Set a function that's called whenever a command's execution didn't complete for one
    // reason or another. For example, when a user has exceeded a rate-limit or a command
    // can only be performed by the bot owner.
    .on_dispatch_error(dispatch_error)
    // Can't be used more than once per 5 seconds:
    .bucket("emoji", |b| b.delay(5)).await
    // Can't be used more than 2 times per 30 seconds, with a 5 second delay applying per channel.
    // Optionally `await_ratelimits` will delay until the command can be executed instead of
    // cancelling the command invocation.
    .bucket("complicated", |b| b.limit(2).time_span(30).delay(5)
        // The target each bucket will apply to.
        .limit_for(LimitedFor::Channel)
        // The maximum amount of command invocations that can be delayed per target.
        // Setting this to 0 (default) will never await/delay commands and cancel the invocation.
        .await_ratelimits(1)
        // A function to call when a rate limit leads to a delay.
        .delay_action(delay_action)
    ).await
    */
}

pub async fn new<F>(
client_builder: ClientBuilder,
user_data_setup: F,
options: FrameworkOptions<U, E>
) -> Result<Arc<Self>, Error>where
F: Send + Sync + 'static + for<'a> FnOnce(&'a Context, &'a Ready, &'a Self) -> BoxFuture<'a, Result<U, E>>,
U: Send + Sync + 'static,
E: Send + 'static,

Setup a new Framework. For more ergonomic setup, please see FrameworkBuilder

This function is async and returns Result because it already initializes the Discord client.

The user data callback is invoked as soon as the bot is logged in. That way, bot data like user ID or connected guilds can be made available to the user data setup function. The user data setup is not allowed to return Result because there would be no reasonable course of action on error.

pub async fn start_with<F: Future<Output = Result<()>>>(
self: Arc<Self>,
start: impl FnOnce(Client) -> F
) -> Result<(), Error>where
U: Send + Sync + 'static,
E: Send + 'static,

Small utility function for starting the framework that is agnostic over client sharding

You can use these shortcut methods to start the framework with a single shard or with automatic sharding: Framework::start, Framework::start_autosharded

Examples

Start shards in a range

let shard_range = [3, 7];
let total_shards = 8;

framework
    .start_with(|mut client| async move {
        client.start_shard_range(shard_range, total_shards).await
   })
   .await?;

pub async fn start(self: Arc<Self>) -> Result<(), Error>where
U: Send + Sync + 'static,
E: Send + 'static,

Starts the framework with a shard. Calls serenity::Client::start internally.

See Framework::start_with for other sharding configurations.

pub async fn start_autosharded(self: Arc<Self>) -> Result<(), Error>where
U: Send + Sync + 'static,
E: Send + 'static,

Starts the framework with automatic sharding. Calls serenity::Client::start_autosharded internally.

See Framework::start_with for other sharding configurations.

pub fn options(&self) -> &FrameworkOptions<U, E>

Return the stored framework options, including commands.

Examples found in repository?

examples/invocation_data/main.rs (line 57)

↕

49
50
51
52
53
54
55
56
57
58
59
60
61
62
63
64
65
66
67
68
69
70
71
72
73
74
75
76
77
78
79
80
81
82
83
84
85
86
87
88
89
90
91
92
93
94
95
96
97
98
99
100
101
102
103
104
105
106
107
108
109
110
111
112
113
114
115
116
117
118
119
120
121
122

async fn main() {
    poise::Framework::builder()
        .token(std::env::var("TOKEN").unwrap())
        .user_data_setup(move |ctx, _, framework| {
            Box::pin(async move {
                poise::serenity_prelude::GuildId(703332075914264606)
                    .set_application_commands(ctx, |b| {
                        *b = poise::samples::create_application_commands(
                            &framework.options().commands,
                        );
                        b
                    })
                    .await
                    .unwrap();
                Ok(())
            })
        })
        .options(poise::FrameworkOptions {
            pre_command: |ctx| {
                Box::pin(async move {
                    println!(
                        "In pre_command: {:?}",
                        ctx.invocation_data::<&str>().await.as_deref()
                    );
                })
            },
            command_check: Some(|ctx| {
                Box::pin(async move {
                    // Global command check is the first callback that's invoked, so let's set the
                    // data here
                    println!("Writing invocation data!");
                    ctx.set_invocation_data("hello").await;

                    println!(
                        "In global check: {:?}",
                        ctx.invocation_data::<&str>().await.as_deref()
                    );

                    Ok(true)
                })
            }),
            post_command: |ctx| {
                Box::pin(async move {
                    println!(
                        "In post_command: {:?}",
                        ctx.invocation_data::<&str>().await.as_deref()
                    );
                })
            },
            on_error: |err| {
                Box::pin(async move {
                    match err {
                        poise::FrameworkError::Command { ctx, .. } => {
                            println!(
                                "In on_error: {:?}",
                                ctx.invocation_data::<&str>().await.as_deref()
                            );
                        }
                        err => poise::samples::on_error(err).await.unwrap(),
                    }
                })
            },

            commands: vec![invocation_data_test()],
            prefix_options: poise::PrefixFrameworkOptions {
                prefix: Some("~".into()),
                ..Default::default()
            },
            ..Default::default()
        })
        .run()
        .await
        .unwrap();
}

pub fn shard_manager(&self) -> &Arc<Mutex<ShardManager>>

Returns the serenity’s client shard manager.

pub fn client(&self) -> impl DerefMut<Target = Client> + '_

Returns the serenity client. Panics if the framework has already started!

pub async fn user_data(&self) -> &U

Retrieves user data, or blocks until it has been initialized (once the Ready event has been received).