Struct poise::FrameworkOptions

pub struct FrameworkOptions<U, E> {
    pub commands: Vec<Command<U, E>>,
    pub on_error: fn(_: FrameworkError<'_, U, E>) -> BoxFuture<'_, ()>,
    pub pre_command: fn(_: Context<'_, U, E>) -> BoxFuture<'_, ()>,
    pub post_command: fn(_: Context<'_, U, E>) -> BoxFuture<'_, ()>,
    pub command_check: Option<fn(_: Context<'_, U, E>) -> BoxFuture<'_, Result<bool, E>>>,
    pub allowed_mentions: Option<CreateAllowedMentions>,
    pub reply_callback: Option<fn(_: Context<'_, U, E>, _: &mut CreateReply<'_>)>,
    pub manual_cooldowns: bool,
    pub require_cache_for_guild_check: bool,
    pub listener: for<'a> fn(_: &'a Context, _: &'a Event<'a>, _: FrameworkContext<'a, U, E>, _: &'a U) -> BoxFuture<'a, Result<(), E>>,
    pub prefix_options: PrefixFrameworkOptions<U, E>,
    pub owners: HashSet<UserId>,
    /* private fields */
}

Framework configuration

Fields
commands: Vec<Command<U, E>>

List of commands in the framework

on_error: fn(_: FrameworkError<'_, U, E>) -> BoxFuture<'_, ()>

Provide a callback to be invoked when any user code yields an error.

pre_command: fn(_: Context<'_, U, E>) -> BoxFuture<'_, ()>

Called before every command

post_command: fn(_: Context<'_, U, E>) -> BoxFuture<'_, ()>

Called after every command if it was successful (returned Ok)

command_check: Option<fn(_: Context<'_, U, E>) -> BoxFuture<'_, Result<bool, E>>>

Provide a callback to be invoked before every command. The command will only be executed if the callback returns true.

If individual commands add their own check, both callbacks are run and must return true.

allowed_mentions: Option<CreateAllowedMentions>

Default set of allowed mentions to use for all responses

By default, user pings are allowed and role pings and everyone pings are filtered

reply_callback: Option<fn(_: Context<'_, U, E>, _: &mut CreateReply<'_>)>

Invoked before every message sent using crate::Context::say or crate::Context::send

Allows you to modify every outgoing message in a central place

manual_cooldowns: bool

If true, disables automatic cooldown handling before every command invocation.

Useful for implementing custom cooldown behavior. See crate::Command::cooldowns and the methods on crate::Cooldowns for how to do that.

require_cache_for_guild_check: bool

If true, changes behavior of guild_only command check to abort execution if the guild is not in cache.

If cache feature is disabled, this has no effect!

listener: for<'a> fn(_: &'a Context, _: &'a Event<'a>, _: FrameworkContext<'a, U, E>, _: &'a U) -> BoxFuture<'a, Result<(), E>>

Called on every Discord event. Can be used to react to non-command events, like messages deletions or guild updates.

prefix_options: PrefixFrameworkOptions<U, E>

Prefix command specific options.

owners: HashSet<UserId>

User IDs which are allowed to use owners_only commands

If using crate::FrameworkBuilder, automatically initialized with the bot application owner and team members

Implementations

impl<U, E> FrameworkOptions<U, E>

pub fn command(
&mut self,
command: Command<U, E>,
meta_builder: impl FnOnce(&mut Command<U, E>) -> &mut Command<U, E> + 'static
)

👎Deprecated:
supply commands in FrameworkOptions directly with commands: vec![...]

Add a new command to the framework