Enum poise::structs::ApplicationCommandOrAutocompleteInteraction

pub enum ApplicationCommandOrAutocompleteInteraction<'a> {
    ApplicationCommand(&'a ApplicationCommandInteraction),
    Autocomplete(&'a AutocompleteInteraction),
}

Abstracts over a refernce to an application command interaction or autocomplete interaction

Used in crate::ApplicationContext. We need to support autocomplete interactions in crate::ApplicationContext because command checks are invoked for autocomplete interactions too: we don’t want poise accidentally leaking sensitive information through autocomplete suggestions

Variants

ApplicationCommand(&'a ApplicationCommandInteraction)

An application command interaction

Autocomplete(&'a AutocompleteInteraction)

An autocomplete interaction

Implementations

impl<'a> ApplicationCommandOrAutocompleteInteraction<'a>

pub fn data(self) -> &'a CommandData

Returns the data field of the underlying interaction

Examples found in repository?

src/structs/context.rs (line 294)

291
292
293
294
295
296

    pub fn invoked_command_name(&self) -> &'a str {
        match self {
            Self::Prefix(ctx) => ctx.invoked_command_name,
            Self::Application(ctx) => &ctx.interaction.data().name,
        }
    }

Hide additional examples

src/slash_argument/slash_trait.rs (line 217)

198
199
200
201
202
203
204
205
206
207
208
209
210
211
212
213
214
215
216
217
218
219
220
221
222
223
224
225

    async fn extract(
        self,
        _: &serenity::Context,
        interaction: crate::ApplicationCommandOrAutocompleteInteraction<'_>,
        value: &Value,
    ) -> Result<serenity::Attachment, SlashArgError> {
        let attachment_id = serenity::AttachmentId(
            value
                .as_str()
                .ok_or(SlashArgError::CommandStructureMismatch(
                    "expected attachment id",
                ))?
                .parse()
                .map_err(|_| {
                    SlashArgError::CommandStructureMismatch("improper attachment id passed")
                })?,
        );

        interaction
            .data()
            .resolved
            .attachments
            .get(&attachment_id)
            .cloned()
            .ok_or(SlashArgError::CommandStructureMismatch(
                "attachment id with no attachment",
            ))
    }

src/dispatch/slash.rs (line 45)

36
37
38
39
40
41
42
43
44
45
46
47
48
49
50
51
52
53
54
55
56
57
58
59
60
61
62
63
64
65
66
67
68
69
70
71
72
73

pub async fn extract_command_and_run_checks<'a, U, E>(
    framework: crate::FrameworkContext<'a, U, E>,
    ctx: &'a serenity::Context,
    interaction: crate::ApplicationCommandOrAutocompleteInteraction<'a>,
    has_sent_initial_response: &'a std::sync::atomic::AtomicBool,
    invocation_data: &'a tokio::sync::Mutex<Box<dyn std::any::Any + Send + Sync>>,
    parent_commands: &'a mut Vec<&'a crate::Command<U, E>>,
) -> Result<crate::ApplicationContext<'a, U, E>, crate::FrameworkError<'a, U, E>> {
    let search_result = find_matching_command(
        &interaction.data().name,
        &interaction.data().options,
        &framework.options.commands,
        parent_commands,
    );
    let (command, leaf_interaction_options) =
        search_result.ok_or(crate::FrameworkError::UnknownInteraction {
            ctx,
            framework,
            interaction,
        })?;

    let ctx = crate::ApplicationContext {
        data: framework.user_data().await,
        serenity_context: ctx,
        framework,
        interaction,
        args: leaf_interaction_options,
        command,
        parent_commands,
        has_sent_initial_response,
        invocation_data,
        __non_exhaustive: (),
    };

    super::common::check_permissions_and_cooldown(ctx.into()).await?;

    Ok(ctx)
}

src/structs/framework_error.rs (line 331)

231
232
233
234
235
236
237
238
239
240
241
242
243
244
245
246
247
248
249
250
251
252
253
254
255
256
257
258
259
260
261
262
263
264
265
266
267
268
269
270
271
272
273
274
275
276
277
278
279
280
281
282
283
284
285
286
287
288
289
290
291
292
293
294
295
296
297
298
299
300
301
302
303
304
305
306
307
308
309
310
311
312
313
314
315
316
317
318
319
320
321
322
323
324
325
326
327
328
329
330
331
332
333
334
335

    fn fmt(&self, f: &mut std::fmt::Formatter<'_>) -> std::fmt::Result {
        match self {
            Self::Setup {
                error: _,
                framework: _,
                data_about_bot: _,
                ctx: _,
            } => write!(f, "poise setup error"),
            Self::EventHandler {
                error: _,
                ctx: _,
                event,
                framework: _,
            } => write!(f, "error in {} event event handler", event.name()),
            Self::Command { error: _, ctx } => {
                write!(f, "error in command `{}`", full_command_name!(ctx))
            }
            Self::ArgumentParse {
                error: _,
                input,
                ctx,
            } => write!(
                f,
                "failed to parse argument in command `{}` on input {:?}",
                full_command_name!(ctx),
                input
            ),
            Self::CommandStructureMismatch { description, ctx } => write!(
                f,
                "unexpected application command structure in command `{}`: {}",
                full_command_name!(crate::Context::Application(*ctx)),
                description
            ),
            Self::CooldownHit {
                remaining_cooldown,
                ctx,
            } => write!(
                f,
                "cooldown hit in command `{}` ({:?} remaining)",
                full_command_name!(ctx),
                remaining_cooldown
            ),
            Self::MissingBotPermissions {
                missing_permissions,
                ctx,
            } => write!(
                f,
                "bot is missing permisions ({}) to execute command `{}`",
                missing_permissions,
                full_command_name!(ctx),
            ),
            Self::MissingUserPermissions {
                missing_permissions,
                ctx,
            } => write!(
                f,
                "user is or may be missing permisions ({:?}) to execute command `{}`",
                missing_permissions,
                full_command_name!(ctx),
            ),
            Self::NotAnOwner { ctx } => write!(
                f,
                "owner-only command `{}` cannot be run by non-owners",
                full_command_name!(ctx)
            ),
            Self::GuildOnly { ctx } => write!(
                f,
                "guild-only command `{}` cannot run in DMs",
                full_command_name!(ctx)
            ),
            Self::DmOnly { ctx } => write!(
                f,
                "DM-only command `{}` cannot run in guilds",
                full_command_name!(ctx)
            ),
            Self::NsfwOnly { ctx } => write!(
                f,
                "nsfw-only command `{}` cannot run in non-nsfw channels",
                full_command_name!(ctx)
            ),
            Self::CommandCheckFailed { error: _, ctx } => write!(
                f,
                "pre-command check for command `{}` either denied access or errored",
                full_command_name!(ctx)
            ),
            Self::DynamicPrefix {
                error: _,
                ctx: _,
                msg,
            } => {
                write!(
                    f,
                    "dynamic prefix callback errored on message {:?}",
                    msg.content
                )
            }
            Self::UnknownCommand { msg_content, .. } => {
                write!(f, "unknown command `{}`", msg_content)
            }
            Self::UnknownInteraction { interaction, .. } => {
                write!(f, "unknown interaction `{}`", interaction.data().name)
            }
            Self::__NonExhaustive => unreachable!(),
        }
    }

src/builtins/mod.rs (line 154)

29
30
31
32
33
34
35
36
37
38
39
40
41
42
43
44
45
46
47
48
49
50
51
52
53
54
55
56
57
58
59
60
61
62
63
64
65
66
67
68
69
70
71
72
73
74
75
76
77
78
79
80
81
82
83
84
85
86
87
88
89
90
91
92
93
94
95
96
97
98
99
100
101
102
103
104
105
106
107
108
109
110
111
112
113
114
115
116
117
118
119
120
121
122
123
124
125
126
127
128
129
130
131
132
133
134
135
136
137
138
139
140
141
142
143
144
145
146
147
148
149
150
151
152
153
154
155
156
157
158
159
160
161

pub async fn on_error<U, E: std::fmt::Display + std::fmt::Debug>(
    error: crate::FrameworkError<'_, U, E>,
) -> Result<(), serenity::Error> {
    match error {
        crate::FrameworkError::Setup { error, .. } => {
            log::error!("Error in user data setup: {}", error);
        }
        crate::FrameworkError::EventHandler { error, event, .. } => log::error!(
            "User event event handler encountered an error on {} event: {}",
            event.name(),
            error
        ),
        crate::FrameworkError::Command { ctx, error } => {
            let error = error.to_string();
            ctx.say(error).await?;
        }
        crate::FrameworkError::ArgumentParse { ctx, input, error } => {
            // If we caught an argument parse error, give a helpful error message with the
            // command explanation if available
            let usage = match ctx.command().help_text {
                Some(help_text) => help_text(),
                None => "Please check the help menu for usage information".into(),
            };
            let response = if let Some(input) = input {
                format!(
                    "**Cannot parse `{}` as argument: {}**\n{}",
                    input, error, usage
                )
            } else {
                format!("**{}**\n{}", error, usage)
            };
            ctx.say(response).await?;
        }
        crate::FrameworkError::CommandStructureMismatch { ctx, description } => {
            log::error!(
                "Error: failed to deserialize interaction arguments for `/{}`: {}",
                ctx.command.name,
                description,
            );
        }
        crate::FrameworkError::CommandCheckFailed { ctx, error } => {
            log::error!(
                "A command check failed in command {} for user {}: {:?}",
                ctx.command().name,
                ctx.author().name,
                error,
            );
        }
        crate::FrameworkError::CooldownHit {
            remaining_cooldown,
            ctx,
        } => {
            let msg = format!(
                "You're too fast. Please wait {} seconds before retrying",
                remaining_cooldown.as_secs()
            );
            ctx.send(|b| b.content(msg).ephemeral(true)).await?;
        }
        crate::FrameworkError::MissingBotPermissions {
            missing_permissions,
            ctx,
        } => {
            let msg = format!(
                "Command cannot be executed because the bot is lacking permissions: {}",
                missing_permissions,
            );
            ctx.send(|b| b.content(msg).ephemeral(true)).await?;
        }
        crate::FrameworkError::MissingUserPermissions {
            missing_permissions,
            ctx,
        } => {
            let response = if let Some(missing_permissions) = missing_permissions {
                format!(
                    "You're lacking permissions for `{}{}`: {}",
                    ctx.prefix(),
                    ctx.command().name,
                    missing_permissions,
                )
            } else {
                format!(
                    "You may be lacking permissions for `{}{}`. Not executing for safety",
                    ctx.prefix(),
                    ctx.command().name,
                )
            };
            ctx.send(|b| b.content(response).ephemeral(true)).await?;
        }
        crate::FrameworkError::NotAnOwner { ctx } => {
            let response = "Only bot owners can call this command";
            ctx.send(|b| b.content(response).ephemeral(true)).await?;
        }
        crate::FrameworkError::GuildOnly { ctx } => {
            let response = "You cannot run this command in DMs.";
            ctx.send(|b| b.content(response).ephemeral(true)).await?;
        }
        crate::FrameworkError::DmOnly { ctx } => {
            let response = "You cannot run this command outside DMs.";
            ctx.send(|b| b.content(response).ephemeral(true)).await?;
        }
        crate::FrameworkError::NsfwOnly { ctx } => {
            let response = "You cannot run this command outside NSFW channels.";
            ctx.send(|b| b.content(response).ephemeral(true)).await?;
        }
        crate::FrameworkError::DynamicPrefix { error, msg, .. } => {
            log::error!(
                "Dynamic prefix failed for message {:?}: {}",
                msg.content,
                error
            );
        }
        crate::FrameworkError::UnknownCommand {
            msg_content,
            prefix,
            ..
        } => {
            log::warn!(
                "Recognized prefix `{}`, but didn't recognize command name in `{}`",
                prefix,
                msg_content,
            );
        }
        crate::FrameworkError::UnknownInteraction { interaction, .. } => {
            log::warn!(
                "received unknown interaction \"{}\"",
                interaction.data().name
            );
        }
        crate::FrameworkError::__NonExhaustive => panic!(),
    }

    Ok(())
}

pub fn id(self) -> InteractionId

Returns the ID of the underlying interaction

Examples found in repository?

src/structs/context.rs (line 216)

214
215
216
217
218
219
220
221
222
223
224
225
226
227
228
229
230
231
232
233
234
235
236
237
238
239
240
241
242
243
244
245
246
247
248
249
250
251
252
253
254
255
256
257

    pub fn created_at(&self) -> serenity::Timestamp {
        match self {
            Self::Application(ctx) => ctx.interaction.id().created_at(),
            Self::Prefix(ctx) => ctx.msg.timestamp,
        }
    }

    /// Get the author of the command message or application command.
    pub fn author(&self) -> &'a serenity::User {
        match self {
            Self::Application(ctx) => ctx.interaction.user(),
            Self::Prefix(ctx) => &ctx.msg.author,
        }
    }

    /// Return a ID that uniquely identifies this command invocation.
    #[cfg(any(feature = "chrono", feature = "time"))]
    pub fn id(&self) -> u64 {
        match self {
            Self::Application(ctx) => ctx.interaction.id().0,
            Self::Prefix(ctx) => {
                let mut id = ctx.msg.id.0;
                if let Some(edited_timestamp) = ctx.msg.edited_timestamp {
                    // We replace the 42 datetime bits with msg.timestamp_edited so that the ID is
                    // unique even after edits

                    // Set existing datetime bits to zero
                    id &= !0 >> 42;

                    // Calculate Discord's datetime representation (millis since Discord epoch) and
                    // insert those bits into the ID

                    #[cfg(feature = "time")]
                    let timestamp_millis = edited_timestamp.unix_timestamp_nanos() / 1_000_000;

                    #[cfg(not(feature = "time"))]
                    let timestamp_millis = edited_timestamp.timestamp_millis();

                    id |= ((timestamp_millis - 1420070400000) as u64) << 22;
                }
                id
            }
        }
    }

pub fn guild_id(self) -> Option<GuildId>

Returns the guild ID of the underlying interaction

Examples found in repository?

src/structs/context.rs (line 161)

159
160
161
162
163
164

    pub fn guild_id(&self) -> Option<serenity::GuildId> {
        match self {
            Self::Application(ctx) => ctx.interaction.guild_id(),
            Self::Prefix(ctx) => ctx.msg.guild_id,
        }
    }

Hide additional examples

src/slash_argument/slash_trait.rs (line 110)

99
100
101
102
103
104
105
106
107
108
109
110
111
112
113
114
115
116
117
118
119

    async fn extract(
        self,
        ctx: &serenity::Context,
        interaction: crate::ApplicationCommandOrAutocompleteInteraction<'_>,
        value: &serenity::json::Value,
    ) -> Result<T, SlashArgError> {
        let string = value
            .as_str()
            .ok_or(SlashArgError::CommandStructureMismatch("expected string"))?;
        T::convert(
            ctx,
            interaction.guild_id(),
            Some(interaction.channel_id()),
            string,
        )
        .await
        .map_err(|e| SlashArgError::Parse {
            error: e.into(),
            input: string.into(),
        })
    }

pub fn channel_id(self) -> ChannelId

Returns the channel ID of the underlying interaction

Examples found in repository?

src/structs/context.rs (line 153)

151
152
153
154
155
156

    pub fn channel_id(&self) -> serenity::ChannelId {
        match self {
            Self::Application(ctx) => ctx.interaction.channel_id(),
            Self::Prefix(ctx) => ctx.msg.channel_id,
        }
    }

Hide additional examples

src/slash_argument/slash_trait.rs (line 111)

99
100
101
102
103
104
105
106
107
108
109
110
111
112
113
114
115
116
117
118
119

    async fn extract(
        self,
        ctx: &serenity::Context,
        interaction: crate::ApplicationCommandOrAutocompleteInteraction<'_>,
        value: &serenity::json::Value,
    ) -> Result<T, SlashArgError> {
        let string = value
            .as_str()
            .ok_or(SlashArgError::CommandStructureMismatch("expected string"))?;
        T::convert(
            ctx,
            interaction.guild_id(),
            Some(interaction.channel_id()),
            string,
        )
        .await
        .map_err(|e| SlashArgError::Parse {
            error: e.into(),
            input: string.into(),
        })
    }

pub fn member(self) -> Option<&'a Member>

Returns the member field of the underlying interaction

Examples found in repository?

src/structs/context.rs (line 203)

201
202
203
204
205
206
207
208
209
210
211

    pub async fn author_member(&'a self) -> Option<Cow<'a, serenity::Member>> {
        if let Self::Application(ctx) = self {
            ctx.interaction.member().map(Cow::Borrowed)
        } else {
            self.guild_id()?
                .member(self.sc(), self.author().id)
                .await
                .ok()
                .map(Cow::Owned)
        }
    }

pub fn user(self) -> &'a User

Returns the user field of the underlying interaction

Examples found in repository?

src/structs/context.rs (line 224)

222
223
224
225
226
227

    pub fn author(&self) -> &'a serenity::User {
        match self {
            Self::Application(ctx) => ctx.interaction.user(),
            Self::Prefix(ctx) => &ctx.msg.author,
        }
    }

pub fn unwrap(self) -> &'a ApplicationCommandInteraction

Returns the inner serenity::ApplicationCommandInteraction and panics otherwise

Examples found in repository?

src/modal.rs (line 58)

53
54
55
56
57
58
59
60
61
62
63
64
65
66
67
68
69
70
71
72
73
74
75
76
77
78
79
80
81
82
83
84
85
86
87
88
89
90
91

pub async fn execute_modal<U: Send + Sync, E, M: Modal>(
    ctx: crate::ApplicationContext<'_, U, E>,
    defaults: Option<M>,
    timeout: Option<std::time::Duration>,
) -> Result<Option<M>, serenity::Error> {
    let interaction = ctx.interaction.unwrap();
    let interaction_id = interaction.id.to_string();

    // Send modal
    interaction
        .create_interaction_response(ctx.serenity_context, |b| {
            *b = M::create(defaults, interaction_id.clone());
            b
        })
        .await?;
    ctx.has_sent_initial_response
        .store(true, std::sync::atomic::Ordering::SeqCst);

    // Wait for user to submit
    let response = serenity::CollectModalInteraction::new(&ctx.serenity_context.shard)
        .filter(move |d| d.data.custom_id == interaction_id)
        .timeout(timeout.unwrap_or(std::time::Duration::from_secs(3600)))
        .await;
    let response = match response {
        Some(x) => x,
        None => return Ok(None),
    };

    // Send acknowledgement so that the pop-up is closed
    response
        .create_interaction_response(ctx.serenity_context, |b| {
            b.kind(serenity::InteractionResponseType::DeferredUpdateMessage)
        })
        .await?;

    Ok(Some(
        M::parse(response.data.clone()).map_err(serenity::Error::Other)?,
    ))
}

pub fn locale(self) -> &'a str

Returns the locale field of the underlying interaction

Examples found in repository?

src/structs/context.rs (line 445)

443
444
445
446
447
448

    pub fn locale(&self) -> Option<&str> {
        match self {
            Context::Application(ctx) => Some(ctx.interaction.locale()),
            Context::Prefix(_) => None,
        }
    }