poise::structs

Enum Context

pub enum Context<'a, U, E> {
    Application(ApplicationContext<'a, U, E>),
    Prefix(PrefixContext<'a, U, E>),
}

Wrapper around either crate::ApplicationContext or crate::PrefixContext

Variants

Application(ApplicationContext<'a, U, E>)

Application command context

Prefix(PrefixContext<'a, U, E>)

Prefix command context

Implementations

impl<'a, U, E> Context<'a, U, E>

pub async fn defer(self) -> Result<(), Error>

Defer the response, giving the bot multiple minutes to respond without the user seeing an “interaction failed error”.

Also sets the crate::ApplicationContext::has_sent_initial_response flag so subsequent responses will be sent in the correct manner.

No-op if this is an autocomplete context

This will make the response public; to make it ephemeral, use Self::defer_ephemeral().

pub async fn defer_ephemeral(self) -> Result<(), Error>

See Self::defer()

This will make the response ephemeral; to make it public, use Self::defer().

pub async fn defer_or_broadcast(self) -> Result<Option<Typing>, Error>

If this is an application command, Self::defer() is called

If this is a prefix command, a typing broadcast is started until the return value is dropped.

pub async fn say( self, text: impl Into<String>, ) -> Result<ReplyHandle<'a>, Error>

Shorthand of crate::say_reply

Note: panics when called in an autocomplete context!

Examples found in repository?

examples/help_generation/main.rs (line 40)

39async fn beer(ctx: Context<'_>) -> Result<(), Error> {
40    ctx.say("🍺").await?;
41    Ok(())
42}
43
44/// Respond with a random fruit
45///
46/// Subcommands can be used to get a specific fruit
47#[poise::command(
48    slash_command,
49    prefix_command,
50    subcommands(
51        "apple",
52        "banana",
53        "orange",
54        "watermelon",
55        "grape",
56        "strawberry",
57        "help"
58    ),
59    category = "Vegan"
60)]
61async fn fruit(ctx: Context<'_>) -> Result<(), Error> {
62    let response = FRUIT[rand::thread_rng().gen_range(0..FRUIT.len())];
63    ctx.say(response).await?;
64    Ok(())
65}
66
67/// Respond with an apple
68#[poise::command(slash_command, prefix_command, subcommands("red", "green"))]
69async fn apple(ctx: Context<'_>) -> Result<(), Error> {
70    ctx.say("🍎").await?;
71    Ok(())
72}
73
74/// Respond with a red apple
75#[poise::command(slash_command, prefix_command)]
76async fn red(ctx: Context<'_>) -> Result<(), Error> {
77    ctx.say("🍎").await?;
78    Ok(())
79}
80
81/// Respond with a green apple
82#[poise::command(slash_command, prefix_command)]
83async fn green(ctx: Context<'_>) -> Result<(), Error> {
84    ctx.say("🍏").await?;
85    Ok(())
86}
87
88/// Respond with a banana
89#[poise::command(slash_command, prefix_command)]
90async fn banana(ctx: Context<'_>) -> Result<(), Error> {
91    ctx.say("🍌").await?;
92    Ok(())
93}
94
95/// Respond with an orange
96#[poise::command(slash_command, prefix_command)]
97async fn orange(ctx: Context<'_>) -> Result<(), Error> {
98    ctx.say("🍊").await?;
99    Ok(())
100}
101
102/// Respond with a watermelon
103#[poise::command(slash_command, prefix_command)]
104async fn watermelon(ctx: Context<'_>) -> Result<(), Error> {
105    ctx.say("🍉").await?;
106    Ok(())
107}
108
109/// Respond with a grape
110#[poise::command(slash_command, prefix_command)]
111async fn grape(ctx: Context<'_>) -> Result<(), Error> {
112    ctx.say("🍇").await?;
113    Ok(())
114}
115
116/// Respond with a strawberry
117#[poise::command(slash_command, prefix_command)]
118async fn strawberry(ctx: Context<'_>) -> Result<(), Error> {
119    ctx.say("🍓").await?;
120    Ok(())
121}
122
123/// Respond with a random vegetable
124#[poise::command(slash_command, prefix_command, category = "Vegan")]
125async fn vegetable(ctx: Context<'_>) -> Result<(), Error> {
126    let response = VEGETABLES[rand::thread_rng().gen_range(0..VEGETABLES.len())];
127    ctx.say(response).await?;
128    Ok(())
129}
130
131/// Respond with a random meat
132#[poise::command(slash_command, prefix_command, category = "Other")]
133async fn meat(ctx: Context<'_>) -> Result<(), Error> {
134    let response = MEAT[rand::thread_rng().gen_range(0..MEAT.len())];
135    ctx.say(response).await?;
136    Ok(())
137}
138
139/// Respond with a random dairy product
140#[poise::command(slash_command, prefix_command, category = "Other")]
141async fn dairy(ctx: Context<'_>) -> Result<(), Error> {
142    let response = DAIRY[rand::thread_rng().gen_range(0..DAIRY.len())];
143    ctx.say(response).await?;
144    Ok(())
145}
146
147/// Give a user some random food
148#[poise::command(context_menu_command = "Give food")]
149async fn context_food(
150    ctx: Context<'_>,
151    #[description = "User to give food to"] user: serenity::User,
152) -> Result<(), Error> {
153    let response = format!(
154        "<@{}>: {}",
155        user.id,
156        FOOD[rand::thread_rng().gen_range(0..FOOD.len())]
157    );
158
159    ctx.say(response).await?;
160    Ok(())
161}
162
163/// Give a user some random fruit
164#[poise::command(
165    slash_command,
166    context_menu_command = "Give fruit",
167    category = "Context menu but also slash/prefix"
168)]
169async fn context_fruit(
170    ctx: Context<'_>,
171    #[description = "User to give fruit to"] user: serenity::User,
172) -> Result<(), Error> {
173    let response = format!(
174        "<@{}>: {}",
175        user.id,
176        FRUIT[rand::thread_rng().gen_range(0..FRUIT.len())]
177    );
178
179    ctx.say(response).await?;
180    Ok(())
181}
182
183/// Give a user some random vegetable
184#[poise::command(
185    prefix_command,
186    context_menu_command = "Give vegetable",
187    category = "Context menu but also slash/prefix"
188)]
189async fn context_vegetable(
190    ctx: Context<'_>,
191    #[description = "User to give vegetable to"] user: serenity::User,
192) -> Result<(), Error> {
193    let response = format!(
194        "<@{}>: {}",
195        user.id,
196        VEGETABLES[rand::thread_rng().gen_range(0..VEGETABLES.len())]
197    );
198
199    ctx.say(response).await?;
200    Ok(())
201}
202
203/// Give a user some random meat
204#[poise::command(
205    prefix_command,
206    slash_command,
207    context_menu_command = "Give meat",
208    category = "Context menu but also slash/prefix"
209)]
210async fn context_meat(
211    ctx: Context<'_>,
212    #[description = "User to give meat to"] user: serenity::User,
213) -> Result<(), Error> {
214    let response = format!(
215        "<@{}>: {}",
216        user.id,
217        MEAT[rand::thread_rng().gen_range(0..MEAT.len())]
218    );
219
220    ctx.say(response).await?;
221    Ok(())
222}
223
224/// React to a message with random food
225// This command intentionally doesn't have a slash/prefix command, and its own
226// category, so that we can test whether the category shows up in the help
227// message. It shouldn't.
228#[poise::command(
229    context_menu_command = "React with food",
230    ephemeral,
231    category = "No slash/prefix",
232    subcommands("fruit_react", "vegetable_react")
233)]
234async fn food_react(
235    ctx: Context<'_>,
236    #[description = "Message to react to (enter a link or ID)"] msg: serenity::Message,
237) -> Result<(), Error> {
238    let reaction = FOOD[rand::thread_rng().gen_range(0..FOOD.len())].to_string();
239    msg.react(ctx, serenity::ReactionType::Unicode(reaction))
240        .await?;
241    ctx.say("Reacted!").await?;
242    Ok(())
243}
244
245// These next two commands are subcommands of `food_react`, so they're not
246// visible in the overview help command. But they should still show up in
247// `?help react with food`
248
249/// React to a message with a random fruit
250#[poise::command(
251    slash_command,
252    context_menu_command = "React with fruit",
253    ephemeral,
254    category = "No slash/prefix"
255)]
256async fn fruit_react(
257    ctx: Context<'_>,
258    #[description = "Message to react to (enter a link or ID)"] msg: serenity::Message,
259) -> Result<(), Error> {
260    let reaction = FRUIT[rand::thread_rng().gen_range(0..FRUIT.len())].to_string();
261    msg.react(ctx, serenity::ReactionType::Unicode(reaction))
262        .await?;
263    ctx.say("Reacted!").await?;
264    Ok(())
265}
266
267/// React to a message with a random vegetable
268#[poise::command(
269    slash_command,
270    context_menu_command = "React with vegetable",
271    ephemeral,
272    category = "No slash/prefix"
273)]
274async fn vegetable_react(
275    ctx: Context<'_>,
276    #[description = "Message to react to (enter a link or ID)"] msg: serenity::Message,
277) -> Result<(), Error> {
278    let reaction = VEGETABLES[rand::thread_rng().gen_range(0..VEGETABLES.len())].to_string();
279    msg.react(ctx, serenity::ReactionType::Unicode(reaction))
280        .await?;
281    ctx.say("Reacted!").await?;
282    Ok(())
283}

Hide additional examples

examples/feature_showcase/subcommands.rs (line 9)

8pub async fn parent(ctx: Context<'_>) -> Result<(), Error> {
9    ctx.say("Hello there!").await?;
10    Ok(())
11}
12
13/// A subcommand of `parent`
14#[poise::command(prefix_command, slash_command)]
15pub async fn child1(ctx: Context<'_>) -> Result<(), Error> {
16    ctx.say("You invoked the first child command!").await?;
17    Ok(())
18}
19
20/// Another subcommand of `parent`
21#[poise::command(prefix_command, slash_command)]
22pub async fn child2(ctx: Context<'_>) -> Result<(), Error> {
23    ctx.say("You invoked the second child command!").await?;
24    Ok(())
25}

examples/manual_dispatch/main.rs (line 14)

13async fn ping(ctx: poise::Context<'_, (), Error>) -> Result<(), Error> {
14    ctx.say("Pong!").await?;
15    Ok(())
16}

examples/feature_showcase/panic_handler.rs (line 8)

7pub async fn div(ctx: Context<'_>, a: i32, b: i32) -> Result<(), Error> {
8    ctx.say((a / b).to_string()).await?;
9    Ok(())
10}

examples/feature_showcase/subcommand_required.rs (line 25)

24pub async fn child1(ctx: Context<'_>) -> Result<(), Error> {
25    ctx.say("You invoked the first child command!").await?;
26    Ok(())
27}
28
29/// Another subcommand of `parent`
30#[poise::command(prefix_command, slash_command)]
31pub async fn child2(ctx: Context<'_>) -> Result<(), Error> {
32    ctx.say("You invoked the second child command!").await?;
33    Ok(())
34}

examples/feature_showcase/user_apps.rs (line 14)

13pub async fn everywhere(ctx: Context<'_>) -> Result<(), Error> {
14    ctx.say("This command is available everywhere!").await?;
15    Ok(())
16}
17
18// also works with `context_menu_command`
19/// Available everywhere
20#[poise::command(
21    context_menu_command = "Everywhere",
22    install_context = "Guild|User",
23    interaction_context = "Guild|BotDm|PrivateChannel"
24)]
25pub async fn everywhere_context(ctx: Context<'_>, msg: serenity::Message) -> Result<(), Error> {
26    msg.reply(ctx, "This context menu is available everywhere!")
27        .await?;
28    Ok(())
29}
30
31/// Available with a user install only
32#[poise::command(
33    slash_command,
34    install_context = "User",
35    interaction_context = "Guild|BotDm|PrivateChannel"
36)]
37pub async fn user_install(ctx: Context<'_>) -> Result<(), Error> {
38    ctx.say("This command is available only with a user install!")
39        .await?;
40    Ok(())
41}
42
43/// Not available in guilds
44#[poise::command(
45    slash_command,
46    install_context = "User",
47    interaction_context = "BotDm|PrivateChannel"
48)]
49pub async fn not_in_guilds(ctx: Context<'_>) -> Result<(), Error> {
50    ctx.say("This command is not available in guilds!").await?;
51    Ok(())
52}
53
54/// User install only in guilds
55#[poise::command(slash_command, install_context = "User", interaction_context = "Guild")]
56pub async fn user_install_guild(ctx: Context<'_>) -> Result<(), Error> {
57    ctx.say("This command is available in guilds only with a user install!")
58        .await?;
59    Ok(())
60}

Additional examples can be found in:

examples/feature_showcase/inherit_checks.rs

examples/feature_showcase/checks.rs

examples/feature_showcase/raw_identifiers.rs

examples/feature_showcase/choice_parameter.rs

examples/feature_showcase/code_block_parameter.rs

examples/feature_showcase/track_edits.rs

examples/fluent_localization/main.rs

examples/generic_commands/main.rs

examples/quickstart/main.rs

examples/feature_showcase/context_menu.rs

examples/feature_showcase/bool_parameter.rs

examples/feature_showcase/parameter_attributes.rs

examples/feature_showcase/attachment_parameter.rs

examples/feature_showcase/autocomplete.rs

examples/basic_structure/commands.rs

examples/feature_showcase/localization.rs

examples/advanced_cooldowns/main.rs

pub async fn reply( self, text: impl Into<String>, ) -> Result<ReplyHandle<'a>, Error>

Like Self::say, but formats the message as a reply to the user’s command message.

Equivalent to .send(CreateReply::default().content("...").reply(true)).

Only has an effect in prefix context, because slash command responses are always formatted as a reply.

Note: panics when called in an autocomplete context!

Examples found in repository?

examples/feature_showcase/response_with_reply.rs (line 5)

4pub async fn reply(ctx: Context<'_>) -> Result<(), Error> {
5    ctx.reply(format!("Hello {}!", ctx.author().name)).await?;
6    Ok(())
7}

pub async fn send(self, builder: CreateReply) -> Result<ReplyHandle<'a>, Error>

Shorthand of crate::send_reply

Note: panics when called in an autocomplete context!

Examples found in repository?

examples/feature_showcase/modal.rs (line 38)

25pub async fn component_modal(ctx: crate::Context<'_>) -> Result<(), Error> {
26    let reply = {
27        let components = vec![serenity::CreateActionRow::Buttons(vec![
28            serenity::CreateButton::new("open_modal")
29                .label("Open modal")
30                .style(poise::serenity_prelude::ButtonStyle::Success),
31        ])];
32
33        poise::CreateReply::default()
34            .content("Click the button below to open the modal")
35            .components(components)
36    };
37
38    ctx.send(reply).await?;
39
40    while let Some(mci) = serenity::ComponentInteractionCollector::new(ctx.serenity_context())
41        .timeout(std::time::Duration::from_secs(120))
42        .filter(move |mci| mci.data.custom_id == "open_modal")
43        .await
44    {
45        let data =
46            poise::execute_modal_on_component_interaction::<MyModal>(ctx, mci, None, None).await?;
47        println!("Got data: {:?}", data);
48    }
49    Ok(())
50}

Hide additional examples

examples/feature_showcase/collector.rs (line 21)

6pub async fn boop(ctx: Context<'_>) -> Result<(), Error> {
7    let uuid_boop = ctx.id();
8
9    let reply = {
10        let components = vec![serenity::CreateActionRow::Buttons(vec![
11            serenity::CreateButton::new(format!("{uuid_boop}"))
12                .style(serenity::ButtonStyle::Primary)
13                .label("Boop me!"),
14        ])];
15
16        CreateReply::default()
17            .content("I want some boops!")
18            .components(components)
19    };
20
21    ctx.send(reply).await?;
22
23    let mut boop_count = 0;
24    while let Some(mci) = serenity::ComponentInteractionCollector::new(ctx)
25        .author_id(ctx.author().id)
26        .channel_id(ctx.channel_id())
27        .timeout(std::time::Duration::from_secs(120))
28        .filter(move |mci| mci.data.custom_id == uuid_boop.to_string())
29        .await
30    {
31        boop_count += 1;
32
33        let mut msg = mci.message.clone();
34        msg.edit(
35            ctx,
36            serenity::EditMessage::new().content(format!("Boop count: {boop_count}")),
37        )
38        .await?;
39
40        mci.create_response(ctx, serenity::CreateInteractionResponse::Acknowledge)
41            .await?;
42    }
43
44    Ok(())
45}

examples/feature_showcase/track_edits.rs (line 25)

5pub async fn test_reuse_response(ctx: Context<'_>) -> Result<(), Error> {
6    let image_url = "https://raw.githubusercontent.com/serenity-rs/serenity/current/logo.png";
7
8    let reply = {
9        let embed = serenity::CreateEmbed::default()
10            .description("embed 1")
11            .image(image_url);
12
13        let components = vec![serenity::CreateActionRow::Buttons(vec![
14            serenity::CreateButton::new("1")
15                .label("button 1")
16                .style(serenity::ButtonStyle::Primary),
17        ])];
18
19        poise::CreateReply::default()
20            .content("message 1")
21            .embed(embed)
22            .components(components)
23    };
24
25    ctx.send(reply).await?;
26
27    tokio::time::sleep(std::time::Duration::from_secs(2)).await;
28
29    let image_url = "https://raw.githubusercontent.com/serenity-rs/serenity/current/examples/e09_create_message_builder/ferris_eyes.png";
30    let reply = {
31        let embed = serenity::CreateEmbed::default()
32            .description("embed 2")
33            .image(image_url);
34
35        let components = vec![serenity::CreateActionRow::Buttons(vec![
36            serenity::CreateButton::new("2")
37                .label("button 2")
38                .style(serenity::ButtonStyle::Danger),
39        ])];
40
41        poise::CreateReply::default()
42            .content("message 2")
43            .embed(embed)
44            .components(components)
45    };
46
47    ctx.send(reply).await?;
48    Ok(())
49}

pub fn serenity_context(self) -> &'a Context

Return the stored serenity::Context within the underlying context type.

Examples found in repository?

examples/feature_showcase/modal.rs (line 40)

25pub async fn component_modal(ctx: crate::Context<'_>) -> Result<(), Error> {
26    let reply = {
27        let components = vec![serenity::CreateActionRow::Buttons(vec![
28            serenity::CreateButton::new("open_modal")
29                .label("Open modal")
30                .style(poise::serenity_prelude::ButtonStyle::Success),
31        ])];
32
33        poise::CreateReply::default()
34            .content("Click the button below to open the modal")
35            .components(components)
36    };
37
38    ctx.send(reply).await?;
39
40    while let Some(mci) = serenity::ComponentInteractionCollector::new(ctx.serenity_context())
41        .timeout(std::time::Duration::from_secs(120))
42        .filter(move |mci| mci.data.custom_id == "open_modal")
43        .await
44    {
45        let data =
46            poise::execute_modal_on_component_interaction::<MyModal>(ctx, mci, None, None).await?;
47        println!("Got data: {:?}", data);
48    }
49    Ok(())
50}

pub fn cooldown_context(self) -> CooldownContext

Create a crate::CooldownContext based off the underlying context type.

Examples found in repository?

examples/advanced_cooldowns/main.rs (line 18)

8async fn dynamic_cooldowns(ctx: Context<'_>) -> Result<(), Error> {
9    {
10        let mut cooldown_tracker = ctx.command().cooldowns.lock().unwrap();
11
12        // You can change the cooldown duration depending on the message author, for example
13        let mut cooldown_durations = poise::CooldownConfig::default();
14        if ctx.author().id == 472029906943868929 {
15            cooldown_durations.user = Some(std::time::Duration::from_secs(10));
16        }
17
18        match cooldown_tracker.remaining_cooldown(ctx.cooldown_context(), &cooldown_durations) {
19            Some(remaining) => {
20                return Err(format!("Please wait {} seconds", remaining.as_secs()).into())
21            }
22            None => cooldown_tracker.start_cooldown(ctx.cooldown_context()),
23        }
24    };
25
26    ctx.say("You successfully invoked the command!").await?;
27    Ok(())
28}

pub fn discord(self) -> &'a Context

👎Deprecated:
poise::Context can now be passed directly into most serenity functions. Otherwise, use .serenity_context() now

See Self::serenity_context.

pub fn framework(self) -> FrameworkContext<'a, U, E>

Returns a view into data stored by the framework, like configuration

Examples found in repository?

examples/feature_showcase/checks.rs (line 15)

14pub async fn shutdown(ctx: Context<'_>) -> Result<(), Error> {
15    ctx.framework().shard_manager().shutdown_all().await;
16    Ok(())
17}

pub fn data(self) -> &'a U

Return a reference to your custom user data

Examples found in repository?

examples/basic_structure/commands.rs (line 33)

27pub async fn vote(
28    ctx: Context<'_>,
29    #[description = "What to vote for"] choice: String,
30) -> Result<(), Error> {
31    // Lock the Mutex in a block {} so the Mutex isn't locked across an await point
32    let num_votes = {
33        let mut hash_map = ctx.data().votes.lock().unwrap();
34        let num_votes = hash_map.entry(choice.clone()).or_default();
35        *num_votes += 1;
36        *num_votes
37    };
38
39    let response = format!("Successfully voted for {choice}. {choice} now has {num_votes} votes!");
40    ctx.say(response).await?;
41    Ok(())
42}
43
44/// Retrieve number of votes
45///
46/// Retrieve the number of votes either in general, or for a specific choice:
47/// ```
48/// ~getvotes
49/// ~getvotes pumpkin
50/// ```
51#[poise::command(prefix_command, track_edits, aliases("votes"), slash_command)]
52pub async fn getvotes(
53    ctx: Context<'_>,
54    #[description = "Choice to retrieve votes for"] choice: Option<String>,
55) -> Result<(), Error> {
56    if let Some(choice) = choice {
57        let num_votes = *ctx.data().votes.lock().unwrap().get(&choice).unwrap_or(&0);
58        let response = match num_votes {
59            0 => format!("Nobody has voted for {} yet", choice),
60            _ => format!("{} people have voted for {}", num_votes, choice),
61        };
62        ctx.say(response).await?;
63    } else {
64        let mut response = String::new();
65        for (choice, num_votes) in ctx.data().votes.lock().unwrap().iter() {
66            response += &format!("{}: {} votes", choice, num_votes);
67        }
68
69        if response.is_empty() {
70            response += "Nobody has voted for anything yet :(";
71        }
72
73        ctx.say(response).await?;
74    };
75
76    Ok(())
77}

Hide additional examples

examples/fluent_localization/translation.rs (line 54)

48pub fn get(
49    ctx: Context,
50    id: &str,
51    attr: Option<&str>,
52    args: Option<&fluent::FluentArgs<'_>>,
53) -> String {
54    let translations = &ctx.data().translations;
55    ctx.locale()
56        // Try to get the language-specific translation
57        .and_then(|locale| format(translations.other.get(locale)?, id, attr, args))
58        // Otherwise, fall back on main translation
59        .or_else(|| format(&translations.main, id, attr, args))
60        // If this message ID is not present in any translation files whatsoever
61        .unwrap_or_else(|| {
62            tracing::warn!("unknown fluent message identifier `{}`", id);
63            id.to_string()
64        })
65}

pub fn channel_id(self) -> ChannelId

Return the channel ID of this context

Examples found in repository?

examples/feature_showcase/collector.rs (line 26)

6pub async fn boop(ctx: Context<'_>) -> Result<(), Error> {
7    let uuid_boop = ctx.id();
8
9    let reply = {
10        let components = vec![serenity::CreateActionRow::Buttons(vec![
11            serenity::CreateButton::new(format!("{uuid_boop}"))
12                .style(serenity::ButtonStyle::Primary)
13                .label("Boop me!"),
14        ])];
15
16        CreateReply::default()
17            .content("I want some boops!")
18            .components(components)
19    };
20
21    ctx.send(reply).await?;
22
23    let mut boop_count = 0;
24    while let Some(mci) = serenity::ComponentInteractionCollector::new(ctx)
25        .author_id(ctx.author().id)
26        .channel_id(ctx.channel_id())
27        .timeout(std::time::Duration::from_secs(120))
28        .filter(move |mci| mci.data.custom_id == uuid_boop.to_string())
29        .await
30    {
31        boop_count += 1;
32
33        let mut msg = mci.message.clone();
34        msg.edit(
35            ctx,
36            serenity::EditMessage::new().content(format!("Boop count: {boop_count}")),
37        )
38        .await?;
39
40        mci.create_response(ctx, serenity::CreateInteractionResponse::Acknowledge)
41            .await?;
42    }
43
44    Ok(())
45}

pub fn guild_id(self) -> Option<GuildId>

Returns the guild ID of this context, if we are inside a guild

Examples found in repository?

examples/feature_showcase/checks.rs (line 48)

47async fn is_ferris(ctx: Context<'_>) -> Result<bool, Error> {
48    let nickname = match ctx.guild_id() {
49        Some(guild_id) => ctx.author().nick_in(ctx, guild_id).await,
50        None => None,
51    };
52    let name = nickname.as_ref().unwrap_or(&ctx.author().name);
53
54    Ok(name.eq_ignore_ascii_case("ferris"))
55}

pub async fn guild_channel(self) -> Option<GuildChannel>

Available on crate feature cache only.

Return the guild channel of this context, if we are inside a guild.

pub fn guild(self) -> Option<GuildRef<'a>>

Available on crate feature cache only.

Return the guild of this context, if we are inside a guild.

pub async fn partial_guild(self) -> Option<PartialGuild>

Return the partial guild of this context, if we are inside a guild.

Attempts to find the guild in cache, if cache feature is enabled. Otherwise, falls back to an HTTP request

Returns None if in DMs, or if the guild HTTP request fails

Examples found in repository?

examples/feature_showcase/checks.rs (line 117)

114pub async fn get_guild_name(ctx: Context<'_>) -> Result<(), Error> {
115    ctx.say(format!(
116        "The name of this guild is: {}",
117        ctx.partial_guild().await.unwrap().name
118    ))
119    .await?;
120
121    Ok(())
122}

Hide additional examples

examples/fluent_localization/main.rs (line 37)

35pub async fn info(ctx: Context<'_>) -> Result<(), Error> {
36    let guild = ctx
37        .partial_guild()
38        .await
39        .ok_or_else(|| tr!(ctx, "not-in-guild-error"))?;
40
41    let response = tr!(ctx, "guild-info",
42        name: &*guild.name,
43        emojiCount: guild.emojis.len(),
44        emojis: guild.emojis.values().map(|e| e.to_string()).collect::<String>(),
45        roleCount: guild.roles.len(),
46        stickerCount: guild.stickers.len(),
47    );
48
49    ctx.say(response).await?;
50    Ok(())
51}

pub async fn author_member(self) -> Option<Cow<'a, Member>>

Returns the author of the invoking message or interaction, as a serenity::Member

Returns a reference to the inner member object if in an crate::ApplicationContext, otherwise clones the member out of the cache, or fetches from the discord API.

Returns None if this command was invoked in DMs, or if the member cache lookup or HTTP request failed

Warning: can clone the entire Member instance out of the cache

pub fn created_at(self) -> Timestamp

Return the datetime of the invoking message or interaction

pub fn author(self) -> &'a User

Get the author of the command message or application command.

Examples found in repository?

examples/feature_showcase/response_with_reply.rs (line 5)

4pub async fn reply(ctx: Context<'_>) -> Result<(), Error> {
5    ctx.reply(format!("Hello {}!", ctx.author().name)).await?;
6    Ok(())
7}

Hide additional examples

examples/feature_showcase/checks.rs (line 49)

47async fn is_ferris(ctx: Context<'_>) -> Result<bool, Error> {
48    let nickname = match ctx.guild_id() {
49        Some(guild_id) => ctx.author().nick_in(ctx, guild_id).await,
50        None => None,
51    };
52    let name = nickname.as_ref().unwrap_or(&ctx.author().name);
53
54    Ok(name.eq_ignore_ascii_case("ferris"))
55}

examples/quickstart/main.rs (line 13)

9async fn age(
10    ctx: Context<'_>,
11    #[description = "Selected user"] user: Option<serenity::User>,
12) -> Result<(), Error> {
13    let u = user.as_ref().unwrap_or_else(|| ctx.author());
14    let response = format!("{}'s account was created at {}", u.name, u.created_at());
15    ctx.say(response).await?;
16    Ok(())
17}

examples/advanced_cooldowns/main.rs (line 14)

8async fn dynamic_cooldowns(ctx: Context<'_>) -> Result<(), Error> {
9    {
10        let mut cooldown_tracker = ctx.command().cooldowns.lock().unwrap();
11
12        // You can change the cooldown duration depending on the message author, for example
13        let mut cooldown_durations = poise::CooldownConfig::default();
14        if ctx.author().id == 472029906943868929 {
15            cooldown_durations.user = Some(std::time::Duration::from_secs(10));
16        }
17
18        match cooldown_tracker.remaining_cooldown(ctx.cooldown_context(), &cooldown_durations) {
19            Some(remaining) => {
20                return Err(format!("Please wait {} seconds", remaining.as_secs()).into())
21            }
22            None => cooldown_tracker.start_cooldown(ctx.cooldown_context()),
23        }
24    };
25
26    ctx.say("You successfully invoked the command!").await?;
27    Ok(())
28}

examples/feature_showcase/collector.rs (line 25)

6pub async fn boop(ctx: Context<'_>) -> Result<(), Error> {
7    let uuid_boop = ctx.id();
8
9    let reply = {
10        let components = vec![serenity::CreateActionRow::Buttons(vec![
11            serenity::CreateButton::new(format!("{uuid_boop}"))
12                .style(serenity::ButtonStyle::Primary)
13                .label("Boop me!"),
14        ])];
15
16        CreateReply::default()
17            .content("I want some boops!")
18            .components(components)
19    };
20
21    ctx.send(reply).await?;
22
23    let mut boop_count = 0;
24    while let Some(mci) = serenity::ComponentInteractionCollector::new(ctx)
25        .author_id(ctx.author().id)
26        .channel_id(ctx.channel_id())
27        .timeout(std::time::Duration::from_secs(120))
28        .filter(move |mci| mci.data.custom_id == uuid_boop.to_string())
29        .await
30    {
31        boop_count += 1;
32
33        let mut msg = mci.message.clone();
34        msg.edit(
35            ctx,
36            serenity::EditMessage::new().content(format!("Boop count: {boop_count}")),
37        )
38        .await?;
39
40        mci.create_response(ctx, serenity::CreateInteractionResponse::Acknowledge)
41            .await?;
42    }
43
44    Ok(())
45}

examples/basic_structure/main.rs (line 75)

40async fn main() {
41    tracing_subscriber::fmt::init();
42
43    // FrameworkOptions contains all of poise's configuration option in one struct
44    // Every option can be omitted to use its default value
45    let options = poise::FrameworkOptions {
46        commands: vec![commands::help(), commands::vote(), commands::getvotes()],
47        prefix_options: poise::PrefixFrameworkOptions {
48            prefix: Some("~".into()),
49            edit_tracker: Some(Arc::new(poise::EditTracker::for_timespan(
50                Duration::from_secs(3600),
51            ))),
52            additional_prefixes: vec![
53                poise::Prefix::Literal("hey bot,"),
54                poise::Prefix::Literal("hey bot"),
55            ],
56            ..Default::default()
57        },
58        // The global error handler for all error cases that may occur
59        on_error: |error| Box::pin(on_error(error)),
60        // This code is run before every command
61        pre_command: |ctx| {
62            Box::pin(async move {
63                println!("Executing command {}...", ctx.command().qualified_name);
64            })
65        },
66        // This code is run after a command if it was successful (returned Ok)
67        post_command: |ctx| {
68            Box::pin(async move {
69                println!("Executed command {}!", ctx.command().qualified_name);
70            })
71        },
72        // Every command invocation must pass this check to continue execution
73        command_check: Some(|ctx| {
74            Box::pin(async move {
75                if ctx.author().id == 123456789 {
76                    return Ok(false);
77                }
78                Ok(true)
79            })
80        }),
81        // Enforce command checks even for owners (enforced by default)
82        // Set to true to bypass checks, which is useful for testing
83        skip_checks_for_owners: false,
84        event_handler: |_ctx, event, _framework, _data| {
85            Box::pin(async move {
86                println!(
87                    "Got an event in event handler: {:?}",
88                    event.snake_case_name()
89                );
90                Ok(())
91            })
92        },
93        ..Default::default()
94    };
95
96    let framework = poise::Framework::builder()
97        .setup(move |ctx, ready, framework| {
98            Box::pin(async move {
99                println!("Logged in as {}", ready.user.name);
100                poise::builtins::register_globally(ctx, &framework.options().commands).await?;
101                Ok(Data {
102                    votes: Mutex::new(HashMap::new()),
103                })
104            })
105        })
106        .options(options)
107        .build();
108
109    let token = var("DISCORD_TOKEN")
110        .expect("Missing `DISCORD_TOKEN` env var, see README for more information.");
111    let intents =
112        serenity::GatewayIntents::non_privileged() | serenity::GatewayIntents::MESSAGE_CONTENT;
113
114    let client = serenity::ClientBuilder::new(token, intents)
115        .framework(framework)
116        .await;
117
118    client.unwrap().start().await.unwrap()
119}

pub fn id(self) -> u64

Available on crate feature chrono only.

Return a ID that uniquely identifies this command invocation.

Examples found in repository?

examples/feature_showcase/collector.rs (line 7)

6pub async fn boop(ctx: Context<'_>) -> Result<(), Error> {
7    let uuid_boop = ctx.id();
8
9    let reply = {
10        let components = vec![serenity::CreateActionRow::Buttons(vec![
11            serenity::CreateButton::new(format!("{uuid_boop}"))
12                .style(serenity::ButtonStyle::Primary)
13                .label("Boop me!"),
14        ])];
15
16        CreateReply::default()
17            .content("I want some boops!")
18            .components(components)
19    };
20
21    ctx.send(reply).await?;
22
23    let mut boop_count = 0;
24    while let Some(mci) = serenity::ComponentInteractionCollector::new(ctx)
25        .author_id(ctx.author().id)
26        .channel_id(ctx.channel_id())
27        .timeout(std::time::Duration::from_secs(120))
28        .filter(move |mci| mci.data.custom_id == uuid_boop.to_string())
29        .await
30    {
31        boop_count += 1;
32
33        let mut msg = mci.message.clone();
34        msg.edit(
35            ctx,
36            serenity::EditMessage::new().content(format!("Boop count: {boop_count}")),
37        )
38        .await?;
39
40        mci.create_response(ctx, serenity::CreateInteractionResponse::Acknowledge)
41            .await?;
42    }
43
44    Ok(())
45}

pub fn parent_commands(self) -> &'a [&'a Command<U, E>]

If the invoked command was a subcommand, these are the parent commands, ordered top-level downwards.

pub fn command(self) -> &'a Command<U, E>

Returns a reference to the command.

Examples found in repository?

examples/basic_structure/main.rs (line 29)

22async fn on_error(error: poise::FrameworkError<'_, Data, Error>) {
23    // This is our custom error handler
24    // They are many errors that can occur, so we only handle the ones we want to customize
25    // and forward the rest to the default handler
26    match error {
27        poise::FrameworkError::Setup { error, .. } => panic!("Failed to start bot: {:?}", error),
28        poise::FrameworkError::Command { error, ctx, .. } => {
29            println!("Error in command `{}`: {:?}", ctx.command().name, error,);
30        }
31        error => {
32            if let Err(e) = poise::builtins::on_error(error).await {
33                println!("Error while handling error: {}", e)
34            }
35        }
36    }
37}
38
39#[tokio::main]
40async fn main() {
41    tracing_subscriber::fmt::init();
42
43    // FrameworkOptions contains all of poise's configuration option in one struct
44    // Every option can be omitted to use its default value
45    let options = poise::FrameworkOptions {
46        commands: vec![commands::help(), commands::vote(), commands::getvotes()],
47        prefix_options: poise::PrefixFrameworkOptions {
48            prefix: Some("~".into()),
49            edit_tracker: Some(Arc::new(poise::EditTracker::for_timespan(
50                Duration::from_secs(3600),
51            ))),
52            additional_prefixes: vec![
53                poise::Prefix::Literal("hey bot,"),
54                poise::Prefix::Literal("hey bot"),
55            ],
56            ..Default::default()
57        },
58        // The global error handler for all error cases that may occur
59        on_error: |error| Box::pin(on_error(error)),
60        // This code is run before every command
61        pre_command: |ctx| {
62            Box::pin(async move {
63                println!("Executing command {}...", ctx.command().qualified_name);
64            })
65        },
66        // This code is run after a command if it was successful (returned Ok)
67        post_command: |ctx| {
68            Box::pin(async move {
69                println!("Executed command {}!", ctx.command().qualified_name);
70            })
71        },
72        // Every command invocation must pass this check to continue execution
73        command_check: Some(|ctx| {
74            Box::pin(async move {
75                if ctx.author().id == 123456789 {
76                    return Ok(false);
77                }
78                Ok(true)
79            })
80        }),
81        // Enforce command checks even for owners (enforced by default)
82        // Set to true to bypass checks, which is useful for testing
83        skip_checks_for_owners: false,
84        event_handler: |_ctx, event, _framework, _data| {
85            Box::pin(async move {
86                println!(
87                    "Got an event in event handler: {:?}",
88                    event.snake_case_name()
89                );
90                Ok(())
91            })
92        },
93        ..Default::default()
94    };
95
96    let framework = poise::Framework::builder()
97        .setup(move |ctx, ready, framework| {
98            Box::pin(async move {
99                println!("Logged in as {}", ready.user.name);
100                poise::builtins::register_globally(ctx, &framework.options().commands).await?;
101                Ok(Data {
102                    votes: Mutex::new(HashMap::new()),
103                })
104            })
105        })
106        .options(options)
107        .build();
108
109    let token = var("DISCORD_TOKEN")
110        .expect("Missing `DISCORD_TOKEN` env var, see README for more information.");
111    let intents =
112        serenity::GatewayIntents::non_privileged() | serenity::GatewayIntents::MESSAGE_CONTENT;
113
114    let client = serenity::ClientBuilder::new(token, intents)
115        .framework(framework)
116        .await;
117
118    client.unwrap().start().await.unwrap()
119}

Hide additional examples

examples/advanced_cooldowns/main.rs (line 10)

8async fn dynamic_cooldowns(ctx: Context<'_>) -> Result<(), Error> {
9    {
10        let mut cooldown_tracker = ctx.command().cooldowns.lock().unwrap();
11
12        // You can change the cooldown duration depending on the message author, for example
13        let mut cooldown_durations = poise::CooldownConfig::default();
14        if ctx.author().id == 472029906943868929 {
15            cooldown_durations.user = Some(std::time::Duration::from_secs(10));
16        }
17
18        match cooldown_tracker.remaining_cooldown(ctx.cooldown_context(), &cooldown_durations) {
19            Some(remaining) => {
20                return Err(format!("Please wait {} seconds", remaining.as_secs()).into())
21            }
22            None => cooldown_tracker.start_cooldown(ctx.cooldown_context()),
23        }
24    };
25
26    ctx.say("You successfully invoked the command!").await?;
27    Ok(())
28}

pub fn prefix(self) -> &'a str

Returns the prefix this command was invoked with, or a slash (/), if this is an application command.

pub fn invoked_command_name(self) -> &'a str

Returns the command name that this command was invoked with

Mainly useful in prefix context, for example to check whether a command alias was used.

In slash contexts, the given command name will always be returned verbatim, since there are no slash command aliases and the user has no control over spelling

Examples found in repository?

examples/help_generation/main.rs (line 296)

287async fn help(
288    ctx: Context<'_>,
289    #[description = "Command to get help for"]
290    #[rest]
291    mut command: Option<String>,
292) -> Result<(), Error> {
293    // This makes it possible to just make `help` a subcommand of any command
294    // `/fruit help` turns into `/help fruit`
295    // `/fruit help apple` turns into `/help fruit apple`
296    if ctx.invoked_command_name() != "help" {
297        command = match command {
298            Some(c) => Some(format!("{} {}", ctx.invoked_command_name(), c)),
299            None => Some(ctx.invoked_command_name().to_string()),
300        };
301    }
302    let extra_text_at_bottom = "\
303Type `?help command` for more info on a command.
304You can edit your `?help` message to the bot and the bot will edit its response.";
305
306    let config = HelpConfiguration {
307        show_subcommands: true,
308        show_context_menu_commands: true,
309        ephemeral: true,
310        extra_text_at_bottom,
311
312        ..Default::default()
313    };
314    poise::builtins::help(ctx, command.as_deref(), config).await?;
315    Ok(())
316}

pub async fn rerun(self) -> Result<(), E>

Re-runs this entire command invocation

Permission checks are omitted; the command code is directly executed as a function. The result is returned by this function

pub fn invocation_string(self) -> String

Returns the string with which this command was invoked.

For example "/slash_command subcommand arg1:value1 arg2:value2".

Examples found in repository?

examples/feature_showcase/inherit_checks.rs (line 18)

17async fn child2(ctx: Context<'_>, _b: bool, _s: String, _i: u32) -> Result<(), Error> {
18    ctx.say(ctx.invocation_string()).await?;
19    Ok(())
20}

pub async fn set_invocation_data<T: 'static + Send + Sync>(self, data: T)

Stores the given value as the data for this command invocation

This data is carried across the pre_command hook, checks, main command execution, and post_command. It may be useful to cache data or pass information to later phases of command execution.

Examples found in repository?

examples/invocation_data/main.rs (line 83)

51async fn main() {
52    let token = std::env::var("TOKEN").unwrap();
53    let intents =
54        serenity::GatewayIntents::non_privileged() | serenity::GatewayIntents::MESSAGE_CONTENT;
55
56    let framework = poise::Framework::builder()
57        .setup(move |ctx, _, framework| {
58            Box::pin(async move {
59                let commands =
60                    poise::builtins::create_application_commands(&framework.options().commands);
61
62                serenity::GuildId::new(703332075914264606)
63                    .set_commands(ctx, commands)
64                    .await
65                    .unwrap();
66                Ok(())
67            })
68        })
69        .options(poise::FrameworkOptions {
70            pre_command: |ctx| {
71                Box::pin(async move {
72                    println!(
73                        "In pre_command: {:?}",
74                        ctx.invocation_data::<&str>().await.as_deref()
75                    );
76                })
77            },
78            command_check: Some(|ctx| {
79                Box::pin(async move {
80                    // Global command check is the first callback that's invoked, so let's set the
81                    // data here
82                    println!("Writing invocation data!");
83                    ctx.set_invocation_data("hello").await;
84
85                    println!(
86                        "In global check: {:?}",
87                        ctx.invocation_data::<&str>().await.as_deref()
88                    );
89
90                    Ok(true)
91                })
92            }),
93            post_command: |ctx| {
94                Box::pin(async move {
95                    println!(
96                        "In post_command: {:?}",
97                        ctx.invocation_data::<&str>().await.as_deref()
98                    );
99                })
100            },
101            on_error: |err| {
102                Box::pin(async move {
103                    match err {
104                        poise::FrameworkError::Command { ctx, .. } => {
105                            println!(
106                                "In on_error: {:?}",
107                                ctx.invocation_data::<&str>().await.as_deref()
108                            );
109                        }
110                        err => poise::builtins::on_error(err).await.unwrap(),
111                    }
112                })
113            },
114
115            commands: vec![invocation_data_test()],
116            prefix_options: poise::PrefixFrameworkOptions {
117                prefix: Some("~".into()),
118                ..Default::default()
119            },
120            ..Default::default()
121        })
122        .build();
123
124    let client = serenity::ClientBuilder::new(token, intents)
125        .framework(framework)
126        .await;
127
128    client.unwrap().start().await.unwrap();
129}

pub async fn invocation_data<T: 'static>( self, ) -> Option<impl DerefMut<Target = T> + 'a>

Attempts to get the invocation data with the requested type

If the stored invocation data has a different type than requested, None is returned

Examples found in repository?

examples/invocation_data/main.rs (line 15)

12async fn my_check(ctx: Context<'_>) -> Result<bool, Error> {
13    println!(
14        "In command specific check: {:?}",
15        ctx.invocation_data::<&str>().await.as_deref()
16    );
17
18    Ok(true)
19}
20
21async fn my_autocomplete(ctx: Context<'_>, _: &str) -> impl Iterator<Item = String> {
22    println!(
23        "In autocomplete: {:?}",
24        ctx.invocation_data::<&str>().await.as_deref()
25    );
26
27    std::iter::empty()
28}
29
30/// Test command to ensure that invocation_data works
31#[poise::command(prefix_command, slash_command, check = "my_check")]
32pub async fn invocation_data_test(
33    ctx: Context<'_>,
34    #[description = "Whether this command should succeed (yes if 0, no otherwise)"]
35    #[autocomplete = "my_autocomplete"]
36    should_succeed: u32,
37) -> Result<(), Error> {
38    println!(
39        "In command: {:?}",
40        ctx.invocation_data::<&str>().await.as_deref()
41    );
42
43    if should_succeed > 0 {
44        Ok(())
45    } else {
46        Err("")
47    }
48}
49
50#[tokio::main]
51async fn main() {
52    let token = std::env::var("TOKEN").unwrap();
53    let intents =
54        serenity::GatewayIntents::non_privileged() | serenity::GatewayIntents::MESSAGE_CONTENT;
55
56    let framework = poise::Framework::builder()
57        .setup(move |ctx, _, framework| {
58            Box::pin(async move {
59                let commands =
60                    poise::builtins::create_application_commands(&framework.options().commands);
61
62                serenity::GuildId::new(703332075914264606)
63                    .set_commands(ctx, commands)
64                    .await
65                    .unwrap();
66                Ok(())
67            })
68        })
69        .options(poise::FrameworkOptions {
70            pre_command: |ctx| {
71                Box::pin(async move {
72                    println!(
73                        "In pre_command: {:?}",
74                        ctx.invocation_data::<&str>().await.as_deref()
75                    );
76                })
77            },
78            command_check: Some(|ctx| {
79                Box::pin(async move {
80                    // Global command check is the first callback that's invoked, so let's set the
81                    // data here
82                    println!("Writing invocation data!");
83                    ctx.set_invocation_data("hello").await;
84
85                    println!(
86                        "In global check: {:?}",
87                        ctx.invocation_data::<&str>().await.as_deref()
88                    );
89
90                    Ok(true)
91                })
92            }),
93            post_command: |ctx| {
94                Box::pin(async move {
95                    println!(
96                        "In post_command: {:?}",
97                        ctx.invocation_data::<&str>().await.as_deref()
98                    );
99                })
100            },
101            on_error: |err| {
102                Box::pin(async move {
103                    match err {
104                        poise::FrameworkError::Command { ctx, .. } => {
105                            println!(
106                                "In on_error: {:?}",
107                                ctx.invocation_data::<&str>().await.as_deref()
108                            );
109                        }
110                        err => poise::builtins::on_error(err).await.unwrap(),
111                    }
112                })
113            },
114
115            commands: vec![invocation_data_test()],
116            prefix_options: poise::PrefixFrameworkOptions {
117                prefix: Some("~".into()),
118                ..Default::default()
119            },
120            ..Default::default()
121        })
122        .build();
123
124    let client = serenity::ClientBuilder::new(token, intents)
125        .framework(framework)
126        .await;
127
128    client.unwrap().start().await.unwrap();
129}

pub fn locale(self) -> Option<&'a str>

If available, returns the locale (selected language) of the invoking user

Examples found in repository?

examples/fluent_localization/translation.rs (line 55)

48pub fn get(
49    ctx: Context,
50    id: &str,
51    attr: Option<&str>,
52    args: Option<&fluent::FluentArgs<'_>>,
53) -> String {
54    let translations = &ctx.data().translations;
55    ctx.locale()
56        // Try to get the language-specific translation
57        .and_then(|locale| format(translations.other.get(locale)?, id, attr, args))
58        // Otherwise, fall back on main translation
59        .or_else(|| format(&translations.main, id, attr, args))
60        // If this message ID is not present in any translation files whatsoever
61        .unwrap_or_else(|| {
62            tracing::warn!("unknown fluent message identifier `{}`", id);
63            id.to_string()
64        })
65}

Hide additional examples

examples/feature_showcase/localization.rs (line 61)

45pub async fn welcome(
46    ctx: Context<'_>,
47    #[name_localized("de", "nutzer")]
48    #[description_localized("de", "Der zu begrüßende Nutzer")]
49    #[name_localized("es-ES", "usuario")]
50    #[description_localized("es-ES", "El usuario a saludar")]
51    #[description = "The user to welcome"]
52    user: serenity::User,
53    #[name_localized("de", "nachricht")]
54    #[description_localized("de", "Die versendete Nachricht")]
55    #[name_localized("es-ES", "mensaje")]
56    #[description_localized("es-ES", "El mensaje enviado")]
57    #[description = "The message to send"]
58    message: WelcomeChoice,
59) -> Result<(), Error> {
60    let message = message
61        .localized_name(ctx.locale().unwrap_or(""))
62        .unwrap_or_else(|| message.name());
63    ctx.say(format!("<@{}> {}", user.id, message)).await?;
64    Ok(())
65}

pub fn reply_builder(self, builder: CreateReply) -> CreateReply

Builds a crate::CreateReply by combining the builder closure with the defaults that were pre-configured in poise.

This is primarily an internal function and only exposed for people who want to manually convert crate::CreateReply instances into Discord requests.

pub fn cache(self) -> &'a Cache

Available on crate feature cache only.

Returns serenity’s cache which stores various useful data received from the gateway

Shorthand for .serenity_context().cache

pub fn http(self) -> &'a Http

Returns serenity’s raw Discord API client to make raw API requests, if needed.

Shorthand for .serenity_context().http

pub async fn ping(self) -> Duration

Returns the current gateway heartbeat latency (::serenity::gateway::Shard::latency()).

If the shard has just connected, this value is zero.