Struct poise::structs::ApplicationContext

pub struct ApplicationContext<'a, U, E> {
    pub serenity_context: &'a Context,
    pub interaction: ApplicationCommandOrAutocompleteInteraction<'a>,
    pub args: &'a [CommandDataOption],
    pub has_sent_initial_response: &'a AtomicBool,
    pub framework: FrameworkContext<'a, U, E>,
    pub parent_commands: &'a [&'a Command<U, E>],
    pub command: &'a Command<U, E>,
    pub data: &'a U,
    pub invocation_data: &'a Mutex<Box<dyn Any + Send + Sync>>,
    /* private fields */
}

Application command specific context passed to command invocations.

Fields
serenity_context: &'a Context

Serenity’s context, like HTTP or cache

interaction: ApplicationCommandOrAutocompleteInteraction<'a>

The interaction which triggered this command execution.

args: &'a [CommandDataOption]

Slash command arguments

Not equivalent to self.interaction.data().options. That one refers to just the top-level command arguments, whereas Self::args is the options of the actual subcommand, if any.

has_sent_initial_response: &'a AtomicBool

Keeps track of whether an initial response has been sent.

Discord requires different HTTP endpoints for initial and additional responses.

framework: FrameworkContext<'a, U, E>

Read-only reference to the framework

Useful if you need the list of commands, for example for a custom help command

parent_commands: &'a [&'a Command<U, E>]

If the invoked command was a subcommand, these are the parent commands, ordered top down.

command: &'a Command<U, E>

The command object which is the current command

data: &'a U

Your custom user data

invocation_data: &'a Mutex<Box<dyn Any + Send + Sync>>

Custom user data carried across a single command invocation

Implementations

impl<U, E> ApplicationContext<'_, U, E>

pub async fn defer_response(&self, ephemeral: bool) -> Result<(), Error>

See crate::Context::defer()

Examples found in repository?

src/structs/context.rs (line 54)

52
53
54
55
56
57
58
59
60
61
62
63
64
65
66
67
68
69
70
71
72
73
74
75
76
77
78
79
80
81
82
83
84
85
86

    pub async fn defer(self) -> Result<(), serenity::Error> {
        if let Self::Application(ctx) = self {
            ctx.defer_response(false).await?;
        }
        Ok(())
    }

    /// See [`Self::defer()`]
    ///
    /// This will make the response ephemeral; to make it public, use [`Self::defer()`].
    pub async fn defer_ephemeral(self) -> Result<(), serenity::Error> {
        if let Self::Application(ctx) = self {
            ctx.defer_response(true).await?;
        }
        Ok(())
    }

    /// If this is an application command, [`Self::defer()`] is called
    ///
    /// If this is a prefix command, a typing broadcast is started until the return value is
    /// dropped.
    // #[must_use = "The typing broadcast will only persist if you store it"] // currently doesn't work
    pub async fn defer_or_broadcast(self) -> Result<Option<serenity::Typing>, serenity::Error> {
        Ok(match self {
            Self::Application(ctx) => {
                ctx.defer_response(false).await?;
                None
            }
            Self::Prefix(ctx) => Some(
                ctx.msg
                    .channel_id
                    .start_typing(&ctx.serenity_context.http)?,
            ),
        })
    }