poise::structs

Struct CommandParameter

pub struct CommandParameter<U, E> {
    pub name: Cow<'static, str>,
    pub name_localizations: Cow<'static, [(Cow<'static, str>, Cow<'static, str>)]>,
    pub description: Option<Cow<'static, str>>,
    pub description_localizations: Cow<'static, [(Cow<'static, str>, Cow<'static, str>)]>,
    pub required: bool,
    pub channel_types: Option<Cow<'static, [ChannelType]>>,
    pub choices: Cow<'static, [CommandParameterChoice]>,
    pub type_setter: Option<fn(CreateCommandOption) -> CreateCommandOption>,
    pub autocomplete_callback: Option<for<'a> fn(ApplicationContext<'a, U, E>, &'a str) -> BoxFuture<'a, CreateAutocompleteResponse>>,
    /* private fields */
}

A single parameter of a crate::Command

Fields
name: Cow<'static, str>

Name of this command parameter

name_localizations: Cow<'static, [(Cow<'static, str>, Cow<'static, str>)]>

Localized names with locale string as the key (slash-only)

description: Option<Cow<'static, str>>

Description of the command. Required for slash commands

description_localizations: Cow<'static, [(Cow<'static, str>, Cow<'static, str>)]>

Localized descriptions with locale string as the key (slash-only)

required: bool

true is this parameter is required, false if it’s optional or variadic

channel_types: Option<Cow<'static, [ChannelType]>>

If this parameter is a channel, users can only enter these channel types in a slash command

Prefix commands are currently unaffected by this

choices: Cow<'static, [CommandParameterChoice]>

If this parameter is a choice parameter, this is the fixed list of options

type_setter: Option<fn(CreateCommandOption) -> CreateCommandOption>

Closure that sets this parameter’s type and min/max value in the given builder

For example a u32 CommandParameter would store this as the Self::type_setter:

|b| b.kind(serenity::CommandOptionType::Integer).min_int_value(0).max_int_value(u64::MAX)

autocomplete_callback: Option<for<'a> fn(ApplicationContext<'a, U, E>, &'a str) -> BoxFuture<'a, CreateAutocompleteResponse>>

Optionally, a callback that is invoked on autocomplete interactions. This closure should extract the partial argument from the given JSON value and generate the autocomplete response which contains the list of autocomplete suggestions.

Implementations

impl<U, E> CommandParameter<U, E>

pub fn create_as_slash_command_option(&self) -> Option<CreateCommandOption>

Generates a slash command parameter builder from this CommandParameter instance. This can be used to register the command on Discord’s servers