Struct poise::structs::PrefixContext

pub struct PrefixContext<'a, U, E> {
    pub serenity_context: &'a Context,
    pub msg: &'a Message,
    pub prefix: &'a str,
    pub invoked_command_name: &'a str,
    pub args: &'a str,
    pub framework: FrameworkContext<'a, U, E>,
    pub parent_commands: &'a [&'a Command<U, E>],
    pub command: &'a Command<U, E>,
    pub data: &'a U,
    pub invocation_data: &'a Mutex<Box<dyn Any + Send + Sync>>,
    pub trigger: MessageDispatchTrigger,
    pub action: fn(_: PrefixContext<'_, U, E>) -> BoxFuture<'_, Result<(), FrameworkError<'_, U, E>>>,
    /* private fields */
}

Prefix-specific context passed to command invocations.

Contains the trigger message, the Discord connection management stuff, and the user data.

Fields
serenity_context: &'a Context

Serenity’s context, like HTTP or cache

msg: &'a Message

The invoking user message

prefix: &'a str

Prefix used by the user to invoke this command

invoked_command_name: &'a str

Command name used by the user to invoke this command

args: &'a str

Entire argument string

framework: FrameworkContext<'a, U, E>

Read-only reference to the framework

Useful if you need the list of commands, for example for a custom help command

parent_commands: &'a [&'a Command<U, E>]

If the invoked command was a subcommand, these are the parent commands, ordered top down.

command: &'a Command<U, E>

The command object which is the current command

data: &'a U

Your custom user data

invocation_data: &'a Mutex<Box<dyn Any + Send + Sync>>

Custom user data carried across a single command invocation

trigger: MessageDispatchTrigger

How this command invocation was triggered

action: fn(_: PrefixContext<'_, U, E>) -> BoxFuture<'_, Result<(), FrameworkError<'_, U, E>>>

The function that is called to execute the actual command

Implementations

impl<'a, U, E> PrefixContext<'a, U, E>

pub async fn defer(self) -> Result<(), Error>

Defer the response, giving the bot multiple minutes to respond without the user seeing an “interaction failed error”.

Also sets the crate::ApplicationContext::has_sent_initial_response flag so subsequent responses will be sent in the correct manner.

No-op if this is an autocomplete context

This will make the response public; to make it ephemeral, use Self::defer_ephemeral().

pub async fn defer_ephemeral(self) -> Result<(), Error>

See Self::defer()

This will make the response ephemeral; to make it public, use Self::defer().

pub async fn defer_or_broadcast(self) -> Result<Option<Typing>, Error>

If this is an application command, Self::defer() is called

If this is a prefix command, a typing broadcast is started until the return value is dropped.

pub async fn say( self, text: impl Into<String> ) -> Result<ReplyHandle<'a>, Error>

Shorthand of crate::say_reply

Note: panics when called in an autocomplete context!

pub async fn reply( self, text: impl Into<String> ) -> Result<ReplyHandle<'a>, Error>

Like Self::say, but formats the message as a reply to the user’s command message.

Equivalent to .send(|b| b.content("...").reply(true)).

Only has an effect in prefix context, because slash command responses are always formatted as a reply.

Note: panics when called in an autocomplete context!

pub async fn send<'att>( self, builder: CreateReply ) -> Result<ReplyHandle<'a>, Error>

Shorthand of crate::send_reply

Note: panics when called in an autocomplete context!

pub fn serenity_context(self) -> &'a Context

Return the stored serenity::Context within the underlying context type.

pub fn cooldown_context(self) -> CooldownContext

Create a crate::CooldownContext based off the underlying context type.

pub fn discord(self) -> &'a Context

👎Deprecated: poise::Context can now be passed directly into most serenity functions. Otherwise, use .serenity_context() now

See Self::serenity_context.

pub fn framework(self) -> FrameworkContext<'a, U, E>

Returns a view into data stored by the framework, like configuration

pub fn data(self) -> &'a U

Return a reference to your custom user data

pub fn channel_id(self) -> ChannelId

Return the channel ID of this context

pub fn guild_id(self) -> Option<GuildId>

Returns the guild ID of this context, if we are inside a guild

pub async fn guild_channel(self) -> Option<GuildChannel>

Available on crate feature cache only.

Return the guild channel of this context, if we are inside a guild.

pub fn guild(self) -> Option<GuildRef<'a>>

Available on crate feature cache only.

Return the guild of this context, if we are inside a guild.

pub async fn partial_guild(self) -> Option<PartialGuild>

Return the partial guild of this context, if we are inside a guild.

Attempts to find the guild in cache, if cache feature is enabled. Otherwise, falls back to an HTTP request

Returns None if in DMs, or if the guild HTTP request fails

pub async fn author_member(self) -> Option<Cow<'a, Member>>

Returns the author of the invoking message or interaction, as a serenity::Member

Returns a reference to the inner member object if in an crate::ApplicationContext, otherwise clones the member out of the cache, or fetches from the discord API.

Returns None if this command was invoked in DMs, or if the member cache lookup or HTTP request failed

Warning: can clone the entire Member instance out of the cache

pub fn created_at(self) -> Timestamp

Return the datetime of the invoking message or interaction

pub fn author(self) -> &'a User

Get the author of the command message or application command.

pub fn id(self) -> u64

Available on crate features chrono or time only.

Return a ID that uniquely identifies this command invocation.

pub fn parent_commands(self) -> &'a [&'a Command<U, E>]

If the invoked command was a subcommand, these are the parent commands, ordered top-level downwards.

pub fn command(self) -> &'a Command<U, E>

Returns a reference to the command.

pub fn prefix(self) -> &'a str

Returns the prefix this command was invoked with, or a slash (/), if this is an application command.

pub fn invoked_command_name(self) -> &'a str

Returns the command name that this command was invoked with

Mainly useful in prefix context, for example to check whether a command alias was used.

In slash contexts, the given command name will always be returned verbatim, since there are no slash command aliases and the user has no control over spelling

pub async fn rerun(self) -> Result<(), E>

Re-runs this entire command invocation

Permission checks are omitted; the command code is directly executed as a function. The result is returned by this function

pub fn invocation_string(self) -> String

Returns the string with which this command was invoked.

For example "/slash_command subcommand arg1:value1 arg2:value2".

pub async fn set_invocation_data<T: 'static + Send + Sync>(self, data: T)

Stores the given value as the data for this command invocation

This data is carried across the pre_command hook, checks, main command execution, and post_command. It may be useful to cache data or pass information to later phases of command execution.

pub async fn invocation_data<T: 'static>( self ) -> Option<impl DerefMut<Target = T> + 'a>

Attempts to get the invocation data with the requested type

If the stored invocation data has a different type than requested, None is returned

pub fn locale(self) -> Option<&'a str>

If available, returns the locale (selected language) of the invoking user

pub fn reply_builder(self, builder: CreateReply) -> CreateReply

Builds a crate::CreateReply by combining the builder closure with the defaults that were pre-configured in poise.

This is primarily an internal function and only exposed for people who want to manually convert crate::CreateReply instances into Discord requests.

pub fn cache(self) -> &'a Cache

Available on crate feature cache only.

Returns serenity’s cache which stores various useful data received from the gateway

Shorthand for .serenity_context().cache

pub fn http(self) -> &'a Http

Returns serenity’s raw Discord API client to make raw API requests, if needed.

Shorthand for .serenity_context().http

pub async fn ping(self) -> Duration

Returns the current gateway heartbeat latency (::serenity::gateway::Shard::latency()).

If the shard has just connected, this value is zero.