Struct poise::track_edits::EditTracker

pub struct EditTracker { /* private fields */ }

Stores messages and the associated bot responses in order to implement poise’s edit tracking feature.

Implementations

impl EditTracker

pub fn for_timespan(duration: Duration) -> RwLock<Self>

Create an edit tracker which tracks messages for the specified duration.

Note: EditTracker will only purge messages outside the duration when Self::purge is called. If you supply the created EditTracker to crate::Framework, the framework will take care of that by calling Self::purge periodically.

pub fn process_message_update(
&mut self,
user_msg_update: &MessageUpdateEvent,
ignore_edits_if_not_yet_responded: bool
) -> Option<(Message, bool)>

Returns a copy of a newly up-to-date cached message, or a brand new generated message when not in cache. Also returns a bool with true if this message was previously tracked

Returns None if the command shouldn’t be re-run, e.g. if the message content wasn’t edited

Examples found in repository?

src/dispatch/mod.rs (lines 77-83)

52
53
54
55
56
57
58
59
60
61
62
63
64
65
66
67
68
69
70
71
72
73
74
75
76
77
78
79
80
81
82
83
84
85
86
87
88
89
90
91
92
93
94
95
96
97
98
99
100
101
102
103
104
105
106
107
108
109
110
111
112
113
114
115
116
117
118
119
120
121
122
123
124
125
126
127
128
129
130
131
132
133
134
135
136
137
138
139
140
141
142
143
144
145
146
147
148
149
150
151
152
153
154
155
156
157
158
159

pub async fn dispatch_event<U: Send + Sync, E>(
    framework: crate::FrameworkContext<'_, U, E>,
    ctx: &serenity::Context,
    event: &crate::Event<'_>,
) {
    match event {
        crate::Event::Message { new_message } => {
            let invocation_data = tokio::sync::Mutex::new(Box::new(()) as _);
            let mut parent_commands = Vec::new();
            let trigger = crate::MessageDispatchTrigger::MessageCreate;
            if let Err(error) = prefix::dispatch_message(
                framework,
                ctx,
                new_message,
                trigger,
                &invocation_data,
                &mut parent_commands,
            )
            .await
            {
                error.handle(framework.options).await;
            }
        }
        crate::Event::MessageUpdate { event, .. } => {
            if let Some(edit_tracker) = &framework.options.prefix_options.edit_tracker {
                let msg = edit_tracker.write().unwrap().process_message_update(
                    event,
                    framework
                        .options()
                        .prefix_options
                        .ignore_edits_if_not_yet_responded,
                );

                if let Some((msg, previously_tracked)) = msg {
                    let invocation_data = tokio::sync::Mutex::new(Box::new(()) as _);
                    let mut parent_commands = Vec::new();
                    let trigger = match previously_tracked {
                        true => crate::MessageDispatchTrigger::MessageEditFromInvalid,
                        false => crate::MessageDispatchTrigger::MessageEdit,
                    };
                    if let Err(error) = prefix::dispatch_message(
                        framework,
                        ctx,
                        &msg,
                        trigger,
                        &invocation_data,
                        &mut parent_commands,
                    )
                    .await
                    {
                        error.handle(framework.options).await;
                    }
                }
            }
        }
        crate::Event::InteractionCreate {
            interaction: serenity::Interaction::ApplicationCommand(interaction),
        } => {
            let invocation_data = tokio::sync::Mutex::new(Box::new(()) as _);
            let mut parent_commands = Vec::new();
            if let Err(error) = slash::dispatch_interaction(
                framework,
                ctx,
                interaction,
                &std::sync::atomic::AtomicBool::new(false),
                &invocation_data,
                &mut parent_commands,
            )
            .await
            {
                error.handle(framework.options).await;
            }
        }
        crate::Event::InteractionCreate {
            interaction: serenity::Interaction::Autocomplete(interaction),
        } => {
            let invocation_data = tokio::sync::Mutex::new(Box::new(()) as _);
            let mut parent_commands = Vec::new();
            if let Err(error) = slash::dispatch_autocomplete(
                framework,
                ctx,
                interaction,
                &std::sync::atomic::AtomicBool::new(false),
                &invocation_data,
                &mut parent_commands,
            )
            .await
            {
                error.handle(framework.options).await;
            }
        }
        _ => {}
    }

    // Do this after the framework's Ready handling, so that get_user_data() doesnt
    // potentially block infinitely
    if let Err(error) =
        (framework.options.event_handler)(ctx, event, framework, framework.user_data().await).await
    {
        let error = crate::FrameworkError::EventHandler {
            ctx,
            error,
            event,
            framework,
        };
        (framework.options.on_error)(error).await;
    }
}

pub fn purge(&mut self)

Forget all of the messages that are older than the specified duration.

Examples found in repository?

src/framework/mod.rs (line 334)

328
329
330
331
332
333
334
335
336
337
338
339
340
341

fn spawn_edit_tracker_purge_task<U: 'static + Send + Sync, E: 'static>(
    framework: std::sync::Arc<Framework<U, E>>,
) -> tokio::task::JoinHandle<()> {
    tokio::spawn(async move {
        if let Some(edit_tracker) = &framework.options.prefix_options.edit_tracker {
            loop {
                edit_tracker.write().unwrap().purge();

                // not sure if the purging interval should be configurable
                tokio::time::sleep(std::time::Duration::from_secs(60)).await;
            }
        }
    })
}

pub fn find_bot_response(&self, user_msg_id: MessageId) -> Option<&Message>

Given a message by a user, find the corresponding bot response, if one exists and is cached.

Examples found in repository?

src/reply/send_reply.rs (line 154)

133
134
135
136
137
138
139
140
141
142
143
144
145
146
147
148
149
150
151
152
153
154
155
156
157
158
159
160
161
162
163
164
165
166
167
168
169
170
171
172
173
174
175
176
177
178
179
180
181
182
183
184
185
186
187
188
189
190
191
192
193
194
195

async fn _send_prefix_reply<'a, U, E>(
    ctx: crate::PrefixContext<'_, U, E>,
    mut reply: crate::CreateReply<'a>,
) -> Result<Box<serenity::Message>, serenity::Error> {
    if let Some(callback) = ctx.framework.options().reply_callback {
        callback(ctx.into(), &mut reply);
    }

    // This must only return None when we _actually_ want to reuse the existing response! There are
    // no checks later
    let lock_edit_tracker = || {
        if ctx.command.reuse_response {
            if let Some(edit_tracker) = &ctx.framework.options().prefix_options.edit_tracker {
                return Some(edit_tracker.write().unwrap());
            }
        }
        None
    };

    let existing_response = lock_edit_tracker()
        .as_mut()
        .and_then(|t| t.find_bot_response(ctx.msg.id))
        .cloned();

    Ok(Box::new(if let Some(mut response) = existing_response {
        response
            .edit(ctx.serenity_context, |f| {
                // Reset the message. We don't want leftovers of the previous message (e.g. user
                // sends a message with `.content("abc")` in a track_edits command, and the edited
                // message happens to contain embeds, we don't want to keep those embeds)
                // (*f = Default::default() won't do)
                f.content("");
                f.set_embeds(Vec::new());
                f.components(|b| b);
                f.0.insert("attachments", serenity::json::json! { [] });

                reply.to_prefix_edit(f);
                f
            })
            .await?;

        // If the entry still exists after the await, update it to the new contents
        if let Some(mut edit_tracker) = lock_edit_tracker() {
            edit_tracker.set_bot_response(ctx.msg, response.clone());
        }

        response
    } else {
        let new_response = ctx
            .msg
            .channel_id
            .send_message(ctx.serenity_context, |m| {
                reply.to_prefix(m, ctx.msg);
                m
            })
            .await?;
        if let Some(track_edits) = &mut lock_edit_tracker() {
            track_edits.set_bot_response(ctx.msg, new_response.clone());
        }

        new_response
    }))
}

pub fn set_bot_response(&mut self, user_msg: &Message, bot_response: Message)

Notify the EditTracker that the given user message should be associated with the given bot response. Overwrites any previous associated bot response

Examples found in repository?

src/reply/send_reply.rs (line 176)

133
134
135
136
137
138
139
140
141
142
143
144
145
146
147
148
149
150
151
152
153
154
155
156
157
158
159
160
161
162
163
164
165
166
167
168
169
170
171
172
173
174
175
176
177
178
179
180
181
182
183
184
185
186
187
188
189
190
191
192
193
194
195

async fn _send_prefix_reply<'a, U, E>(
    ctx: crate::PrefixContext<'_, U, E>,
    mut reply: crate::CreateReply<'a>,
) -> Result<Box<serenity::Message>, serenity::Error> {
    if let Some(callback) = ctx.framework.options().reply_callback {
        callback(ctx.into(), &mut reply);
    }

    // This must only return None when we _actually_ want to reuse the existing response! There are
    // no checks later
    let lock_edit_tracker = || {
        if ctx.command.reuse_response {
            if let Some(edit_tracker) = &ctx.framework.options().prefix_options.edit_tracker {
                return Some(edit_tracker.write().unwrap());
            }
        }
        None
    };

    let existing_response = lock_edit_tracker()
        .as_mut()
        .and_then(|t| t.find_bot_response(ctx.msg.id))
        .cloned();

    Ok(Box::new(if let Some(mut response) = existing_response {
        response
            .edit(ctx.serenity_context, |f| {
                // Reset the message. We don't want leftovers of the previous message (e.g. user
                // sends a message with `.content("abc")` in a track_edits command, and the edited
                // message happens to contain embeds, we don't want to keep those embeds)
                // (*f = Default::default() won't do)
                f.content("");
                f.set_embeds(Vec::new());
                f.components(|b| b);
                f.0.insert("attachments", serenity::json::json! { [] });

                reply.to_prefix_edit(f);
                f
            })
            .await?;

        // If the entry still exists after the await, update it to the new contents
        if let Some(mut edit_tracker) = lock_edit_tracker() {
            edit_tracker.set_bot_response(ctx.msg, response.clone());
        }

        response
    } else {
        let new_response = ctx
            .msg
            .channel_id
            .send_message(ctx.serenity_context, |m| {
                reply.to_prefix(m, ctx.msg);
                m
            })
            .await?;
        if let Some(track_edits) = &mut lock_edit_tracker() {
            track_edits.set_bot_response(ctx.msg, new_response.clone());
        }

        new_response
    }))
}

pub fn track_command(&mut self, user_msg: &Message)

Store that this command is currently running; so that if the command is editing its own invocation message (e.g. removing embeds), we don’t accidentally treat it as an execute_untracked_edits situation and start an infinite loop

Examples found in repository?

src/dispatch/prefix.rs (line 311)

279
280
281
282
283
284
285
286
287
288
289
290
291
292
293
294
295
296
297
298
299
300
301
302
303
304
305
306
307
308
309
310
311
312
313
314
315
316
317
318
319
320

pub async fn run_invocation<U, E>(
    ctx: crate::PrefixContext<'_, U, E>,
) -> Result<(), crate::FrameworkError<'_, U, E>> {
    // Check if we should disregard this invocation if it was triggered by an edit
    if ctx.trigger == crate::MessageDispatchTrigger::MessageEdit && !ctx.command.invoke_on_edit {
        return Ok(());
    }
    if ctx.trigger == crate::MessageDispatchTrigger::MessageEditFromInvalid
        && !ctx.framework.options.prefix_options.execute_untracked_edits
    {
        return Ok(());
    }

    super::common::check_permissions_and_cooldown(ctx.into()).await?;

    // Typing is broadcasted as long as this object is alive
    let _typing_broadcaster = if ctx.command.broadcast_typing {
        ctx.msg
            .channel_id
            .start_typing(&ctx.serenity_context.http)
            .ok()
    } else {
        None
    };

    (ctx.framework.options.pre_command)(crate::Context::Prefix(ctx)).await;

    // Store that this command is currently running; so that if the invocation message is being
    // edited before a response message is registered, we don't accidentally treat it as an
    // execute_untracked_edits situation and start an infinite loop
    // Reported by vicky5124 https://discord.com/channels/381880193251409931/381912587505500160/897981367604903966
    if let Some(edit_tracker) = &ctx.framework.options.prefix_options.edit_tracker {
        edit_tracker.write().unwrap().track_command(ctx.msg);
    }

    // Execute command
    (ctx.action)(ctx).await?;

    (ctx.framework.options.post_command)(crate::Context::Prefix(ctx)).await;

    Ok(())
}