Trait poise::PopArgumentAsync

pub trait PopArgumentAsync<'a>: Sized {
    type Err;
    fn async_pop_from<'life0, 'life1, 'life2, 'async_trait>(
        ctx: &'life0 Context, 
        msg: &'life1 Message, 
        args: &'life2 ArgString<'a>
    ) -> Pin<Box<dyn Future<Output = Result<(ArgString<'a>, Self), Self::Err>> + Send + 'async_trait>>
    where
        'a: 'async_trait,
        'life0: 'async_trait,
        'life1: 'async_trait,
        'life2: 'async_trait,
        Self: 'async_trait;
}

Parse a value out of a string by popping off the front of the string. Discord message context is available for parsing, and IO may be done as part of the parsing.

Implementors should assume that a string never starts with whitespace, and fail to parse if it does. This is for consistency’s sake and also because it keeps open the possibility of parsing whitespace.

Associated Types

type Err

This error type should implement std::error::Error most of the time

Required methods

fn async_pop_from<'life0, 'life1, 'life2, 'async_trait>(
ctx: &'life0 Context,
msg: &'life1 Message,
args: &'life2 ArgString<'a>
) -> Pin<Box<dyn Future<Output = Result<(ArgString<'a>, Self), Self::Err>> + Send + 'async_trait>> where
'a: 'async_trait,
'life0: 'async_trait,
'life1: 'async_trait,
'life2: 'async_trait,
Self: 'async_trait,

Parse Self from the front of the given string and return a tuple of the remaining string and Self.