Trait poise::SlashArgument

pub trait SlashArgument: Sized {
    fn extract<'life0, 'life1, 'async_trait>(
        ctx: &'life0 Context, 
        guild: Option<GuildId>, 
        channel: Option<ChannelId>, 
        value: &'life1 Value
    ) -> Pin<Box<dyn Future<Output = Result<Self, SlashArgError>> + Send + 'async_trait>>
    where
        'life0: 'async_trait,
        'life1: 'async_trait,
        Self: 'async_trait;

    fn create(
        builder: &mut CreateApplicationCommandOption
    ) -> &mut CreateApplicationCommandOption;
}

Implement this trait on types that you want to use as a slash command parameter.

Required methods

fn extract<'life0, 'life1, 'async_trait>(
ctx: &'life0 Context,
guild: Option<GuildId>,
channel: Option<ChannelId>,
value: &'life1 Value
) -> Pin<Box<dyn Future<Output = Result<Self, SlashArgError>> + Send + 'async_trait>> where
'life0: 'async_trait,
'life1: 'async_trait,
Self: 'async_trait,

Extract a Rust value of type T from the slash command argument, given via a serde_json::Value.

fn create(
builder: &mut CreateApplicationCommandOption
) -> &mut CreateApplicationCommandOption

Create a slash command parameter equivalent to type T.

Only fields about the argument type are filled in. The caller is still responsible for filling in name(), description(), and possibly required() or other fields.