Crate poise

Poise is an opinionated Discord bot framework with a few distinctive features:

edit tracking: when user edits their message, automatically update bot response

slash commands: completely define both normal and slash commands with a single function

flexible argument parsing: command parameters are defined with normal Rust types and parsed automatically

I initially this framework mainly for personal use (rustbot and etternabot). Features are added on demand, since it’s easier to draft a good design when you know exactly which practical needs it should cover.

Warning: API details are subject to change

Quickstart

use poise::serenity_prelude as serenity;

struct Data {}
type Error = Box<dyn std::error::Error + Send + Sync>;
type Context<'a> = poise::Context<'a, Data, Error>;

/// Display your or another user's account creation date
#[poise::command(prefix_command, slash_command, track_edits)]
async fn age(
    ctx: Context<'_>,
    #[description = "Selected user"] user: Option<serenity::User>,
) -> Result<(), Error> {
    let user = user.as_ref().unwrap_or(ctx.author());
    ctx.say(format!("{}'s account was created at {}", user.name, user.created_at())).await?;

    Ok(())
}

#[tokio::main]
async fn main() {
    poise::Framework::build()
        .token(std::env::var("DISCORD_BOT_TOKEN").unwrap())
        .user_data_setup(move |_ctx, _ready, _framework| Box::pin(async move {
            Ok(Data {})
        }))
        .options(poise::FrameworkOptions {
            // configure framework here
            prefix_options: poise::PrefixFrameworkOptions {
                prefix: Some("~".into()),
                ..Default::default()
            },
            commands: vec![age()],
            ..Default::default()
        })
        .run().await.unwrap();
}

A full functioning bot would contain a help command as well as a register command to register slash commands. See [examples/framework_usage] for examples on that as well as other features of poise.

You can run the framework_usage example with cargo run --example=framework_usage --features collector

Introduction to slash commands

Discord slash commands can be a bit unintuitive at first. If you’re unfamiliar, please read this

To activate a slash command, your bot needs to register in on Discord. You may want to do this manually, with a register command (poise provides builtins::register_application_commands as a starting point for that), or you may want to re-register commands automatically on every bot startup. Choose what you prefer

Commands can be registered globally or per guild. Global commands are available on every guild your bot is invited on, but it takes up to an hour for global registration to roll out. Per guild registration only updates a single guild, but it happens instantly, which is useful for testing.

Your bot also needs to be invited with the applications.commands scope. For example, in Discord’s invite link generator (discord.com/developers/applications/XXX/oauth2/url-generator), tick the applications.commands box.

Crate features

collector: enables serenity’s collector feature

How to use

Create commands

Every command is represented by a function annotated with #[poise::command]:

/// Description of the command here
///
/// Here you can explain how the command \
/// is used and how it works.
#[poise::command(prefix_command, /* add more optional command settings here, like slash_command */)]
async fn command_name(
    ctx: Context<'_>,
    #[description = "Description of arg1 here"] arg1: serenity::Member,
    #[description = "Description of arg2 here"] arg2: Option<u32>,
) -> Result<(), Error> {
    // Command code here

    Ok(())
}

See #[poise::command] for detailed information.

Big example to showcase many command features

use poise::serenity_prelude as serenity;
type Data = ();
type Error = Box<dyn std::error::Error + Send + Sync>;
type Context<'a> = poise::Context<'a, Data, Error>;

/// A test command for poise
#[poise::command(
    prefix_command,
    track_edits,
    hide_in_help,
    required_permissions = "SEND_MESSAGES",
    aliases("bigounce", "abomination"),
    explanation_fn = "my_huge_ass_command_help",
    check = "check",
    on_error = "error_handler",
)]
async fn my_huge_ass_command(
    ctx: Context<'_>,
    #[description = "Consectetur"] ip_addr: std::net::IpAddr, // implements FromStr
    #[description = "Amet"] user: serenity::Member, // implements ArgumentConvert
    #[description = "Sit"] code_block: poise::CodeBlock, // implements PopArgument
    #[description = "Dolor"] #[flag] my_flag: bool,
    #[description = "Ipsum"] #[lazy] always_none: Option<String>,
    #[description = "Lorem"] #[rest] rest: String,
) -> Result<(), Error> {
    Ok(())
}

fn my_huge_ass_command_help() -> String {
    String::from("\
Example usage:
~my_huge_ass_command 127.0.0.1 @kangalioo `i = i + 1` my_flag rest of the message")
}

async fn check(ctx: Context<'_>) -> Result<bool, Error> {
    // We discriminate against users starting with an X
    Ok(!ctx.author().name.starts_with('X'))
}

async fn error_handler(error: poise::FrameworkError<'_, Data, Error>) {
    println!("Oh noes, we got an error: {:?}", error);
}

Create and configure framework

// Use `Framework::build()` to create a framework builder and supply basic data to the framework:
poise::Framework::build()
    .token("...")
    .user_data_setup(|_, _, _| Box::pin(async move {
        // construct user data here (invoked when bot connects to Discord)
        Ok(())
    }))

    // Most configuration is done via the `FrameworkOptions` struct, which you can define with
    // a struct literal (hint: use `..Default::default()` to fill uninitialized
    // settings with their default value):
    .options(poise::FrameworkOptions {
        on_error: |err| Box::pin(my_error_function(err)),
        prefix_options: poise::PrefixFrameworkOptions {
            prefix: Some("~".into()),
            edit_tracker: Some(poise::EditTracker::for_timespan(std::time::Duration::from_secs(3600))),
            case_insensitive_commands: true,
            ..Default::default()
        },
        // This is also where commands go
        commands: vec![
            command1(),
            command2(),
            // To add subcommands, modify the `subcommands` field of the `Command` struct returned
            // by the command functions
            poise::Command {
                subcommands: vec![command3_1(), command3_2()],
                ..command3()
            }
        ],
        ..Default::default()
    })

    .run().await?;

Tips and tricks

Type aliases

As seen in the examples, it’s useful to define type aliases for Context with your bot’s error type and user data type filled in:

type Context<'a> = poise::Context<'a, UserData, ErrorType>;

Serenity prelude

When you’re too lazy to import serenity items from their full path which can be quite lengthy at times, you can use poise::serenity_prelude: a module which reexports almost all items from serenity.

use poise::serenity_prelude as serenity;

// Short paths!
serenity::Member, serenity::UserId, serenity::ReactionType, serenity::GatewayIntents

About the weird name

I’m bad at names. Google lists “poise” as a synonym to “serenity” which is the Discord library underlying this framework, so that’s what I chose.

Re-exports

pub use async_trait::async_trait;

pub use futures_core;

pub use futures_util;

pub use serenity;

Modules

builtins

Building blocks for common commands like help commands or application command registration

samplesDeprecated

See builtins

serenity_prelude

This module re-exports a bunch of items from all over serenity. Useful if you can’t remember the full paths of serenity items.

Macros

autocomplete_argument_into_json

Full version of crate::Autocompletable::into_json.

create_slash_argument

Full version of crate::SlashArgument::create.

extract_autocomplete_argument

Full version of crate::Autocompletable::extract_partial.

extract_slash_argument

Full version of crate::SlashArgument::extract.

parse_prefix_args

Macro for parsing an argument string into multiple parameter types.

parse_slash_args

Macro for extracting and parsing slash command arguments out of an array of serenity::ApplicationCommandInteractionDataOption.

pop_prefix_argument

Full version of crate::PopArgument::pop_from.

Structs

ApplicationContext

Application command specific context passed to command invocations.

AutocompleteChoice

A single autocomplete choice, displayed in Discord UI

CodeBlock

A command parameter type for Discord code blocks

CodeBlockError

Error thrown when parsing a malformed CodeBlock (CodeBlock::pop_from)

Command

Type returned from #[poise::command] annotated functions, which contains all of the generated prefix and application commands

CommandParameter

A single parameter of a crate::Command

CooldownConfig

Configuration struct for Cooldowns

Cooldowns

Handles cooldowns for a single command

CreateReply

Message builder that abstracts over prefix and application command responses

EditTracker

Stores messages and the associated bot responses in order to implement poise’s edit tracking feature.

EventWrapper

A serenity::EventHandler implementation that wraps every received event into the Event enum and propagates it to a callback.

Framework

The main framework struct which stores all data and handles message and interaction dispatch.

FrameworkBuilder

A builder to configure and run a framework.

FrameworkOptions

Framework configuration

InvalidBool

Error thrown when the user enters a string that is not recognized as a boolean

InvalidChoice

Error thrown when the user enters a string that is not recognized by a ChoiceParameter-derived enum

KeyValueArgs

A command parameter type for key-value args

MissingAttachment

Error thrown in prefix invocation when there’s too few attachments

PartialContext

Trimmed down, more general version of Context

PrefixContext

Prefix-specific context passed to command invocations.

PrefixFrameworkOptions

Prefix-specific framework configuration

TooFewArguments

Error thrown if user passes too few arguments to a command

TooManyArguments

Error thrown if user passes too many arguments to a command

Enums

ApplicationCommandOrAutocompleteInteraction

Abstracts over a refernce to an application command interaction or autocomplete interaction

Context

Wrapper around either crate::ApplicationContext or crate::PrefixContext

ContextMenuCommandAction

Possible actions that a context menu entry can have

Event

This enum stores every possible event that a serenity::EventHandler can receive.

FrameworkError

Any error that can occur while the bot runs. Either thrown by user code (those variants will have an error field with your error type E in it), or originating from within the framework.

Prefix

Possible ways to define a command prefix

ReplyHandle

Returned from send_reply to retrieve the sent message object.

SlashArgError

Possible errors when parsing slash command arguments

Traits

Autocompletable

Types that can be marked autocompletable in a slash command parameter.

ContextMenuParameter

Implemented for all types that can be used in a context menu command

PopArgument

Parse a value out of a string by popping off the front of the string. Discord message context is available for parsing, and IO may be done as part of the parsing.

SlashArgument

Implement this trait on types that you want to use as a slash command parameter.

Functions

dispatch_message

Manually dispatches a message with the prefix framework.

find_command

Find a command or subcommand within &[Command], given a command invocation without a prefix. Returns the verbatim command name string as well as the command arguments (i.e. the remaining string).

say_reply

Shorthand of send_reply for text-only messages

send_application_reply

Send a response to an interaction (slash command or context menu command invocation).

send_prefix_reply

Prefix-specific reply function. For more details, see crate::send_reply.

send_reply

Send a message in the given context: normal message if prefix command, interaction response if application command.

Type Definitions

BoxFuture

Shorthand for a wrapped async future with a lifetime, used by many parts of this framework.

Attribute Macros

command

This macro transforms plain functions into poise bot commands.

Derive Macros

ChoiceParameter

Use this derive macro on an enum to easily generate a choice parameter type. A choice parameter is mainly useful in slash commands. It allows you to constrain input to a fixed set of choices.

SlashChoiceParameter

See ChoiceParameter