Module poise::builtins

Building blocks for common commands like help commands or application command registration

This file provides sample commands and utility functions like help menus or error handlers to use as a starting point for the framework.

Structs

HelpConfiguration

Optional configuration for how the help message from help() looks

Functions

autocomplete_command

An autocomplete function that can be used for the command parameter in your help function.

create_application_commands

Collects all commands into a serenity::CreateApplicationCommands builder, which can be used to register the commands on Discord

help

A help command that outputs text in a code block, groups commands by categories, and annotates commands with a slash if they exist as slash commands.

on_error

An error handler that prints the error into the console and also into the Discord chat. If the user invoked the command wrong (crate::FrameworkError::ArgumentParse), the command help is displayed and the user is directed to the help menu.

register_application_commands

Wraps create_application_commands and does some sane default permission checks, as well as actually send the registration HTTP request to Discord

servers

Lists servers of which the bot is a member of, including their member counts, sorted descendingly by member count.