Module poise::builtins

Building blocks for common commands like help commands or application command registration

This file provides sample commands and utility functions like help menus or error handlers to use as a starting point for the framework.

Enums

HelpResponseMode

Whether to send the help message as ephemeral (if possible), or as a normal message

Functions

autocomplete_command

An autocomplete function that can be used for the command parameter in your help function.

help

A help command that outputs text in a code block, groups commands by categories, and annotates commands with a slash if they exist as slash commands.

on_error

An error handler that prints the error into the console and also into the Discord chat. If the user invoked the command wrong (i.e. an crate::ArgumentParseError), the command help is displayed and the user is directed to the help menu.

register_application_commands

Generic function to register application commands. Only allows server owners to invoke.

servers

Lists servers of which the bot is a member of, including their member counts, sorted descendingly by member count.