Module poise::builtins

Building blocks for common commands like help commands or application command registration

This file provides sample commands and utility functions like help menus or error handlers to use as a starting point for the framework.

Structs

HelpConfiguration

Optional configuration for how the help message from help() looks

PrettyHelpConfiguration

Optional configuration for how the help message from pretty_help() looks

Functions

autocomplete_command

An autocomplete function that can be used for the command parameter in your help function.

create_application_commands

Collects all commands into a Vec<serenity::CreateCommand> builder, which can be used to register the commands on Discord

help

A help command that outputs text in a code block, groups commands by categories, and annotates commands with a slash if they exist as slash commands.

on_error

An error handler that logs errors either via the tracing crate or via a Discord message. Set up a logger (e.g. env_logger::init()) or a tracing subscriber (e.g. tracing_subscriber::fmt::init()) to see the logged errors from this method.

paginatechrono or time

This is an example implementation of pagination. To tweak the behavior, copy the source code and adjust to your needs:

pretty_help

A help command that works similarly to builtin::help butt outputs text in an embed.

register_application_commands

Note: you probably want register_application_commands_buttons instead; it’s easier and more powerful

register_application_commands_buttons

Spawns four buttons to register or delete application commands globally or in the current guild

register_globally

Registers the given list of application commands to Discord as global commands.

register_in_guild

Registers the given list of application commands to Discord as guild-specific commands.

serverscache

Lists servers of which the bot is a member of, including their member counts, sorted descendingly by member count.