Function poise::builtins::help

pub async fn help<U, E>(
    ctx: Context<'_, U, E>,
    command: Option<&str>,
    config: HelpConfiguration<'_>
) -> Result<(), Error>

A help command that outputs text in a code block, groups commands by categories, and annotates commands with a slash if they exist as slash commands.

Example usage from Ferris, the Discord bot running in the Rust community server:

/// Show this menu
#[poise::command(prefix_command, track_edits, slash_command)]
pub async fn help(
    ctx: Context<'_>,
    #[description = "Specific command to show help about"] command: Option<String>,
) -> Result<(), Error> {
    let config = poise::builtins::HelpConfiguration {
        extra_text_at_bottom: "\
Type ?help command for more info on a command.
You can edit your message to the bot and the bot will edit its response.",
        ..Default::default()
    };
    poise::builtins::help(ctx, command.as_deref(), config).await?;
    Ok(())
}

Output:

Playground:
  ?play        Compile and run Rust code in a playground
  ?eval        Evaluate a single Rust expression
  ?miri        Run code and detect undefined behavior using Miri
  ?expand      Expand macros to their raw desugared form
  ?clippy      Catch common mistakes using the Clippy linter
  ?fmt         Format code using rustfmt
  ?microbench  Benchmark small snippets of code
  ?procmacro   Compile and use a procedural macro
  ?godbolt     View assembly using Godbolt
  ?mca         Run performance analysis using llvm-mca
  ?llvmir      View LLVM IR using Godbolt
Crates:
  /crate       Lookup crates on crates.io
  /doc         Lookup documentation
Moderation:
  /cleanup     Deletes the bot's messages for cleanup
  /ban         Bans another person
  ?move        Move a discussion to another channel
  /rustify     Adds the Rustacean role to members
Miscellaneous:
  ?go          Evaluates Go code
  /source      Links to the bot GitHub repo
  /help        Show this menu

Type ?help command for more info on a command.
You can edit your message to the bot and the bot will edit its response.

Examples found in repository?

examples/serenity_example_port/main.rs (line 118)

↕

106
107
108
109
110
111
112
113
114
115
116
117
118
119
120
121

async fn help(
    ctx: Context<'_>,
    #[description = "Command to display specific information about"] command: Option<String>,
) -> Result<(), Error> {
    let config = poise::builtins::HelpConfiguration {
        extra_text_at_bottom: "\
Hello! こんにちは！Hola! Bonjour! 您好! 안녕하세요~

If you want more information about a specific command, just pass the command as argument.",
        ..Default::default()
    };

    poise::builtins::help(ctx, command.as_deref(), config).await?;

    Ok(())
}

Hide additional examples

examples/framework_usage/main.rs (lines 26-35)

↕

20
21
22
23
24
25
26
27
28
29
30
31
32
33
34
35
36
37
38

async fn help(
    ctx: Context<'_>,
    #[description = "Specific command to show help about"]
    #[autocomplete = "poise::builtins::autocomplete_command"]
    command: Option<String>,
) -> Result<(), Error> {
    poise::builtins::help(
        ctx,
        command.as_deref(),
        poise::builtins::HelpConfiguration {
            extra_text_at_bottom: "\
This is an example bot made to showcase features of my custom Discord bot framework",
            show_context_menu_commands: true,
            ..Default::default()
        },
    )
    .await?;
    Ok(())
}