Function poise::builtins::on_error

pub async fn on_error<U: Debug, E: Display + Debug>(
    error: FrameworkError<'_, U, E>
) -> Result<(), Error>

An error handler that prints the error into the console and also into the Discord chat. If the user invoked the command wrong (i.e. an crate::ArgumentParseError), the command help is displayed and the user is directed to the help menu.

Can return an error if sending the Discord error message failed. You can decide for yourself how to handle this, for example:

if let Err(e) = poise::builtins::on_error(error).await {
    println!("Fatal error while sending error message: {}", e);
}

Examples found in repository

examples/framework_usage/main.rs (line 60)

↕

50
51
52
53
54
55
56
57
58
59
60
61
62
63
64
65

async fn on_error(error: poise::FrameworkError<'_, Data, Error>) {
    // This is our custom error handler
    // They are many errors that can occur, so we only handle the ones we want to customize
    // and forward the rest to the default handler
    match error {
        poise::FrameworkError::Setup { error } => panic!("Failed to start bot: {:?}", error),
        poise::FrameworkError::Command { error, ctx } => {
            println!("Error in command `{}`: {:?}", ctx.command().name, error,);
        }
        error => {
            if let Err(e) = poise::builtins::on_error(error).await {
                println!("Error while handling error: {}", e)
            }
        }
    }
}

examples/serenity_example_port/main.rs (line 178)

↕

161
162
163
164
165
166
167
168
169
170
171
172
173
174
175
176
177
178
179
180
181
182
183

async fn on_error(error: poise::FrameworkError<'_, Data, Error>) {
    match error {
        poise::FrameworkError::Command { error, ctx } => {
            println!(
                "Command '{}' returned error {:?}",
                ctx.command().name,
                error
            );
        }
        poise::FrameworkError::Listener { error, event } => {
            println!(
                "Listener returned error during {} event: {:?}",
                event.name(),
                error
            );
        }
        error => {
            if let Err(e) = poise::builtins::on_error(error).await {
                println!("Error while handling error: {}", e)
            }
        }
    }
}