Function poise::builtins::on_error

pub async fn on_error<U, E: Display + Debug>(
    error: FrameworkError<'_, U, E>
) -> Result<(), Error>

An error handler that logs errors either via the log crate or via a Discord message. Set up a logger (e.g. env_logger::init()) to see the logged errors from this method.

If the user invoked the command wrong (crate::FrameworkError::ArgumentParse), the command help is displayed and the user is directed to the help menu.

Can return an error if sending the Discord error message failed. You can decide for yourself how to handle this, for example:

if let Err(e) = poise::builtins::on_error(error).await {
    log::error!("Fatal error while sending error message: {}", e);
}

Examples found in repository?

src/structs/framework_options.rs (line 96)

90
91
92
93
94
95
96
97
98
99
100
101
102
103
104
105
106
107
108
109
110
111
112
113
114
115
116
117
118
119
120

    fn default() -> Self {
        #[allow(deprecated)] // we need to set the listener field
        Self {
            commands: Vec::new(),
            on_error: |error| {
                Box::pin(async move {
                    if let Err(e) = crate::builtins::on_error(error).await {
                        log::error!("Error while handling error: {}", e);
                    }
                })
            },
            event_handler: |_, _, _, _| Box::pin(async { Ok(()) }),
            listener: (),
            pre_command: |_| Box::pin(async {}),
            post_command: |_| Box::pin(async {}),
            command_check: None,
            skip_checks_for_owners: false,
            allowed_mentions: Some({
                let mut f = serenity::CreateAllowedMentions::default();
                // Only support direct user pings by default
                f.empty_parse().parse(serenity::ParseValue::Users);
                f
            }),
            reply_callback: None,
            manual_cooldowns: false,
            require_cache_for_guild_check: false,
            prefix_options: Default::default(),
            owners: Default::default(),
            __non_exhaustive: (),
        }
    }