Function poise::builtins::on_error

pub async fn on_error<D>(
    e: Box<dyn Error + Send + Sync>, 
    ctx: ErrorContext<'_, D, Box<dyn Error + Send + Sync>>
)

An error handler that prints the error into the console and also into the Discord chat. If the user invoked the command wrong (i.e. an crate::ArgumentParseError), the command help is displayed and the user is directed to the help menu.