Function poise::builtins::on_error

pub async fn on_error<U, E: Display + Debug>(
    error: FrameworkError<'_, U, E>
) -> Result<(), Error>

An error handler that logs errors either via the tracing crate or via a Discord message. Set up a logger (e.g. env_logger::init()) or a tracing subscriber (e.g. tracing_subscriber::fmt::init()) to see the logged errors from this method.

If the user invoked the command wrong (crate::FrameworkError::ArgumentParse), the command help is displayed and the user is directed to the help menu.

Can return an error if sending the Discord error message failed. You can decide for yourself how to handle this, for example:

if let Err(e) = poise::builtins::on_error(error).await {
    tracing::error!("Fatal error while sending error message: {}", e);
}