Function poise::builtins::register_application_commands

pub async fn register_application_commands<U, E>(
    ctx: Context<'_, U, E>, 
    global: bool
) -> Result<(), Error>

Wraps create_application_commands and does some sane default permission checks, as well as actually send the registration HTTP request to Discord

Permission checks:

global command registration is only allowed for bot owners

guild-specific command registration is allowed for guild owners and bot owners

This function is supposed to be ready-to-use implementation for a ~register command of your bot. So if you want, you can copy paste this help message for the command:

Registers application commands in this guild or globally

Run with no arguments to register in guild, run with argument "global" to register globally.

Examples found in repository?

examples/framework_usage/main.rs (line 45)

44
45
46
47
48

async fn register(ctx: Context<'_>, #[flag] global: bool) -> Result<(), Error> {
    poise::builtins::register_application_commands(ctx, global).await?;

    Ok(())
}

Hide additional examples

examples/serenity_example_port/main.rs (line 128)

127
128
129
130
131

async fn register(ctx: Context<'_>, #[flag] global: bool) -> Result<(), Error> {
    poise::builtins::register_application_commands(ctx, global).await?;

    Ok(())
}