Function poise::builtins::register_application_commands

pub async fn register_application_commands<U, E>(
    ctx: Context<'_, U, E>,
    global: bool
) -> Result<(), Error>

Note: you probably want register_application_commands_buttons instead; it’s easier and more powerful

Wraps create_application_commands and adds a bot owner permission check and status messages.

This function is supposed to be a ready-to-use implementation for a ~register command of your bot. So if you want, you can copy paste this help message for the command:

Registers application commands in this guild or globally

Run with no arguments to register in guild, run with argument "global" to register globally.