Function poise::builtins::servers

pub async fn servers<U, E>(ctx: Context<'_, U, E>) -> Result<(), Error>

Available on crate feature cache only.

Lists servers of which the bot is a member of, including their member counts, sorted descendingly by member count.

Non-public guilds are hidden to preserve privacy. When the command is invoked by the bot owner as an application command, the response will be made ephemeral and private guilds are unhidden.

Example:

I am currently in three servers!

A public server (7123 members)

[3 private servers with 456 members total]

Examples found in repository?

examples/framework_usage/commands/general.rs (line 326)

325
326
327
328

pub async fn servers(ctx: Context<'_>) -> Result<(), Error> {
    poise::builtins::servers(ctx).await?;
    Ok(())
}