poise
Module dispatch

Contains all code to dispatch incoming events onto framework commands

Structs

FrameworkContext

A view into data stored by crate::Framework

Functions

check_permissions_and_cooldown

Checks if the invoker is allowed to execute this command at this point in time

dispatch_autocomplete

Dispatches this interaction onto framework commands, i.e. runs the associated autocomplete callback

dispatch_event

Central event handling function of this library

dispatch_interaction

Dispatches this interaction onto framework commands, i.e. runs the associated command

dispatch_message

Manually dispatches a message with the prefix framework

extract_command_and_run_checks

Given an interaction, finds the matching framework command and checks if the user is allowed access

find_command

Find a command or subcommand within &[Command], given a command invocation without a prefix. Returns the verbatim command name string as well as the command arguments (i.e. the remaining string).

parse_invocation

Given a Message and some context data, parses prefix, command etc. out of the message and returns the resulting crate::PrefixContext. To run the command, see run_invocation.

run_invocation

Given an existing parsed command invocation from parse_invocation, run it, including all the before and after code like checks and built in filters from edit tracking