Function poise::dispatch::find_command

pub fn find_command<'a, U, E>(
    commands: &'a [Command<U, E>],
    remaining_message: &'a str,
    case_insensitive: bool,
    parent_commands: &mut Vec<&'a Command<U, E>>
) -> Option<(&'a Command<U, E>, &'a str, &'a str)>where
    U: Send + Sync,

Find a command or subcommand within &[Command], given a command invocation without a prefix. Returns the verbatim command name string as well as the command arguments (i.e. the remaining string).

The API must be like this (as opposed to just taking the command name upfront) because of subcommands.

#[poise::command(prefix_command)]
async fn command1(ctx: poise::Context<'_, (), ()>) -> Result<(), ()> { Ok(()) }
#[poise::command(prefix_command, subcommands("command3"))]
async fn command2(ctx: poise::Context<'_, (), ()>) -> Result<(), ()> { Ok(()) }
#[poise::command(prefix_command)]
async fn command3(ctx: poise::Context<'_, (), ()>) -> Result<(), ()> { Ok(()) }
let commands = vec![command1(), command2()];

let mut parent_commands = Vec::new();
assert_eq!(
    poise::find_command(&commands, "command1 my arguments", false, &mut parent_commands),
    Some((&commands[0], "command1", "my arguments")),
);
assert!(parent_commands.is_empty());

parent_commands.clear();
assert_eq!(
    poise::find_command(&commands, "command2 command3 my arguments", false, &mut parent_commands),
    Some((&commands[1].subcommands[0], "command3", "my arguments")),
);
assert_eq!(&parent_commands, &[&commands[1]]);

parent_commands.clear();
assert_eq!(
    poise::find_command(&commands, "CoMmAnD2 cOmMaNd99 my arguments", true, &mut parent_commands),
    Some((&commands[1], "CoMmAnD2", "cOmMaNd99 my arguments")),
);
assert!(parent_commands.is_empty());

Examples found in repository?

src/dispatch/prefix.rs (lines 171-176)

139
140
141
142
143
144
145
146
147
148
149
150
151
152
153
154
155
156
157
158
159
160
161
162
163
164
165
166
167
168
169
170
171
172
173
174
175
176
177
178
179
180
181
182
183
184
185
186
187
188
189
190
191
192
193
194
195
196
197
198
199
200
201
202
203
204
205
206
207
208
209
210
211
212
213
214
215
216
217
218
219
220
221
222
223
224
225
226
227
228
229
230
231
232
233
234
235
236
237
238
239
240
241
242
243
244
245
246
247
248
249
250
251
252
253
254
255
256
257
258
259
260
261
262
263
264
265
266
267
268
269
270
271
272
273
274
275

pub fn find_command<'a, U, E>(
    commands: &'a [crate::Command<U, E>],
    remaining_message: &'a str,
    case_insensitive: bool,
    parent_commands: &mut Vec<&'a crate::Command<U, E>>,
) -> Option<(&'a crate::Command<U, E>, &'a str, &'a str)>
where
    U: Send + Sync,
{
    let string_equal = if case_insensitive {
        |a: &str, b: &str| a.eq_ignore_ascii_case(b)
    } else {
        |a: &str, b: &str| a == b
    };

    let (command_name, remaining_message) = {
        let mut iter = remaining_message.splitn(2, char::is_whitespace);
        (iter.next().unwrap(), iter.next().unwrap_or("").trim_start())
    };

    for command in commands {
        let primary_name_matches = string_equal(&command.name, command_name);
        let alias_matches = command
            .aliases
            .iter()
            .any(|alias| string_equal(alias, command_name));
        if !primary_name_matches && !alias_matches {
            continue;
        }

        parent_commands.push(command);
        return Some(
            find_command(
                &command.subcommands,
                remaining_message,
                case_insensitive,
                parent_commands,
            )
            .unwrap_or_else(|| {
                parent_commands.pop();
                (command, command_name, remaining_message)
            }),
        );
    }

    None
}

/// Manually dispatches a message with the prefix framework
pub async fn dispatch_message<'a, U: Send + Sync, E>(
    framework: crate::FrameworkContext<'a, U, E>,
    ctx: &'a serenity::Context,
    msg: &'a serenity::Message,
    trigger: crate::MessageDispatchTrigger,
    invocation_data: &'a tokio::sync::Mutex<Box<dyn std::any::Any + Send + Sync>>,
    parent_commands: &'a mut Vec<&'a crate::Command<U, E>>,
) -> Result<(), crate::FrameworkError<'a, U, E>> {
    if let Some(ctx) = parse_invocation(
        framework,
        ctx,
        msg,
        trigger,
        invocation_data,
        parent_commands,
    )
    .await?
    {
        run_invocation(ctx).await?;
    }
    Ok(())
}

/// Given a Message and some context data, parses prefix, command etc. out of the message and
/// returns the resulting [`crate::PrefixContext`]. To run the command, see [`run_invocation`].
pub async fn parse_invocation<'a, U: Send + Sync, E>(
    framework: crate::FrameworkContext<'a, U, E>,
    ctx: &'a serenity::Context,
    msg: &'a serenity::Message,
    trigger: crate::MessageDispatchTrigger,
    invocation_data: &'a tokio::sync::Mutex<Box<dyn std::any::Any + Send + Sync>>,
    parent_commands: &'a mut Vec<&'a crate::Command<U, E>>,
) -> Result<Option<crate::PrefixContext<'a, U, E>>, crate::FrameworkError<'a, U, E>> {
    // Check if we're allowed to invoke from bot messages
    if msg.author.bot && framework.options.prefix_options.ignore_bots {
        return Ok(None);
    }

    // Check if we're allowed to execute our own messages
    if framework.bot_id == msg.author.id && !framework.options.prefix_options.execute_self_messages
    {
        return Ok(None);
    }

    // Strip prefix, trim whitespace between prefix and rest, split rest into command name and args
    let (prefix, msg_content) = match strip_prefix(framework, ctx, msg).await {
        Some(x) => x,
        None => return Ok(None),
    };
    let msg_content = msg_content.trim_start();

    let (command, invoked_command_name, args) = find_command(
        &framework.options.commands,
        msg_content,
        framework.options.prefix_options.case_insensitive_commands,
        parent_commands,
    )
    .ok_or(crate::FrameworkError::UnknownCommand {
        ctx,
        msg,
        prefix,
        msg_content,
        framework,
        invocation_data,
        trigger,
    })?;
    let action = match command.prefix_action {
        Some(x) => x,
        // This command doesn't have a prefix implementation
        None => return Ok(None),
    };

    Ok(Some(crate::PrefixContext {
        serenity_context: ctx,
        msg,
        prefix,
        invoked_command_name,
        args,
        framework,
        data: framework.user_data().await,
        parent_commands,
        command,
        invocation_data,
        trigger,
        action,
        __non_exhaustive: (),
    }))
}