poise::dispatch
Function parse_invocation

pub async fn parse_invocation<'a, U: Send + Sync, E>(
    framework: FrameworkContext<'a, U, E>,
    msg: &'a Message,
    trigger: MessageDispatchTrigger,
    invocation_data: &'a Mutex<Box<dyn Any + Send + Sync>>,
    parent_commands: &'a mut Vec<&'a Command<U, E>>,
) -> Result<Option<PrefixContext<'a, U, E>>, FrameworkError<'a, U, E>>

Given a Message and some context data, parses prefix, command etc. out of the message and returns the resulting crate::PrefixContext. To run the command, see run_invocation.

Returns Ok(None) if the message does not look like a command invocation. Returns Err(...) if the message does look like a command invocation, but cannot be fully parsed.