poise::dispatch
Function run_invocation

pub async fn run_invocation<U, E>(
    ctx: PrefixContext<'_, U, E>,
) -> Result<(), FrameworkError<'_, U, E>>

Given an existing parsed command invocation from parse_invocation, run it, including all the before and after code like checks and built in filters from edit tracking