Enum poise::Context

pub enum Context<'a, U, E> {
    Application(ApplicationContext<'a, U, E>),
    Prefix(PrefixContext<'a, U, E>),
}

Wrapper around either crate::ApplicationContext or crate::PrefixContext

Variants

Application(ApplicationContext<'a, U, E>)

Application command context

Tuple Fields of Application

0: ApplicationContext<'a, U, E>

Prefix(PrefixContext<'a, U, E>)

Prefix command context

Tuple Fields of Prefix

0: PrefixContext<'a, U, E>

Implementations

impl<U, E> Context<'_, U, E>

pub fn discord(&self) -> &Context

Return the stored serenity::Context within the underlying context type.

pub fn framework(&self) -> &Framework<U, E>

Return a read-only reference to crate::Framework.

pub fn data(&self) -> &U

Return a reference to your custom user data

pub fn channel_id(&self) -> ChannelId

Return the channel ID of this context

pub fn guild_id(&self) -> Option<GuildId>

Returns the guild ID of this context, if we are inside a guild

pub fn guild(&self) -> Option<Guild>

Return the guild of this context, if we are inside a guild.

Warning: clones the entire Guild instance out of the cache

pub fn created_at(&self) -> DateTime<Utc>

Return the datetime of the invoking message or interaction

pub fn author(&self) -> &User

Get the author of the command message or application command.

pub fn id(&self) -> u64

Return a ID that uniquely identifies this command invocation.