Struct poise::framework::Framework

pub struct Framework<U, E> { /* private fields */ }

The main framework struct which stores all data and handles message and interaction dispatch.

Technically, this is just an optional abstraction over crate::dispatch_event with some additional conveniences built-in:

fills in correct values for crate::Command::qualified_name: set_qualified_names

spawns a background task to periodically clear edit tracker cache

sets up user data on the first Ready event

keeps track of shard manager and bot ID automatically

You can build a bot without Framework: see the manual_dispatch example in the repository

Implementations

impl<U, E> Framework<U, E>

pub fn build() -> FrameworkBuilder<U, E>

👎Deprecated: Please use Framework::builder instead

Create a framework builder to configure, create and run a framework.

For more information, see FrameworkBuilder

pub fn builder() -> FrameworkBuilder<U, E>

Create a framework builder to configure, create and run a framework.

For more information, see FrameworkBuilder

pub async fn new<F>(
client_builder: ClientBuilder,
setup: F,
options: FrameworkOptions<U, E>
) -> Result<Arc<Self>, Error>where
F: Send + Sync + 'static + for<'a> FnOnce(&'a Context, &'a Ready, &'a Self) -> BoxFuture<'a, Result<U, E>>,
U: Send + Sync + 'static,
E: Send + 'static,

Setup a new Framework. For more ergonomic setup, please see FrameworkBuilder

This function is async and returns Result because it already initializes the Discord client.

The user data callback is invoked as soon as the bot is logged in. That way, bot data like user ID or connected guilds can be made available to the user data setup function. The user data setup is not allowed to return Result because there would be no reasonable course of action on error.

Examples found in repository?

src/framework/builder.rs (line 225)

187
188
189
190
191
192
193
194
195
196
197
198
199
200
201
202
203
204
205
206
207
208
209
210
211
212
213
214
215
216
217
218
219
220
221
222
223
224
225
226

    pub async fn build(self) -> Result<std::sync::Arc<crate::Framework<U, E>>, serenity::Error>
    where
        U: Send + Sync + 'static,
        E: Send + 'static,
    {
        // Aggregate required values or panic if not provided
        // WHEN CHANGING THESE, UPDATE FrameworkBuilder docs!
        let token = self.token.expect("No token was provided to the framework");
        let intents = self.intents.expect(
            "

No gateway intents were provided to the framework via `FrameworkBuilder::intents()`. If you're \
unsure, use
`serenity::GatewayIntents::non_privileged() | serenity::GatewayIntents::MESSAGE_CONTENT`
and enable MESSAGE_CONTENT in your Discord bot dashboard

",
        );
        let setup = self
            .setup
            .expect("No user data setup function was provided to the framework");
        let mut options = self.options.expect("No framework options provided");

        // Build framework options by concatenating user-set options with commands and owners
        options.commands.extend(self.commands);
        if self.initialize_owners {
            if let Err(e) = super::insert_owners_from_http(&token, &mut options.owners).await {
                log::warn!("Failed to insert owners from HTTP: {}", e);
            }
        }

        // Create serenity client
        let mut client_builder = serenity::ClientBuilder::new(token, intents);
        if let Some(client_settings) = self.client_settings {
            client_builder = client_settings(client_builder);
        }

        // Create framework with specified settings
        crate::Framework::new(client_builder, setup, options).await
    }

pub async fn start_with<F: Future<Output = Result<()>>>(
self: Arc<Self>,
start: impl FnOnce(Client) -> F
) -> Result<(), Error>where
U: Send + Sync + 'static,
E: Send + 'static,

Small utility function for starting the framework that is agnostic over client sharding

You can use these shortcut methods to start the framework with a single shard or with automatic sharding: Framework::start, Framework::start_autosharded

Examples

Start shards in a range

let shard_range = [3, 7];
let total_shards = 8;

framework
    .start_with(|mut client| async move {
        client.start_shard_range(shard_range, total_shards).await
   })
   .await?;

Examples found in repository?

src/framework/mod.rs (line 175)

170
171
172
173
174
175
176
177
178
179
180
181
182
183
184
185
186
187
188
189
190

    pub async fn start(self: std::sync::Arc<Self>) -> Result<(), serenity::Error>
    where
        U: Send + Sync + 'static,
        E: Send + 'static,
    {
        self.start_with(|mut c| async move { c.start().await })
            .await
    }

    /// Starts the framework with automatic sharding.
    /// Calls [`serenity::Client::start_autosharded`] internally.
    ///
    /// See [`Framework::start_with`] for other sharding configurations.
    pub async fn start_autosharded(self: std::sync::Arc<Self>) -> Result<(), serenity::Error>
    where
        U: Send + Sync + 'static,
        E: Send + 'static,
    {
        self.start_with(|mut c| async move { c.start_autosharded().await })
            .await
    }

pub async fn start(self: Arc<Self>) -> Result<(), Error>where
U: Send + Sync + 'static,
E: Send + 'static,

Starts the framework with a shard. Calls serenity::Client::start internally.

See Framework::start_with for other sharding configurations.

Examples found in repository?

src/framework/builder.rs (line 237)

232
233
234
235
236
237
238

    pub async fn run(self) -> Result<(), serenity::Error>
    where
        U: Send + Sync + 'static,
        E: Send + 'static,
    {
        self.build().await?.start().await
    }

pub async fn start_autosharded(self: Arc<Self>) -> Result<(), Error>where
U: Send + Sync + 'static,
E: Send + 'static,

Starts the framework with automatic sharding. Calls serenity::Client::start_autosharded internally.

See Framework::start_with for other sharding configurations.

Examples found in repository?

src/framework/builder.rs (line 246)

241
242
243
244
245
246
247

    pub async fn run_autosharded(self) -> Result<(), serenity::Error>
    where
        U: Send + Sync + 'static,
        E: Send + 'static,
    {
        self.build().await?.start_autosharded().await
    }

pub fn options(&self) -> &FrameworkOptions<U, E>

Return the stored framework options, including commands.

pub fn shard_manager(&self) -> &Arc<Mutex<ShardManager>>

Returns the serenity’s client shard manager.

pub fn client(&self) -> impl DerefMut<Target = Client> + '_

Returns the serenity client. Panics if the framework has already started!

pub async fn user_data(&self) -> &U

Retrieves user data, or blocks until it has been initialized (once the Ready event has been received).

Examples found in repository?

src/framework/mod.rs (line 255)

224
225
226
227
228
229
230
231
232
233
234
235
236
237
238
239
240
241
242
243
244
245
246
247
248
249
250
251
252
253
254
255
256
257
258
259
260
261
262
263
264
265
266
267

async fn raw_dispatch_event<U, E>(
    framework: &crate::Framework<U, E>,
    ctx: &serenity::Context,
    event: &crate::Event<'_>,
) where
    U: Send + Sync,
{
    if let crate::Event::Ready { data_about_bot } = event {
        let _: Result<_, _> = framework.bot_id.set(data_about_bot.user.id);
        let setup = Option::take(&mut *framework.setup.lock().unwrap());
        if let Some(setup) = setup {
            match setup(ctx, data_about_bot, framework).await {
                Ok(user_data) => {
                    let _: Result<_, _> = framework.user_data.set(user_data);
                }
                Err(error) => {
                    (framework.options.on_error)(crate::FrameworkError::Setup {
                        error,
                        framework,
                        data_about_bot,
                        ctx,
                    })
                    .await
                }
            }
        } else {
            // ignoring duplicate Discord bot ready event
            // (happens regularly when bot is online for long period of time)
        }
    }

    let user_data = framework.user_data().await;
    let bot_id = *framework
        .bot_id
        .get()
        .expect("bot ID not set even though we awaited Ready");
    let framework = crate::FrameworkContext {
        bot_id,
        options: &framework.options,
        user_data,
        shard_manager: &framework.shard_manager,
    };
    crate::dispatch_event(framework, ctx, event).await;
}