Struct poise::framework::Framework

pub struct Framework<U, E> { /* private fields */ }

The main framework struct which stores all data and handles message and interaction dispatch.

Technically, this is just an optional abstraction over crate::dispatch_event with some additional conveniences built-in:

fills in correct values for crate::Command::qualified_name: set_qualified_names

spawns a background task to periodically clear edit tracker cache

sets up user data on the first Ready event

keeps track of shard manager and bot ID automatically

You can build a bot without Framework: see the manual_dispatch example in the repository

Implementations

impl<U, E> Framework<U, E>

pub fn build() -> FrameworkBuilder<U, E>

๐Deprecated: Please use Framework::builder instead

Create a framework builder to configure, create and run a framework.

For more information, see FrameworkBuilder

pub fn builder() -> FrameworkBuilder<U, E>

Create a framework builder to configure, create and run a framework.

For more information, see FrameworkBuilder

pub async fn new<F>( client_builder: ClientBuilder, setup: F, options: FrameworkOptions<U, E> ) -> Result<Arc<Self>, Error>where F: Send + Sync + 'static + for<'a> FnOnce(&'a Context, &'a Ready, &'a Self) -> BoxFuture<'a, Result<U, E>>, U: Send + Sync + 'static, E: Send + 'static,

Setup a new Framework. For more ergonomic setup, please see FrameworkBuilder

This function is async and returns Result because it already initializes the Discord client.

The user data callback is invoked as soon as the bot is logged in. That way, bot data like user ID or connected guilds can be made available to the user data setup function. The user data setup is not allowed to return Result because there would be no reasonable course of action on error.

pub async fn start_with<F: Future<Output = Result<()>>>( self: Arc<Self>, start: impl FnOnce(Client) -> F ) -> Result<(), Error>where U: Send + Sync + 'static, E: Send + 'static,

Small utility function for starting the framework that is agnostic over client sharding

You can use these shortcut methods to start the framework with a single shard or with automatic sharding: Framework::start, Framework::start_autosharded

Examples

Start shards in a range

let shard_range = [3, 7];
let total_shards = 8;

framework
    .start_with(|mut client| async move {
        client.start_shard_range(shard_range, total_shards).await
   })
   .await?;

pub async fn start(self: Arc<Self>) -> Result<(), Error>where U: Send + Sync + 'static, E: Send + 'static,

Starts the framework with a shard. Calls serenity::Client::start internally.

See Framework::start_with for other sharding configurations.

pub async fn start_autosharded(self: Arc<Self>) -> Result<(), Error>where U: Send + Sync + 'static, E: Send + 'static,

Starts the framework with automatic sharding. Calls serenity::Client::start_autosharded internally.

See Framework::start_with for other sharding configurations.

pub fn options(&self) -> &FrameworkOptions<U, E>

Return the stored framework options, including commands.

pub fn shard_manager(&self) -> &Arc<Mutex<ShardManager>>

Returns the serenityโs client shard manager.

pub fn client(&self) -> impl DerefMut<Target = Client> + '_

Returns the serenity client. Panics if the framework has already started!

pub async fn user_data(&self) -> &U

Retrieves user data, or blocks until it has been initialized (once the Ready event has been received).