Struct poise::framework::FrameworkBuilder

pub struct FrameworkBuilder<U, E> { /* private fields */ }

A builder to configure a framework.

If one of the following required values is missing, the builder will panic on start:

Self::setup

Self::options

Implementations

impl<U, E> FrameworkBuilder<U, E>

pub fn setup<F>(self, setup: F) -> Self
where F: Send + Sync + 'static + for<'a> FnOnce(&'a Context, &'a Ready, &'a Framework<U, E>) -> BoxFuture<'a, Result<U, E>>,

Sets the setup callback which also returns the user data struct.

pub fn options(self, options: FrameworkOptions<U, E>) -> Self

Configure framework options

pub fn initialize_owners(self, initialize_owners: bool) -> Self

Whether to add this bot application’s owner and team members to crate::FrameworkOptions::owners automatically

true by default

pub fn build(self) -> Framework<U, E>
where U: Send + Sync + 'static, E: Send + 'static,

Build the framework with the specified configuration.

For more information, see FrameworkBuilder