Function poise::modal::execute_modal

pub async fn execute_modal<U: Send + Sync, E, M: Modal>(
    ctx: ApplicationContext<'_, U, E>,
    defaults: Option<M>,
    timeout: Option<Duration>
) -> Result<Option<M>, Error>

Convenience function for showing the modal and waiting for a response.

If the user doesn’t submit before the timeout expires, None is returned.

Note: a modal must be the first response to a command. You cannot send any messages before, or the modal will fail.

This function:

sends the modal via Modal::create()

waits for the user to submit via serenity::ModalInteractionCollector

acknowledges the submitted data so that Discord closes the pop-up for the user

parses the submitted data via Modal::parse(), wrapping errors in serenity::Error::Other

If you need more specialized behavior, you can copy paste the implementation of this function and adjust to your needs. The code of this function is just a starting point.