Trait poise::prefix_argument::PopArgument

pub trait PopArgument<'a>: Sized {
    // Required method
    fn pop_from<'life0, 'life1, 'async_trait>(
        args: &'a str,
        attachment_index: usize,
        ctx: &'life0 Context,
        msg: &'life1 Message,
    ) -> Pin<Box<dyn Future<Output = Result<(&'a str, usize, Self), (Box<dyn Error + Send + Sync>, Option<String>)>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'a: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait;
}

Parse a value out of a string by popping off the front of the string. Discord message context is available for parsing, and IO may be done as part of the parsing.

Implementors should assume that a string never starts with whitespace, and fail to parse if it does. This is for consistency’s sake and also because it keeps open the possibility of parsing whitespace.

Similar in spirit to std::str::FromStr.

Required Methods

fn pop_from<'life0, 'life1, 'async_trait>( args: &'a str, attachment_index: usize, ctx: &'life0 Context, msg: &'life1 Message, ) -> Pin<Box<dyn Future<Output = Result<(&'a str, usize, Self), (Box<dyn Error + Send + Sync>, Option<String>)>> + Send + 'async_trait>>
where Self: 'async_trait, 'a: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait,

Parse Self from the front of the given string and return a tuple of the remaining string and Self. If parsing failed, an error is returned and, if applicable, the string on which parsing failed.

If parsing fails because the string is empty, use the TooFewArguments type as the error.

Don’t call this method directly! Use crate::pop_prefix_argument!

Object Safety

This trait is not object safe.