Struct poise::reply::CreateReply

pub struct CreateReply { /* private fields */ }

Message builder that abstracts over prefix and application command responses

Implementations

impl CreateReply

pub fn new() -> Self

Creates a blank CreateReply. Equivalent to Self::default.

pub fn content(self, content: impl Into<String>) -> Self

Set the content of the message.

pub fn embed(self, embed: CreateEmbed) -> Self

Adds an embed to the message.

Existing embeds are kept.

pub fn components(self, components: Vec<CreateActionRow>) -> Self

Set components (buttons and select menus) for this message.

Any previously set components will be overwritten.

pub fn attachment(self, attachment: CreateAttachment) -> Self

Add an attachment.

This will not have an effect in a slash command’s initial response!

pub fn ephemeral(self, ephemeral: bool) -> Self

Toggles whether the message is an ephemeral response (only invoking user can see it).

This only has an effect in slash commands!

pub fn allowed_mentions(self, allowed_mentions: CreateAllowedMentions) -> Self

Set the allowed mentions for the message.

See serenity::CreateAllowedMentions for more information.

pub fn reply(self, reply: bool) -> Self

Makes this message an inline reply to another message like serenity::Message::reply (prefix-only, because slash commands are always inline replies anyways).

To disable the ping, set Self::allowed_mentions with serenity::CreateAllowedMentions::replied_user set to false.

impl CreateReply

Methods to create a message builder from any type from this CreateReply. Used by poise internally to actually send a response to Discord

pub fn to_slash_initial_response( self, builder: CreateInteractionResponseMessage, ) -> CreateInteractionResponseMessage

Serialize this response builder to a serenity::CreateInteractionResponseMessage

pub fn to_slash_followup_response( self, builder: CreateInteractionResponseFollowup, ) -> CreateInteractionResponseFollowup

Serialize this response builder to a serenity::CreateInteractionResponseFollowup

pub fn to_slash_initial_response_edit( self, builder: EditInteractionResponse, ) -> EditInteractionResponse

Serialize this response builder to a serenity::EditInteractionResponse

pub fn to_prefix_edit(self, builder: EditMessage) -> EditMessage

Serialize this response builder to a serenity::EditMessage

pub fn to_prefix(self, invocation_message: MessageReference) -> CreateMessage

Serialize this response builder to a serenity::CreateMessage