Module poise::serenity_prelude

This module re-exports a bunch of items from all over serenity. Useful if you can’t remember the full paths of serenity items.

One way to use this prelude module in your project is

use poise::serenity_prelude as serenity;

Re-exports

pub use serenity::async_trait;

pub use serenity::model::application::interaction::Interaction;

pub use serenity::model::application::interaction::InteractionResponseType;

pub use serenity::model::application::interaction::InteractionType;

pub use serenity::model::application::interaction::MessageFlags as InteractionResponseFlags;

pub use serenity::model::application::interaction::MessageInteraction;

pub use serenity::model::channel::MessageFlags;

pub use serenity::builder::*;

builder

pub use serenity::client::bridge::gateway::event::*;

pub use serenity::client::bridge::gateway::*;

gateway

pub use serenity::client::*;

client

pub use serenity::collector::*;

collector

pub use serenity::http::*;

http

pub use serenity::model::application::command::*;

pub use serenity::model::application::component::*;

pub use serenity::model::application::interaction::application_command::*;

pub use serenity::model::application::interaction::autocomplete::*;

pub use serenity::model::application::interaction::message_component::*;

pub use serenity::model::application::interaction::modal::*;

pub use serenity::model::application::interaction::*;

pub use serenity::model::event::*;

pub use serenity::model::prelude::*;

pub use serenity::prelude::*;

pub use serenity::utils::*;

utils

pub use serenity::*;