Module poise::serenity_prelude

This module re-exports a bunch of items from all over serenity. Useful if you can’t remember the full paths of serenity items.

One way to use this prelude module in your project is

use poise::serenity_prelude as serenity;

Modules

audit_log

Audit log types for administrative actions within guilds.

automod

Auto moderation types

builderbuilder

A set of builders used to make using methods on certain structs simpler to use.

cachecache

A cache containing data received from Shards.

clientclient

The Client contains information about a single bot’s token, as well as event handlers. Dispatching events to configured handlers and starting the shards’ connections are handled directly via the client. In addition, the http module and Cache are also automatically handled by the Client module for you.

close_codes

collectorcollector

colours

Colour constants used by Discord for their branding, role colour palette, etc.

constants

A set of constants used by the library.

create_poll

frameworkframework

The framework is a customizable method of separating commands.

futures

Abstractions for asynchronous programming.

gatewaygateway

The gateway module contains the pieces - primarily the Shard - responsible for maintaining a WebSocket connection with Discord.

httphttp

The HTTP module which provides functions for performing requests to endpoints in Discord’s API.

json

This module exports different types for JSON interactions. It encapsulates the differences between serde_json and simd-json to allow ignoring those in the rest of the codebase.

model

Mappings of objects received from the API, with optional helper methods for ease of use.

prelude

A set of exports which can be helpful to use.

token

Utilities to parse and validate Discord tokens.

utilsutils

A set of utilities to help with common use cases that are not required to fully use the library.

Structs

ActionExecution

Gateway event payload sent when a rule is triggered and an action is executed (e.g. message is blocked).

ActionRow

An action row.

Activity

Representation of an activity that a User is performing.

ActivityAssets

The assets for an activity.

ActivityButton

Discord docs.

ActivityData

Activity data of the current user.

ActivityEmoji

Representation of an emoji used in a custom status

ActivityFlags

A set of flags defining what is in an activity’s payload.

ActivityParty

Information about an activity’s party.

ActivitySecrets

Secrets for an activity.

ActivityTimestamps

Timestamps of when a user started and/or is ending their activity.

AddMember

A builder to add parameters when using GuildId::add_member.

AffectedRole

AfkMetadata

AnswerId

An identifier for a Poll Answer.

ApplicationFlags

The flags of the application.

ApplicationId

An identifier for an Application.

Attachment

A file uploaded with a message. Not to be confused with Embeds.

AttachmentId

An identifier for an attachment.

AuditLogEntry

Discord docs.

AuditLogEntryId

An identifier for an audit log entry.

AuditLogs

Discord docs.

AutoModActionExecutionEvent

Requires GatewayIntents::AUTO_MODERATION_EXECUTION.

AutoModRuleCreateEvent

Requires GatewayIntents::AUTO_MODERATION_CONFIGURATION.

AutoModRuleDeleteEvent

Requires GatewayIntents::AUTO_MODERATION_CONFIGURATION.

AutoModRuleUpdateEvent

Requires GatewayIntents::AUTO_MODERATION_CONFIGURATION.

AutocompleteChoice

AutocompleteOption

The focused option for autocomplete interactions return by CommandData::autocomplete.

Ban

A representation of a banning of a user.

BotGateway

A representation of the data retrieved from the bot gateway endpoint.

BulkBanResponse

The response from GuildId::bulk_ban.

Button

A button component.

Cache

A cache containing data received from Shards.

CacheRef

ChannelCreateEvent

Event data for the channel creation event.

ChannelDeleteEvent

Requires GatewayIntents::GUILDS.

ChannelFlags

Describes extra features of the channel.

ChannelId

An identifier for a Channel

ChannelMention

Discord docs.

ChannelPinsUpdateEvent

Requires GatewayIntents::GUILDS or GatewayIntents::DIRECT_MESSAGES.

ChannelUpdateEvent

Requires GatewayIntents::GUILDS.

Clientgateway

The Client is the way to be able to start sending authenticated requests over the REST API, as well as initializing a WebSocket connection through Shards. Refer to the documentation on using sharding for more information.

ClientBuildergateway

A builder implementing IntoFuture building a Client to interact with Discord.

ClientStatus

Information detailing the current active status of a User.

CollectorCallback

Newtype around a callback that will be called on every incoming request. As long as this collector should still receive events, it should return true. Once it returns false, it is removed.

Colour

A utility struct to help with working with the basic representation of a colour. This is particularly useful when working with a Role’s colour, as the API works with an integer value instead of an RGB value.

Command

The base command model that belongs to an application.

CommandData

The command data payload.

CommandDataOption

A set of a parameter and a value from the user.

CommandDataResolved

The resolved data of a command data interaction payload. It contains the objects of CommandDataOptions.

CommandId

An identifier for a slash command.

CommandInteraction

An interaction when a user invokes a slash command.

CommandOption

The parameters for an Command.

CommandOptionChoice

The only valid values a user can pick in an CommandOption.

CommandPermission

The CommandPermission data.

CommandPermissionId

An identifier for a slash command permission Id. Can contain a RoleId or UserId.

CommandPermissions

An Command permission.

CommandPermissionsUpdateEvent

Requires no gateway intents.

CommandVersionId

An identifier for a slash command version Id.

ComponentInteraction

An interaction triggered by a message component.

ComponentInteractionCollector

A ComponentInteractionCollector receives ComponentInteraction’s match the given filters for a set duration.

ComponentInteractionData

A message component interaction data, provided by ComponentInteraction::data

Connection

Information about a connection between the current user and a third party service.

Content

Describes formatting on string content

ContentSafeOptions

Struct that allows to alter content_safe’s behaviour.

Context

The context is a general utility struct provided on event dispatches, which helps with dealing with the current “context” of the event dispatch. The context also acts as a general high-level interface over the associated Shard which received the event, or the low-level http module.

CountDetails

A representation of reaction count details.

CreateAllowedMentions

A builder to manage the allowed mentions on a message, used by the ChannelId::send_message and ChannelId::edit_message methods.

CreateAttachment

Enum that allows a user to pass a Path or a File type to send_files

CreateAutocompleteResponse

Discord docs

CreateBotAuthParameters

A builder for constructing an invite link with custom OAuth2 scopes.

CreateButton

A builder for creating a button component in a message

CreateChannel

A builder for creating a new GuildChannel in a Guild.

CreateCommand

A builder for creating a new Command.

CreateCommandOption

A builder for creating a new CommandOption.

CreateCommandPermission

A builder for creating an CommandPermission.

CreateEmbed

A builder to create an embed in a message

CreateEmbedAuthor

A builder to create the author data of an emebd. See CreateEmbed::author

CreateEmbedFooter

A builder to create the footer data for an embed. See CreateEmbed::footer

CreateForumPost

Discord docs.

CreateForumTag

Discord docs

CreateGuildWelcomeChannel

A builder for creating a GuildWelcomeChannel.

CreateInputText

A builder for creating an input text component in a modal

CreateInteractionResponseFollowup

Discord docs

CreateInteractionResponseMessage

Discord docs.

CreateInvite

A builder to create a RichInvite for use via GuildChannel::create_invite.

CreateMessage

A builder to specify the contents of an send message request, primarily meant for use through ChannelId::send_message.

CreateModal

Discord docs.

CreatePoll

CreatePollAnswer

CreateQuickModalcollector

Convenience builder to create a modal, wait for the user to submit and parse the response.

CreateScheduledEvent

Discord docs

CreateSelectMenu

A builder for creating a select menu component in a message

CreateSelectMenuOption

A builder for creating an option of a select menu component in a message

CreateStageInstance

Builder for creating a stage instance

CreateSticker

A builder to create a guild sticker

CreateThread

Discord docs:

CreateWebhook

Discord docs

CurrentApplicationInfo

Information about the current application and its owner.

CurrentUser

Information about the current user.

CustomMessage

A builder for constructing a personal Message instance. This can be useful for emitting a manual dispatch to the framework, but you don’t have a message in hand, or just have a fragment of its data.

DiscordJsonError

DiscordJsonSingleError

EditAttachments

You can add new attachments and edit existing ones using this builder.

EditAutoModRule

A builder for creating or editing guild AutoMod rules.

EditChannel

A builder to edit a GuildChannel for use via GuildChannel::edit.

EditCommandPermissions

A builder for creating several CommandPermission.

EditGuild

A builder to optionally edit certain fields of a guild

EditGuildWelcomeScreen

A builder to edit the welcome screen of a guild

EditGuildWidget

A builder to specify the fields to edit in a GuildWidget.

EditInteractionResponse

Discord docs

EditMember

A builder which edits the properties of a Member, to be used in conjunction with Member::edit.

EditMessage

A builder to specify the fields to edit in an existing message.

EditProfile

A builder to edit the current user’s settings, to be used in conjunction with CurrentUser::edit.

EditRole

A builder to create or edit a Role for use via a number of model methods.

EditScheduledEvent

Discord docs

EditStageInstance

Edits a StageInstance.

EditSticker

A builder to create or edit a Sticker for use via a number of model methods.

EditThread

Discord docs.

EditVoiceState

A builder which edits a user’s voice state, to be used in conjunction with GuildChannel::edit_voice_state.

EditWebhook

Discord docs

EditWebhookMessage

A builder to specify the fields to edit in an existing Webhook’s message.

Embed

Represents a rich embed which allows using richer markdown, multiple fields and more. This was heavily inspired by slack’s attachments.

EmbedAuthor

An author object in an embed.

EmbedField

A field object in an embed.

EmbedFooter

Footer information for an embed.

EmbedImage

An image object in an embed.

EmbedProvider

The provider of an embed.

EmbedThumbnail

The dimensions and URL of an embed thumbnail.

EmbedVideo

Video information for an embed.

Emoji

Represents a custom guild emoji, which can either be created using the API, or via an integration. Emojis created using the API only work within the guild it was created in.

EmojiId

An identifier for an Emoji

EmojiIdentifier

A version of an emoji used only when solely the animated state, Id, and name are known.

EmojiIdentifierParseErrormodel and utils

Entitlement

Represents that a user or guild has access to a premium offering in the application.

EntitlementCreateEvent

Requires no gateway intents.

EntitlementDeleteEvent

Requires no gateway intents.

EntitlementId

An identifier for an entitlement.

EntitlementUpdateEvent

Requires no gateway intents.

ErrorResponse

EventCollectorDeprecated

A EventCollector receives Event’s match the given filters for a set duration.

ExecuteWebhook

A builder to create the content for a Webhook’s execution.

FollowedChannel

A container for the IDs returned by following a news channel.

FormattedTimestamp

Represents a combination of a timestamp and a style for formatting time in messages.

FormattedTimestampParseError

An error that can occur when parsing a FormattedTimestamp from a string.

ForumTag

An object that represents a tag able to be applied to a thread in a GUILD_FORUM channel.

ForumTagId

An identifier for a forum tag.

Gateway

A representation of the data retrieved from the gateway endpoint.

GatewayIntents

Gateway Intents will limit the events your bot will receive via the gateway. By default, all intents except Privileged Intents are selected.

GenericId

An identifier for an unspecific entity.

GetMessages

Builds a request to the API to retrieve messages.

Guild

Information about a Discord guild, such as channels, emojis, etc.

GuildAuditLogEntryCreateEvent

Requires GatewayIntents::GUILD_MODERATION and Permissions::VIEW_AUDIT_LOG.

GuildBanAddEvent

Requires GatewayIntents::GUILD_MODERATION.

GuildBanRemoveEvent

Requires GatewayIntents::GUILD_MODERATION.

GuildChannel

Represents a guild’s text, news, or voice channel. Some methods are available only for voice channels and some are only available for text channels. News channels are a subset of text channels and lack slow mode hence Self::rate_limit_per_user will be None.

GuildCreateEvent

Requires GatewayIntents::GUILDS.

GuildDeleteEvent

Requires GatewayIntents::GUILDS.

GuildEmojisUpdateEvent

Requires GatewayIntents::GUILD_EMOJIS_AND_STICKERS.

GuildId

An identifier for a Guild

GuildInfo

Variant of Guild returned from Http::get_guilds.

GuildIntegrationsUpdateEvent

Requires GatewayIntents::GUILD_INTEGRATIONS.

GuildMemberAddEvent

Requires GatewayIntents::GUILD_MEMBERS.

GuildMemberFlags

Flags for a guild member.

GuildMemberRemoveEvent

Requires GatewayIntents::GUILD_MEMBERS.

GuildMemberUpdateEvent

Requires GatewayIntents::GUILD_MEMBERS.

GuildMembersChunkEvent

Requires no gateway intents.

GuildPreview

Preview Guild information.

GuildPrune

Representation of the number of members that would be pruned by a guild prune operation.

GuildRoleCreateEvent

Requires GatewayIntents::GUILDS.

GuildRoleDeleteEvent

Requires GatewayIntents::GUILDS.

GuildRoleUpdateEvent

Requires GatewayIntents::GUILDS.

GuildScheduledEventCreateEvent

Requires GatewayIntents::GUILD_SCHEDULED_EVENTS.

GuildScheduledEventDeleteEvent

Requires GatewayIntents::GUILD_SCHEDULED_EVENTS.

GuildScheduledEventUpdateEvent

Requires GatewayIntents::GUILD_SCHEDULED_EVENTS.

GuildScheduledEventUserAddEvent

Requires GatewayIntents::GUILD_SCHEDULED_EVENTS.

GuildScheduledEventUserRemoveEvent

Requires GatewayIntents::GUILD_SCHEDULED_EVENTS.

GuildStickersUpdateEvent

Requires GatewayIntents::GUILD_EMOJIS_AND_STICKERS.

GuildUpdateEvent

Requires GatewayIntents::GUILDS.

GuildWelcomeChannel

A channel shown in the GuildWelcomeScreen.

GuildWelcomeScreen

Information relating to a guild’s welcome screen.

GuildWidget

A Guild widget.

Http

Note: For all member functions that return a Result, the Error kind will be either Error::Http or Error::Json.

HttpBuilder

A builder for the underlying Http client that performs requests to Discord’s HTTP API. If you do not need to use a proxy or do not need to disable the rate limiter, you can use Http::new instead.

ImageHash

An image hash returned from the Discord API.

Incident

An incident retrieved from the Discord status page.

IncidentUpdate

An update to an incident from the Discord status page.

InputText

An input text component for modal interactions

InstallParams

Settings for the application’s default in-app authorization link

Integration

Various information about integrations.

IntegrationAccount

Integration account object.

IntegrationApplication

Integration application object.

IntegrationCreateEvent

Requires GatewayIntents::GUILD_INTEGRATIONS.

IntegrationDeleteEvent

Requires GatewayIntents::GUILD_INTEGRATIONS.

IntegrationId

An identifier for an Integration

IntegrationUpdateEvent

Requires GatewayIntents::GUILD_INTEGRATIONS.

InteractionCreateEvent

Requires no gateway intents.

InteractionId

An identifier for an interaction.

InteractionResponseFlags

The flags for an interaction response message.

InvalidToken

Error that can be return by validate.

Invite

Information about an invite code.

InviteChannel

A minimal amount of information about the channel an invite points to.

InviteCreateEvent

Requires GatewayIntents::GUILD_INVITES and [`Permissions::MANAGE_CHANNELS´] permission.

InviteDeleteEvent

Requires GatewayIntents::GUILD_INVITES and [`Permissions::MANAGE_CHANNELS´] permission.

InviteGuild

Subset of Guild used in Invite.

InviteStageInstance

Discord docs.

Maintenance

A Discord status maintenance message. This can be either for active maintenances or for scheduled maintenances.

Member

Information about a member of a guild.

MembersItermodel

A helper class returned by GuildId::members_iter

Message

A representation of a message over a guild’s text channel, a group, or a private channel.

MessageActivity

Rich Presence activity information.

MessageApplication

Rich Presence application information.

MessageBuilder

The Message Builder is an ergonomic utility to easily build a message, by adding text and mentioning mentionable structs.

MessageCollector

A MessageCollector receives Message’s match the given filters for a set duration.

MessageCreateEvent

Requires GatewayIntents::GUILD_MESSAGES or GatewayIntents::DIRECT_MESSAGES.

MessageDeleteBulkEvent

Requires GatewayIntents::GUILD_MESSAGES or GatewayIntents::DIRECT_MESSAGES.

MessageDeleteEvent

Requires GatewayIntents::GUILD_MESSAGES or GatewayIntents::DIRECT_MESSAGES.

MessageFlags

Describes extra features of the message.

MessageId

An identifier for a Message

MessageInteraction

Sent when a Message is a response to an Interaction.

MessagePollVoteAddEvent

Requires GatewayIntents::GUILD_MESSAGE_POLLS or GatewayIntents::DIRECT_MESSAGE_POLLS.

MessagePollVoteRemoveEvent

Requires GatewayIntents::GUILD_MESSAGE_POLLS or GatewayIntents::DIRECT_MESSAGE_POLLS.

MessageReaction

A representation of a reaction to a message.

MessageReference

Reference data sent with crossposted messages.

MessageUpdateEvent

Requires GatewayIntents::GUILD_MESSAGES.

MessagesItermodel

A helper class returned by ChannelId::messages_iter

ModalInteraction

An interaction triggered by a modal submit.

ModalInteractionCollector

A ModalInteractionCollector receives ModalInteraction’s match the given filters for a set duration.

ModalInteractionData

A modal submit interaction data, provided by ModalInteraction::data

Multipart

Holder for multipart body. Contains upload data, multipart fields, and payload_json for creating requests with attachments.

Options

Discord docs.

PartialChannel

A container for any partial channel.

PartialCurrentApplicationInfo

Partial information about the given application.

PartialGuild

Partial information about a Guild. This does not include information like member data.

PartialGuildChannel

A partial guild channel.

PartialIntegration

Partial version of Integration, used in AuditLogs::integrations.

PartialMember

A partial amount of data for a member.

PartialThreadMember

PermissionOverwrite

A channel-specific permission overwrite for a member or role.

Permissions

A set of permissions that can be assigned to Users and Roles via PermissionOverwrites, roles globally in a Guild, and to GuildChannels.

PingInteraction

A ping interaction, which can only be received through an endpoint url.

Poll

A poll that has been attached to a Message.

PollAnswer

A possible answer for a Poll.

PollAnswerCount

The count of a single PollAnswer’s results.

PollMedia

A piece of data used in mutliple parts of the Poll structure.

PollResults

The model for the results of a Poll.

Presence

Information detailing the current online status of a User.

PresenceData

Presence data of the current user.

PresenceUpdateEvent

Requires GatewayIntents::GUILD_PRESENCES.

PresenceUser

Information about the user of a Presence event.

PresencesReplaceEventDeprecated

Not officially documented.

PrivateChannel

A Direct Message text channel with another user.

QuickModalResponsecollector

Ratelimit

A set of data containing information about the ratelimits for a particular RatelimitingBucket, which is stored in Http.

RatelimitInfo

Passed to the Ratelimiter::set_ratelimit_callback callback. If using Client, that callback is initialized to call the EventHandler::ratelimit() method.

Ratelimiter

Ratelimiter for requests to the Discord API.

RatelimitingBucket

Used to group requests together for ratelimiting.

Reaction

An emoji reaction to a message.

ReactionAddEvent

Requires GatewayIntents::GUILD_MESSAGE_REACTIONS or GatewayIntents::DIRECT_MESSAGE_REACTIONS.

ReactionCollector

A ReactionCollector receives Reaction’s match the given filters for a set duration.

ReactionConversionError

ReactionRemoveAllEvent

Requires GatewayIntents::GUILD_MESSAGE_REACTIONS or GatewayIntents::DIRECT_MESSAGE_REACTIONS.

ReactionRemoveEmojiEvent

Requires GatewayIntents::GUILD_MESSAGE_REACTIONS or GatewayIntents::DIRECT_MESSAGE_REACTIONS.

ReactionRemoveEvent

Requires GatewayIntents::GUILD_MESSAGE_REACTIONS or GatewayIntents::DIRECT_MESSAGE_REACTIONS.

Ready

An initial set of information given after IDENTIFYing to the gateway.

ReadyEvent

The “Ready” event, containing initial ready cache

Request

ResolvedOption

ResumedEvent

Requires no gateway intents.

RichInvite

Detailed information about an invite. This information can only be retrieved by anyone with the Manage Guild permission. Otherwise, a minimal amount of information can be retrieved via the Invite struct.

Role

Information about a role within a guild. A role represents a set of permissions, and can be attached to one or multiple users. A role has various miscellaneous configurations, such as being assigned a colour. Roles are unique per guild and do not cross over to other guilds in any way, and can have channel-specific permission overrides in addition to guild-level permissions.

RoleId

An identifier for a Role

RoleSubscriptionData

Discord docs

RoleTags

The tags of a Role.

Rule

Configured auto moderation rule.

RuleId

An identifier for an auto moderation rule

ScheduledEvent

Information about a guild scheduled event.

ScheduledEventId

An identifier for a Scheduled Event

ScheduledEventMetadata

Discord docs.

ScheduledEventUser

Discord docs.

SelectMenu

A select menu component.

SelectMenuOption

A select menu component options.

SessionStartLimit

Information describing how many gateway sessions you can initiate within a ratelimit period.

Settings

Settings for the cache.

Shard

A Shard is a higher-level handler for a websocket connection to Discord’s gateway. The shard allows for sending and receiving messages over the websocket, such as setting the active activity, reconnecting, syncing guilds, and more.

ShardId

An identifier for a Shard.

ShardInfo

ShardManager

A manager for handling the status of shards by starting them, restarting them, and stopping them when required.

ShardManagerOptions

ShardMessenger

A handle to a ShardRunner.

ShardQueuer

The shard queuer is a simple loop that runs indefinitely to manage the startup of shards.

ShardRunner

A runner for managing a Shard and its respective WebSocket client.

ShardRunnerInfo

Information about a ShardRunner.

ShardRunnerOptions

Options to be passed to ShardRunner::new.

ShardStageUpdateEvent

An event denoting that a shard’s connection stage was changed.

Sku

A premium offering that can be made available to an application’s users and guilds.

SkuFlags

Differentates between user and server subscriptions.

SkuId

An identifier for a SKU.

StageInstance

Discord docs.

StageInstanceCreateEvent

Requires GatewayIntents::GUILDS.

StageInstanceDeleteEvent

Requires GatewayIntents::GUILDS.

StageInstanceId

An identifier for a stage channel instance.

StageInstanceUpdateEvent

Requires GatewayIntents::GUILDS.

StatusCode

An HTTP status code (status-code in RFC 7230 et al.).

Sticker

A sticker sent with a message.

StickerId

An identifier for a sticker.

StickerItem

The smallest amount of data required to render a sticker.

StickerPack

A sticker sent with a message.

StickerPackBannerId

An identifier for a sticker pack banner.

StickerPackId

An identifier for a sticker pack.

SystemChannelFlags

Describes a system channel flags.

TargetId

An identifier for a slash command target Id. Can contain a UserId or MessageId.

Team

Information about the Team group of the application.

TeamMember

Information about a Member on a Team.

ThreadCreateEvent

Requires GatewayIntents::GUILDS.

ThreadDeleteEvent

Requires GatewayIntents::GUILDS.

ThreadListSyncEvent

Requires GatewayIntents::GUILDS.

ThreadMember

Discord docs, extra fields.

ThreadMemberFlags

Describes extra features of the message.

ThreadMemberUpdateEvent

Requires GatewayIntents::GUILDS, and, to receive this event for other users, GatewayIntents::GUILD_MEMBERS.

ThreadMembersUpdateEvent

Requires GatewayIntents::GUILD_MEMBERS.

ThreadMetadata

A thread data.

ThreadUpdateEvent

Requires GatewayIntents::GUILDS.

ThreadsData

A response to getting several threads channels.

Timestamp

TriggerMetadata

Individual change for trigger metadata within an audit log entry.

Typing

A struct to start typing in a Channel for an indefinite period of time.

TypingStartEvent

Requires GatewayIntents::GUILD_MESSAGE_TYPING or GatewayIntents::DIRECT_MESSAGE_TYPING.

UnavailableGuild

Data for an unavailable guild.

UnknownEvent

User

Information about a user.

UserId

An identifier for a User

UserPublicFlags

User’s public flags

UserUpdateEvent

Sent when properties about the current bot’s user change.

VoiceChannelStatusUpdateEvent

Requires GatewayIntents::GUILDS.

VoiceRegion

Information about an available voice region.

VoiceServerUpdateEvent

Requires no gateway intents.

VoiceState

A user’s state within a voice channel.

VoiceStateUpdateEvent

Requires GatewayIntents::GUILD_VOICE_STATES.

Webhook

A representation of a webhook, which is a low-effort way to post messages to channels. They do not necessarily require a bot user or authentication to use.

WebhookChannel

WebhookGuild

The guild object returned by a Webhook, of type WebhookType::ChannelFollower.

WebhookId

An identifier for a Webhook

WebhookUpdateEvent

Requires GatewayIntents::GUILD_WEBHOOKS.

WsClient

Enums

Action

An action which will execute whenever a rule is triggered.

ActionRowComponent

A component which can be inside of an ActionRow.

ActionType

See Action

ActivityType

Discord docs.

AfkTimeout

The Guild AFK timeout length.

AutoArchiveDuration

See ThreadMetadata::auto_archive_duration.

AutoModAction

Discord docs.

AutomodEventType

Indicates in what event context a rule should be checked.

ButtonKind

ButtonStyle

The style of a button.

Change

Channel

A container for any channel.

ChannelAction

Discord docs.

ChannelOverwriteAction

Discord docs.

ChannelParseError

Error that can be returned from Channel::convert.

ChannelType

A representation of a type of channel.

ChunkGuildFilter

Discord docs.

ClientError

An error returned from the Client.

CommandDataOptionValue

The value of an CommandDataOption.

CommandOptionType

The type of an CommandOption.

CommandPermissionType

The type of a CommandPermission.

CommandType

The type of an application command.

ComponentInteractionDataKind

ComponentType

The type of a component

ConnectionStage

Indicates the current connection stage of a Shard.

ConnectionVisibility

The visibility of a user connection on a user’s profile.

ContentModifier

Formatting modifiers for MessageBuilder content pushes

CreateActionRow

A builder for creating a components action row in a message.

CreateInteractionResponse

Discord docs.

CreateSelectMenuKind

Discord docs.

CreatorMonetizationAction

Discord docs.

DefaultMessageNotificationLevel

Default message notification level for a guild.

EmojiAction

Discord docs.

EmojiParseError

Error that can be returned from Emoji::convert.

EntitlementKind

Differentiates between Entitlement types.

EntitlementOwner

EntityType

Error

A common error enum returned by most of the library’s functionality within a custom Result.

Event

Event received over a websocket connection

ExplicitContentFilter

Setting used to filter explicit messages from members.

FormattedTimestampStyle

Enum representing various styles for formatting time in messages.

ForumEmoji

An object that specifies the emoji to use for Forum related emoji parameters.

ForumLayoutType

See GuildChannel::default_forum_layout.

FullEvent

This enum stores every possible event that an EventHandler can receive.

GatewayError

An error that occurred while attempting to deal with the gateway.

GatewayEvent

Discord docs.

GuildChannelParseError

Error that can be returned from GuildChannel::convert.

GuildPagination

Representation of the method of a query to send for the Http::get_guilds function.

GuildParseError

Error that can be returned from Guild::convert.

GuildWelcomeChannelEmoji

A GuildWelcomeScreen emoji.

GuildWidgetStyle

HttpError

ImageHashParseError

An error returned when ImageHash is passed an erronous value.

InputTextStyle

The style of the input text

IntegrationAction

Discord docs.

IntegrationExpireBehaviour

The behavior once the integration expires.

Interaction

Discord docs

InteractionType

The type of an Interaction.

InviteAction

Discord docs.

InviteTargetType

Type of target for a voice channel invite.

KeywordPresetType

Internally pre-defined wordsets which will be searched for in content.

LightMethod

An method used for ratelimiting special routes.

MemberAction

Discord docs.

MemberParseError

Error that can be returned from Member::convert.

MembershipState

Discord docs.

Mention

A struct that represents some way to insert a notification, link, or emoji into a message.

MentionParseErrormodel and utils

MessageAction

Discord docs.

MessageActivityKind

Discord docs.

MessagePagination

MessageParseError

Error that can be returned from Message::convert.

MessageType

Differentiates between regular and different types of system messages.

MfaLevel

Multi-Factor Authentication level for guild moderators.

ModelError

An error returned from the model module.

MultipartUpload

Nonce

NsfwLevel

The Guild nsfw level.

OnlineStatus

The representation of a user’s status.

Opcode

An enum representing the [gateway opcodes].

PermissionOverwriteType

The type of edit being made to a Channel’s permissions.

PollLayoutType

Represents the different layouts that a Poll may have.

PollMediaEmoji

The “Partial Emoji” attached to a PollMedia model.

PremiumTier

The guild’s premium tier, depends on the amount of users boosting the guild currently

PremiumType

Premium types denote the level of premium a user has. Visit the Nitro page to learn more about the premium plans Discord currently offers.

ReactionType

The type of a Reaction sent.

ReactionTypes

A list of types a reaction can be.

ReconnectType

The type of reconnection that should be performed.

ResolvedTarget

The resolved value of a CommandData::target_id.

ResolvedValue

The resolved value of a CommandDataOption.

RoleAction

Discord docs.

RoleParseError

Error that can be returned from Role::convert.

Route

ScheduledEventAction

Discord docs.

ScheduledEventPrivacyLevel

See ScheduledEvent::privacy_level.

ScheduledEventStatus

Discord docs.

ScheduledEventType

Discord docs.

Scope

The available OAuth2 Scopes.

ShardAction

ShardQueuerMessage

A message to be sent to the ShardQueuer.

ShardRunnerMessage

A message to send from a shard over a WebSocket.

SkuKind

Differentiates between SKU classes.

SortOrder

The sort order for threads in a forum.

StageInstanceAction

Discord docs.

StageInstancePrivacyLevel

See StageInstance::privacy_level.

StickerAction

Discord docs.

StickerFormatType

Differentiates between sticker formats.

StickerType

Differentiates between sticker types.

TeamMemberRole

ThreadAction

Discord docs.

Trigger

Characterizes the type of content which can trigger the rule.

TriggerType

Type of Trigger.

Unresolved

Option value variants that couldn’t be resolved by CommandData::options().

UserPagination

Representation of the method of a query to send for the Http::get_scheduled_event_users and Http::get_bans functions.

UserParseError

Error that can be returned from User::convert.

VerificationLevel

The level to set as criteria prior to a user being able to send messages in a Guild.

VideoQualityMode

The video quality mode for a voice channel.

VoiceChannelStatusAction

Incomplete documentation

WebhookAction

Discord docs.

WebhookType

A representation of a type of webhook.

Constants

ALREADY_AUTHENTICATED

More than one identify payload was sent.

AUTHENTICATION_FAILED

The account token sent with the identify payload was incorrect.

DECODE_ERROR

An invalid payload was sent.

DISALLOWED_GATEWAY_INTENTS

Disallowed gateway intents have been provided.

EMBED_MAX_COUNT

The maximum number of embeds in a message.

EMBED_MAX_LENGTH

The maximum length of the textual size of an embed.

GATEWAY_VERSION

The gateway version used by the library. The gateway URL is retrieved via the REST API.

INVALID_GATEWAY_INTENTS

Undocumented gateway intents have been provided.

INVALID_SEQUENCE

The sequence sent when resuming the session was invalid.

INVALID_SHARD

An invalid shard when identifying was sent.

LARGE_THRESHOLD

The large threshold to send on identify.

MEMBER_FETCH_LIMIT

The maximum number of members the bot can fetch at once

MESSAGE_CODE_LIMIT

The maximum unicode code points allowed within a message by Discord.

NOT_AUTHENTICATED

A payload was sent prior to identifying.

PRESET_GENERAL

Returns a set of permissions with the original @everyone permissions set to true.

PRESET_TEXT

Returns a set of text-only permissions with the original @everyone permissions set to true.

PRESET_VOICE

Returns a set of voice-only permissions with the original @everyone permissions set to true.

RATE_LIMITED

Payloads were being sent too quickly.

SESSION_TIMEOUT

A session timed out.

SHARDING_REQUIRED

The session would have handled too many guilds.

STICKER_MAX_COUNT

The maximum number of stickers in a message.

UNKNOWN_ERROR

Unknown error; try reconnecting?

UNKNOWN_OPCODE

Invalid Gateway OP Code.

USER_AGENT

The UserAgent sent along with every request.

Traits

ArgumentConvert

Parse a value from a string in context of a received message.

Builderhttp

Common trait for all HTTP request builders in this module.

CacheHttp

This trait will be required by functions that need Http and can optionally use a Cache to potentially avoid REST-requests.

CacheUpdate

Trait used for updating the cache with a type.

EmbedMessageBuilding

A trait with additional functionality over the MessageBuilder for creating content with additional functionality available only in embeds.

EventHandler

The core trait for handling events by serenity.

Framework

A trait for defining your own framework for serenity to use.

FutureExt

An extension trait for Futures that provides a variety of convenient adapters.

Mentionable

Allows something - such as a channel or role - to be mentioned in a message.

RawEventHandler

This core trait for handling raw events

Functions

collect

Fundamental collector function. All collector types in this module are just wrappers around this function.

content_safe

Transforms role, channel, user, @everyone and @here mentions into raw text by using the Cache and the users passed in with users.

deserialize_errors

parse_channelDeprecated

parse_channel_mention

Retrieves an Id from a channel mention.

parse_emoji

Retrieves the animated state, name and Id from an emoji mention, in the form of an EmojiIdentifier.

parse_invite

Retrieves the “code” part of an invite out of a URL.

parse_message_id_pair

Retrieves IDs from “{channel ID}-{message ID}” (retrieved by shift-clicking on “Copy ID”).

parse_message_url

Retrieves guild, channel, and message ID from a message URL.

parse_quotes

Turns a string into a vector of string arguments, splitting by spaces, but parsing content within quotes as one individual argument.

parse_roleDeprecated

parse_role_mention

Retrieves an Id from a role mention.

parse_user_mention

Retrieves an Id from a user mention.

parse_user_tag

Retrieves the username and discriminator out of a user tag (name#discrim). In order to accomodate next gen Discord usernames, this will also accept name style tags.

parse_usernameDeprecated

parse_webhook

Parses the id and token from a webhook url. Expects a url::Url rather than a &str.

shard_id

Calculates the Id of the shard responsible for a guild, given its Id and total number of shards used.

validate_token

Validates that a token is likely in a valid format.

Type Aliases

AttachmentTypeDeprecatedmodel

ChannelMessagesRef

Color

CurrentUserRef

GuildChannelRef

GuildChannelsRef

GuildRef

GuildRoleRef

GuildRolesRef

MemberRef

MessageRef

RequestBuilderDeprecated

Result

The common result type between most library functions.

SettingsRef

UserRef

Attribute Macros

async_trait