Module poise::serenity_prelude::prelude

A set of exports which can be helpful to use.

Note that the SerenityError re-export is equivalent to serenity::Error, although is re-exported as a separate name to remove likely ambiguity with other crate error enums.

Examples

Import all of the exports:

use serenity::prelude::*;

Structs

Clientgateway

The Client is the way to be able to start sending authenticated requests over the REST API, as well as initializing a WebSocket connection through Shards. Refer to the documentation on using sharding for more information.

Context

The context is a general utility struct provided on event dispatches, which helps with dealing with the current “context” of the event dispatch. The context also acts as a general high-level interface over the associated Shard which received the event, or the low-level http module.

GatewayIntents

Gateway Intents will limit the events your bot will receive via the gateway. By default, all intents except Privileged Intents are selected.

Mutex

An asynchronous Mutex-like type.

RwLock

An asynchronous reader-writer lock.

TypeMap

TypeMap is a simple abstraction around the standard library’s HashMap type, where types are its keys. This allows for statically-checked value retrieval.

Enums

ClientError

An error returned from the Client.

GatewayError

An error that occurred while attempting to deal with the gateway.

HttpError

ModelError

An error returned from the model module.

SerenityError

A common error enum returned by most of the library’s functionality within a custom Result.

Traits

CacheHttp

This trait will be required by functions that need Http and can optionally use a Cache to potentially avoid REST-requests.

EventHandler

The core trait for handling events by serenity.

Mentionable

Allows something - such as a channel or role - to be mentioned in a message.

RawEventHandler

This core trait for handling raw events

TypeMapKey

TypeMapKey is used to declare key types that are eligible for use with TypeMap.