poise
Module slash_argument

Application command argument handling code

Macros

create_slash_argument

Full version of crate::SlashArgument::create.

extract_slash_argument

Full version of crate::SlashArgument::extract.

parse_slash_args

Macro for extracting and parsing slash command arguments out of an array of serenity::CommandDataOption.

slash_argument_choices

Full version of crate::SlashArgument::choices.

Enums

SlashArgError

Possible errors when parsing slash command arguments

Traits

ContextMenuParameter

Implemented for all types that can be used in a context menu command

SlashArgument

Implement this trait on types that you want to use as a slash command parameter.