poise::slash_argument
Trait ContextMenuParameter

pub trait ContextMenuParameter<U, E> {
    // Required method
    fn to_action(
        action: fn(_: ApplicationContext<'_, U, E>, _: Self) -> BoxFuture<'_, Result<(), FrameworkError<'_, U, E>>>,
    ) -> ContextMenuCommandAction<U, E>;
}

Implemented for all types that can be used in a context menu command

Required Methods

fn to_action( action: fn(_: ApplicationContext<'_, U, E>, _: Self) -> BoxFuture<'_, Result<(), FrameworkError<'_, U, E>>>, ) -> ContextMenuCommandAction<U, E>

Convert an action function pointer that takes Self as an argument into the appropriate crate::ContextMenuCommandAction variant.

Dyn Compatibility

This trait is not dyn compatible.

In older versions of Rust, dyn compatibility was called "object safety", so this trait is not object safe.