Trait poise::slash_argument::ContextMenuParameter

pub trait ContextMenuParameter<U, E> {
    // Required method
    fn to_action(
        action: fn(_: ApplicationContext<'_, U, E>, _: Self) -> BoxFuture<'_, Result<(), FrameworkError<'_, U, E>>>
    ) -> ContextMenuCommandAction<U, E>;
}

Implemented for all types that can be used in a context menu command

Required Methods

fn to_action( action: fn(_: ApplicationContext<'_, U, E>, _: Self) -> BoxFuture<'_, Result<(), FrameworkError<'_, U, E>>> ) -> ContextMenuCommandAction<U, E>

Convert an action function pointer that takes Self as an argument into the appropriate crate::ContextMenuCommandAction variant.

Object Safety

This trait is not object safe.