Trait poise::slash_argument::SlashArgument

pub trait SlashArgument: Sized {
    // Required methods
    fn extract<'life0, 'life1, 'life2, 'life3, 'async_trait>(
        ctx: &'life0 Context,
        interaction: &'life1 CommandInteraction,
        value: &'life2 ResolvedValue<'life3>
    ) -> Pin<Box<dyn Future<Output = Result<Self, SlashArgError>> + Send + 'async_trait>>
       where Self: 'async_trait,
             'life0: 'async_trait,
             'life1: 'async_trait,
             'life2: 'async_trait,
             'life3: 'async_trait;
    fn create(builder: CreateCommandOption) -> CreateCommandOption;

    // Provided method
    fn choices() -> Vec<CommandParameterChoice> { ... }
}

Implement this trait on types that you want to use as a slash command parameter.

Required Methods

fn extract<'life0, 'life1, 'life2, 'life3, 'async_trait>( ctx: &'life0 Context, interaction: &'life1 CommandInteraction, value: &'life2 ResolvedValue<'life3> ) -> Pin<Box<dyn Future<Output = Result<Self, SlashArgError>> + Send + 'async_trait>>
where Self: 'async_trait, 'life0: 'async_trait, 'life1: 'async_trait, 'life2: 'async_trait, 'life3: 'async_trait,

Extract a Rust value of type T from the slash command argument, given via a serenity::json::Value.

Don’t call this method directly! Use crate::extract_slash_argument!

fn create(builder: CreateCommandOption) -> CreateCommandOption

Create a slash command parameter equivalent to type T.

Only fields about the argument type are filled in. The caller is still responsible for filling in name(), description(), and possibly required() or other fields.

Don’t call this method directly! Use crate::create_slash_argument!

Provided Methods

fn choices() -> Vec<CommandParameterChoice>

If this is a choice parameter, returns the choices

Don’t call this method directly! Use crate::slash_argument_choices!

Object Safety

This trait is not object safe.