Module poise::structs

Plain data structs that define the framework configuration.

Structs

ApplicationContext

Application command specific context passed to command invocations.

Command

Type returned from #[poise::command] annotated functions, which contains all of the generated prefix and application commands

CommandParameter

A single parameter of a crate::Command

CommandParameterChoice

A single drop-down choice in a slash command choice parameter

FrameworkOptions

Framework configuration

PartialContext

Trimmed down, more general version of Context

PrefixContext

Prefix-specific context passed to command invocations.

PrefixFrameworkOptions

Prefix-specific framework configuration

Enums

ApplicationCommandOrAutocompleteInteraction

Abstracts over a refernce to an application command interaction or autocomplete interaction

Context

Wrapper around either crate::ApplicationContext or crate::PrefixContext

ContextMenuCommandAction

Possible actions that a context menu entry can have

FrameworkError

Any error that can occur while the bot runs. Either thrown by user code (those variants will have an error field with your error type E in it), or originating from within the framework.

MessageDispatchTrigger

The event that triggered a prefix command execution

Prefix

Possible ways to define a command prefix