Struct poise::track_edits::EditTracker

pub struct EditTracker { /* private fields */ }

Stores messages and the associated bot responses in order to implement poise’s edit tracking feature.

Implementations

impl EditTracker

pub fn for_timespan(duration: Duration) -> RwLock<Self>

Create an edit tracker which tracks messages for the specified duration.

Note: EditTracker will only purge messages outside the duration when Self::purge is called. If you supply the created EditTracker to crate::Framework, the framework will take care of that by calling Self::purge periodically.

pub fn process_message_update( &mut self, user_msg_update: &MessageUpdateEvent, ignore_edits_if_not_yet_responded: bool ) -> Option<(Message, bool)>

Returns a copy of a newly up-to-date cached message, or a brand new generated message when not in cache. Also returns a bool with true if this message was previously tracked

Returns None if the command shouldn’t be re-run, e.g. if the message content wasn’t edited

pub fn process_message_delete( &mut self, deleted_message_id: MessageId ) -> Option<Message>

Removes this command invocation from the cache and returns the associated bot response, if the command invocation is cached, and it has an associated bot response, and the command is marked track_deletion

pub fn purge(&mut self)

Forget all of the messages that are older than the specified duration.

pub fn find_bot_response(&self, user_msg_id: MessageId) -> Option<&Message>

Given a message by a user, find the corresponding bot response, if one exists and is cached.

pub fn set_bot_response( &mut self, user_msg: &Message, bot_response: Message, track_deletion: bool )

Notify the EditTracker that the given user message should be associated with the given bot response. Overwrites any previous associated bot response

pub fn track_command(&mut self, user_msg: &Message, track_deletion: bool)

Store that this command is currently running; so that if the command is editing its own invocation message (e.g. removing embeds), we don’t accidentally treat it as an execute_untracked_edits situation and start an infinite loop