Trait poise::Autocompletable

pub trait Autocompletable {
    type Partial;

    fn extract_partial(value: &Value) -> Result<Self::Partial, SlashArgError>;
    fn into_json(self) -> Value;
}

Types that can be marked autocompletable in a slash command parameter.

Includes almost all types that can be used as a slash command parameter in general, except some built-in model types (User, Member, Role…)

Required Associated Types

type Partial

Type of the partial input. This should be Self except in cases where a partial input cannot be parsed into Self (e.g. an IP address)

Required Methods

fn extract_partial(value: &Value) -> Result<Self::Partial, SlashArgError>

Try extracting the partial input from the JSON value

Equivalent to crate::SlashArgument::extract

Don’t call this method directly! Use crate::extract_autocomplete_argument!

fn into_json(self) -> Value

Serialize an autocompletion choice as a JSON value.

This is the counterpart to Self::extract_partial

Don’t call this method directly! Use crate::autocomplete_argument_into_json!