Trait poise::SlashArgument

pub trait SlashArgument: Sized {
    fn extract<'life0, 'life1, 'life2, 'async_trait>(
        ctx: &'life0 Context,
        interaction: ApplicationCommandOrAutocompleteInteraction<'life1>,
        value: &'life2 Value
    ) -> Pin<Box<dyn Future<Output = Result<Self, SlashArgError>> + Send + 'async_trait>>
   where
        'life0: 'async_trait,
        'life1: 'async_trait,
        'life2: 'async_trait,
        Self: 'async_trait;
    fn create(builder: &mut CreateApplicationCommandOption);
    fn choices() -> Vec<CommandParameterChoice>;
}

Implement this trait on types that you want to use as a slash command parameter.

Required Methods

fn extract<'life0, 'life1, 'life2, 'async_trait>(
ctx: &'life0 Context,
interaction: ApplicationCommandOrAutocompleteInteraction<'life1>,
value: &'life2 Value
) -> Pin<Box<dyn Future<Output = Result<Self, SlashArgError>> + Send + 'async_trait>>where
'life0: 'async_trait,
'life1: 'async_trait,
'life2: 'async_trait,
Self: 'async_trait,

Extract a Rust value of type T from the slash command argument, given via a serenity::json::Value.

Don’t call this method directly! Use crate::extract_slash_argument!

fn create(builder: &mut CreateApplicationCommandOption)

Create a slash command parameter equivalent to type T.

Only fields about the argument type are filled in. The caller is still responsible for filling in name(), description(), and possibly required() or other fields.

Don’t call this method directly! Use crate::create_slash_argument!

fn choices() -> Vec<CommandParameterChoice>

If this is a choice parameter, returns the choices

Don’t call this method directly! Use crate::slash_argument_choices!